Combine methods into family named tableau files
