--- src/tableau/rkv76.rs
@@ -1,632 +1,0 @@
-use crate::tableau::{
-    ButcherTableau,
-};
-
-impl ButcherTableau<10,13> {
-    /// A 'most efficient' Runge-Kutta (10:6(7)) pair with 6th-order interpolation.
-    ///
-    /// # Overview
-    /// This provides a 10-stage Runge-Kutta method with:
-    /// - Primary order: 6
-    /// - Embedded order: 7 (for error estimation)
-    /// - Number of stages: 10 primary + 3 additional stages for interpolation
-    /// - Dominant stage order: 3
-    ///
-    /// # Efficiency
-    /// This method is considered "most efficient" in the sense that for a specified
-    /// maximum coefficient from `b` and `A`, it almost minimizes the local truncation
-    /// error 2-norm:
-    ///
-    /// T_82 ~ 0.000003389
-    ///
-    /// (Formulas with slightly different nodes may have a slightly smaller error norm,
-    /// perhaps achieved by having a larger maximum coefficient.)
-    ///
-    /// # Interpolation
-    /// - Node c[11]=1 was chosen to make the interpolants differentiable
-    /// - Node c[12] was chosen as a zero of a quadratic polynomial to enable efficient interpolation
-    /// - The interpolant has continuous derivatives
-    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
-    ///
-    ///   Ti_72 ~ 0.0000165
-    ///
-    /// - The interpolation error has four local maximum values on [0,1]
-    ///
-    /// # References
-    /// - J.H. Verner, SIAM NA 1978, "Explicit Runge-Kutta methods with estimates of the Local Truncation Error"
-    /// - J.H. Verner, SIAM NA 30, 1993, "Differentiable Interpolants for high-order Runge-Kutta methods"
-    pub const fn rkv766e() -> Self {
-        let mut c = [0.0; 13];
-        let mut a = [[0.0; 13]; 13];
-        let mut b = [0.0; 10];
-        let mut bh = [0.0; 10];
-        let mut bi6 = [[0.0; 13]; 13];
-
-        c[0] = 0.0;
-        c[1] = 0.69e-1;
-        c[2] = 0.118;
-        c[3] = 0.177;
-        c[4] = 0.501;
-        c[5] = 0.7737799115305331003715765296862487670813;
-        c[6] = 0.994;
-        c[7] = 0.998;
-        c[8] = 1.0;
-        c[9] = 1.0;
-
-        a[1][0] = 0.69e-1;
-
-        a[2][0] = 0.1710144927536231884057971014492753623188e-1;
-        a[2][1] = 0.1008985507246376811594202898550724637681;
-
-        a[3][0] = 0.4425e-1;
-        a[3][1] = 0.0;
-        a[3][2] = 0.13275;
-
-        a[4][0] = 0.7353445130709566216604424016087331226659;
-        a[4][1] = 0.0;
-        a[4][2] = -2.830160657856937661591496696351623096811;
-        a[4][3] = 2.595816144785981039931054294742889974145;
-
-        a[5][0] = -12.21580485360407974005910916471598682362;
-        a[5][1] = 0.0;
-        a[5][2] = 48.82665485823736062335980699373053427134;
-        a[5][3] = -38.55615592319928364666616600329792491404;
-        a[5][4] = 2.719085830096535863737044703969626233400;
-
-        a[6][0] = 108.8614188704176574066699618897203578466;
-        a[6][1] = 0.0;
-        a[6][2] = -432.4521181775777896358931629332707752654;
-        a[6][3] = 343.9115281800118289547200158889409233641;
-        a[6][4] = -20.55041135925273709189369488701721016265;
-        a[6][5] = 1.223582486401040366396880041626704217305;
-
-        a[7][0] = 113.4755131883738522204615568160304033854;
-        a[7][1] = 0.0;
-        a[7][2] = -450.8122021555997002820400438087344405365;
-        a[7][3] = 358.5132765190089889943579090008312808216;
-        a[7][4] = -21.45046667648445540174055882443151176550;
-        a[7][5] = 1.274053318605952891766776667539031508649;
-        a[7][6] = -0.2174193904638422805639851234763413667602e-2;
-
-        a[8][0] = 115.6996223324232534824963925993127275021;
-        a[8][1] = 0.0;
-        a[8][2] = -459.6635446100248030478961869239726305957;
-        a[8][3] = 365.5534717131745930309149378867953890507;
-        a[8][4] = -21.88511586349784824146225495848432937529;
-        a[8][5] = 1.298718109698721459187976480852777474315;
-        a[8][6] = -0.5318700918481883515898878747322241917739e-4;
-        a[8][7] = -0.3098494764731864405706095716460833640254e-2;
-
-        a[9][0] = 124.1543935612464600014576130437603883332;
-        a[9][1] = 0.0;
-        a[9][2] = -493.2318713314597046194663569971348299332;
-        a[9][3] = 392.2086219315800762927575562172365337929;
-        a[9][4] = -23.48641564290853341361596821616234280392;
-        a[9][5] = 1.362322948908907509911149920532561575254;
-        a[9][6] = -0.7051467367205771043993968232310964220061e-2;
-        a[9][7] = 0.0;
-        a[9][8] = 0.0;
-
-        b[0] = 0.5163520172057869163393251056217968836723e-1;
-        b[1] = 0.0;
-        b[2] = 0.0;
-        b[3] = 0.2767172535461648728769641534539952501983;
-        b[4] = 0.3374175285287150670818592701488271741753;
-        b[5] = 0.1884488267810967803491085059046161195540;
-        b[6] = 24.54134121634868026791753618430192161716;
-        b[7] = -68.81190284469011946382716084194838780382;
-        b[8] = 44.41634281776488378396776021757684795437;
-        b[9] = 0.0;
-
-        bh[0] = 0.5089676583692947576073561095512200263213e-1;
-        bh[1] = 0.0;
-        bh[2] = 0.0;
-        bh[3] = 0.2793777374763233901369432426263934138476;
-        bh[4] = 0.3281330142746535239936396881369403928344;
-        bh[5] = 0.2241721218186151033581794837350130009230;
-        bh[6] = 0.7874574778015076584344903106189416715189;
-        bh[7] = 0.0;
-        bh[8] = 0.0;
-        bh[9] = -0.6700371172080291516839883360724104817561;
-
-        c[10] = 1.0;
-
-        a[10][0] = 0.5163520172057869163393251056217968836723e-1;
-        a[10][1] = 0.0;
-        a[10][2] = 0.0;
-        a[10][3] = 0.2767172535461648728769641534539952501983;
-        a[10][4] = 0.3374175285287150670818592701488271741753;
-        a[10][5] = 0.1884488267810967803491085059046161195540;
-        a[10][6] = 24.54134121634868026791753618430192161716;
-        a[10][7] = -68.81190284469011946382716084194838780382;
-        a[10][8] = 44.41634281776488378396776021757684795437;
-        a[10][9] = 0.0;
-
-        c[11] = 0.3206919742864321488153177349612198724445;
-
-        a[11][0] = 0.5595947882055415742583956081637901452552e-1;
-        a[11][1] = 0.0;
-        a[11][2] = 0.0;
-        a[11][3] = 0.2481262830322509327139022005325437452965;
-        a[11][4] = 0.2103052990591148013587987543073879445482e-1;
-        a[11][5] = -0.1086652415454008449407261022406674546105e-1;
-        a[11][6] = 5.710311335454033086222886871347348898943;
-        a[11][7] = -16.59749151785386284126327966975768025349;
-        a[11][8] = 10.90440686603898964031684903021468235733;
-        a[11][9] = 0.0;
-        a[11][10] = -0.1078447695690422224268752339872593914865e-1;
-
-        c[12] = 0.107;
-
-        a[12][0] = 0.5365176199729764566758655497507692759771e-1;
-        a[12][1] = 0.0;
-        a[12][2] = 0.0;
-        a[12][3] = 0.9828513377552059696782507697799478414148e-1;
-        a[12][4] = 0.2236190536978546688920574502539925812364e-1;
-        a[12][5] = -0.3196168594283279436559562148268690316884e-2;
-        a[12][6] = 0.5451003257307239985591851329642383758416;
-        a[12][7] = -1.638732002009432676897859879241483267007;
-        a[12][8] = 1.093896392979466849054156189707873746269;
-        a[12][9] = 0.0;
-        a[12][10] = -0.4657483105197485920611507484730881742515e-4;
-        a[12][11] = -0.6432077441802662594433314318598382583251e-1;
-
-        bi6[0][0] = 1.0;
-        bi6[0][1] = -7.582446684249578716313054661512721933240;
-        bi6[0][2] = 21.65339697937461679177171822544604176620;
-        bi6[0][3] = -26.50313170111562393346845843813196066333;
-        bi6[0][4] = 13.68567041142918450667757590934976189120;
-        bi6[0][5] = -2.201853803718019957033848524588941372473;
-
-        bi6[1][0] = 0.0;
-        bi6[1][1] = 0.0;
-        bi6[1][2] = 0.0;
-        bi6[1][3] = 0.0;
-        bi6[1][4] = 0.0;
-        bi6[1][5] = 0.0;
-
-        bi6[2][0] = 0.0;
-        bi6[2][1] = 0.0;
-        bi6[2][2] = 0.0;
-        bi6[2][3] = 0.0;
-        bi6[2][4] = 0.0;
-        bi6[2][5] = 0.0;
-
-        bi6[3][0] = 0.0;
-        bi6[3][1] = 5.889034278272535871631390961135086381406;
-        bi6[3][2] = -57.96960154935775491821473273509196795388;
-        bi6[3][3] = 174.1066668861731135303452708623511580529;
-        bi6[3][4] = -196.2003627160861165553101224432427862687;
-        bi6[3][5] = 74.45098035454438694442515750830250503840;
-
-        bi6[4][0] = 0.0;
-        bi6[4][1] = 3.196839349796654162957352828525358112148;
-        bi6[4][2] = -31.32397572700528310044209947396113560366;
-        bi6[4][3] = 93.05298102970384392438258309708366195791;
-        bi6[4][4] = -102.8968871064061647967771234654923865084;
-        bi6[4][5] = 38.30845998243966487696114628399332921615;
-
-        bi6[5][0] = 0.0;
-        bi6[5][1] = -0.8787886596190119861699566896289502028596e-1;
-        bi6[5][2] = 1.013797983077119427760544302767092045959;
-        bi6[5][3] = -4.100173002382458810769471031390844351476;
-        bi6[5][4] = 6.641160480067744814819942865759689363544;
-        bi6[5][5] = -3.278457768019407452844911962268425918186;
-
-        bi6[6][0] = 0.0;
-        bi6[6][1] = -208.3972099204096382331570396018818793788;
-        bi6[6][2] = 2077.909119755834934676745809403064182251;
-        bi6[6][3] = -6428.966291404445362174964426975619327549;
-        bi6[6][4] = 7605.042110521116554859824801255385155563;
-        bi6[6][5] = -3021.046387735747808860531607896646209269;
-
-        bi6[7][0] = 0.0;
-        bi6[7][1] = 594.3593292228751204027648218620251141015;
-        bi6[7][2] = -5925.388971699224459341618710912833599773;
-        bi6[7][3] = 18326.52193738665917733083306878451733081;
-        bi6[7][4] = -21667.18569363528617503083225733032464553;
-        bi6[7][5] = 8602.881495880286217175025916754667412586;
-
-        bi6[8][0] = 0.0;
-        bi6[8][1] = -386.8811915634934821635422144581019035900;
-        bi6[8][2] = 3856.672838888456624774287551141874741487;
-        bi6[8][3] = -11926.20151472972403478718897872831668331;
-        bi6[8][4] = 14096.40733595464142660949072316887784424;
-        bi6[8][5] = -5595.581125732115650649079320906757150875;
-
-        bi6[9][0] = 0.0;
-        bi6[9][1] = 0.0;
-        bi6[9][2] = 0.0;
-        bi6[9][3] = 0.0;
-        bi6[9][4] = 0.0;
-        bi6[9][5] = 0.0;
-
-        bi6[10][0] = 0.0;
-        bi6[10][1] = 0.6111099397023196267050121119527364301482;
-        bi6[10][2] = -6.151956371109289894189681549817017604566;
-        bi6[10][3] = 19.45873699463664083616118469228645171564;
-        bi6[10][4] = -23.90604463475469049657337318293279633818;
-        bi6[10][5] = 9.988154071525019927896857928510625796953;
-
-        bi6[11][0] = 0.0;
-        bi6[11][1] = -8.007634773786193306062361780345514409318;
-        bi6[11][2] = 74.60031392403404792779391526165329391946;
-        bi6[11][3] = -199.6748294772657466686035644277447081048;
-        bi6[11][4] = 207.5792562775741227780748301919115920885;
-        bi6[11][5] = -74.49710595055623073120281924547466349383;
-
-        bi6[12][0] = 0.0;
-        bi6[12][1] = 6.900049017254163553633088407166619306472;
-        bi6[12][2] = -11.01496218408055634389431366310163053450;
-        bi6[12][3] = -27.69438198223954924672720783503507856034;
-        bi6[12][4] = 60.83345444770411331060500303070857149829;
-        bi6[12][5] = -29.02415929863817127361656993973848170993;
-
-        Self {
-            c,
-            a,
-            b,
-            bh: Some(bh),
-            bi: Some(bi6),
-        }
-    }
-}
-
-impl ButcherTableau<10, 16> {
-    /// A 'most efficient' Runge-Kutta (10:7(6)) pair with 7th-order interpolation.
-    ///
-    /// # Overview
-    /// This provides a 10-stage Runge-Kutta method with:
-    /// - Primary order: 7
-    /// - Embedded order: 6 (for error estimation)
-    /// - Number of stages: 10 primary + 6 additional stages for interpolation
-    /// - Dominant stage order: 3
-    ///
-    /// # Efficiency
-    /// This method is considered "most efficient" in the sense that for a specified
-    /// maximum coefficient from `b` and `A`, it almost minimizes the local truncation
-    /// error 2-norm:
-    ///
-    /// T_82 ~ 0.000003389
-    ///
-    /// # Interpolation
-    /// - Additional stages allow computation of approximations at any point with order up to 7
-    /// - The interpolant has continuous derivatives
-    /// - Maximum 2-norm of the local truncation error for the 7th-order interpolant:
-    ///
-    ///   Ti_82 ~ 0.000003389
-    ///
-    /// - The 2-norm is bounded by this value on [0,1] and is nearly monotone increasing
-    /// - The 7th-order interpolant requires 6 additional stages (16 total)
-    ///
-    /// # Notes
-    /// - This method uses the same primary stages as the 6th-order method (rkv766e)
-    /// - The higher-order interpolant requires more additional stages
-    ///
-    /// # References
-    /// - J.H. Verner, SIAM NA 1978, "Explicit Runge-Kutta methods with estimates of the Local Truncation Error"
-    /// - J.H. Verner, SIAM NA 30, 1993, "Differentiable Interpolants for high-order Runge-Kutta methods"
-    pub const fn rkv767e() -> Self {
-        let mut c = [0.0; 16];
-        let mut a = [[0.0; 16]; 16];
-        let mut b = [0.0; 10];
-        let mut bh = [0.0; 10];
-        let mut bi7 = [[0.0; 16]; 16];
-
-        c[0] = 0.0;
-        c[1] = 0.69e-1;
-        c[2] = 0.118;
-        c[3] = 0.177;
-        c[4] = 0.501;
-        c[5] = 0.7737799115305331003715765296862487670813;
-        c[6] = 0.994;
-        c[7] = 0.998;
-        c[8] = 1.0;
-        c[9] = 1.0;
-
-        a[1][0] = 0.69e-1;
-
-        a[2][0] = 0.1710144927536231884057971014492753623188e-1;
-        a[2][1] = 0.1008985507246376811594202898550724637681;
-
-        a[3][0] = 0.4425e-1;
-        a[3][1] = 0.0;
-        a[3][2] = 0.13275;
-
-        a[4][0] = 0.7353445130709566216604424016087331226659;
-        a[4][1] = 0.0;
-        a[4][2] = -2.830160657856937661591496696351623096811;
-        a[4][3] = 2.595816144785981039931054294742889974145;
-
-        a[5][0] = -12.21580485360407974005910916471598682362;
-        a[5][1] = 0.0;
-        a[5][2] = 48.82665485823736062335980699373053427134;
-        a[5][3] = -38.55615592319928364666616600329792491404;
-        a[5][4] = 2.719085830096535863737044703969626233400;
-
-        a[6][0] = 108.8614188704176574066699618897203578466;
-        a[6][1] = 0.0;
-        a[6][2] = -432.4521181775777896358931629332707752654;
-        a[6][3] = 343.9115281800118289547200158889409233641;
-        a[6][4] = -20.55041135925273709189369488701721016265;
-        a[6][5] = 1.223582486401040366396880041626704217305;
-
-        a[7][0] = 113.4755131883738522204615568160304033854;
-        a[7][1] = 0.0;
-        a[7][2] = -450.8122021555997002820400438087344405365;
-        a[7][3] = 358.5132765190089889943579090008312808216;
-        a[7][4] = -21.45046667648445540174055882443151176550;
-        a[7][5] = 1.274053318605952891766776667539031508649;
-        a[7][6] = -0.2174193904638422805639851234763413667602e-2;
-
-        a[8][0] = 115.6996223324232534824963925993127275021;
-        a[8][1] = 0.0;
-        a[8][2] = -459.6635446100248030478961869239726305957;
-        a[8][3] = 365.5534717131745930309149378867953890507;
-        a[8][4] = -21.88511586349784824146225495848432937529;
-        a[8][5] = 1.298718109698721459187976480852777474315;
-        a[8][6] = -0.5318700918481883515898878747322241917739e-4;
-        a[8][7] = -0.3098494764731864405706095716460833640254e-2;
-
-        a[9][0] = 124.1543935612464600014576130437603883332;
-        a[9][1] = 0.0;
-        a[9][2] = -493.2318713314597046194663569971348299332;
-        a[9][3] = 392.2086219315800762927575562172365337929;
-        a[9][4] = -23.48641564290853341361596821616234280392;
-        a[9][5] = 1.362322948908907509911149920532561575254;
-        a[9][6] = -0.7051467367205771043993968232310964220061e-2;
-        a[9][7] = 0.0;
-        a[9][8] = 0.0;
-
-        b[0] = 0.5163520172057869163393251056217968836723e-1;
-        b[1] = 0.0;
-        b[2] = 0.0;
-        b[3] = 0.2767172535461648728769641534539952501983;
-        b[4] = 0.3374175285287150670818592701488271741753;
-        b[5] = 0.1884488267810967803491085059046161195540;
-        b[6] = 24.54134121634868026791753618430192161716;
-        b[7] = -68.81190284469011946382716084194838780382;
-        b[8] = 44.41634281776488378396776021757684795437;
-        b[9] = 0.0;
-
-        bh[0] = 0.5089676583692947576073561095512200263213e-1;
-        bh[1] = 0.0;
-        bh[2] = 0.0;
-        bh[3] = 0.2793777374763233901369432426263934138476;
-        bh[4] = 0.3281330142746535239936396881369403928344;
-        bh[5] = 0.2241721218186151033581794837350130009230;
-        bh[6] = 0.7874574778015076584344903106189416715189;
-        bh[7] = 0.0;
-        bh[8] = 0.0;
-        bh[9] = -0.6700371172080291516839883360724104817561;
-
-        c[10] = 1.0;
-
-        a[10][0] = 0.5163520172057869163393251056217968836723e-1;
-        a[10][1] = 0.0;
-        a[10][2] = 0.0;
-        a[10][3] = 0.2767172535461648728769641534539952501983;
-        a[10][4] = 0.3374175285287150670818592701488271741753;
-        a[10][5] = 0.1884488267810967803491085059046161195540;
-        a[10][6] = 24.54134121634868026791753618430192161716;
-        a[10][7] = -68.81190284469011946382716084194838780382;
-        a[10][8] = 44.41634281776488378396776021757684795437;
-        a[10][9] = 0.0;
-
-        c[11] = 0.3206919742864321488153177349612198724445;
-
-        a[11][0] = 0.5595947882055415742583956081637901452552e-1;
-        a[11][1] = 0.0;
-        a[11][2] = 0.0;
-        a[11][3] = 0.2481262830322509327139022005325437452965;
-        a[11][4] = 0.2103052990591148013587987543073879445482e-1;
-        a[11][5] = -0.1086652415454008449407261022406674546105e-1;
-        a[11][6] = 5.710311335454033086222886871347348898943;
-        a[11][7] = -16.59749151785386284126327966975768025349;
-        a[11][8] = 10.90440686603898964031684903021468235733;
-        a[11][9] = 0.0;
-        a[11][10] = -0.1078447695690422224268752339872593914865e-1;
-
-        c[12] = 0.107;
-
-        a[12][0] = 0.5365176199729764566758655497507692759771e-1;
-        a[12][1] = 0.0;
-        a[12][2] = 0.0;
-        a[12][3] = 0.9828513377552059696782507697799478414148e-1;
-        a[12][4] = 0.2236190536978546688920574502539925812364e-1;
-        a[12][5] = -0.3196168594283279436559562148268690316884e-2;
-        a[12][6] = 0.5451003257307239985591851329642383758416;
-        a[12][7] = -1.638732002009432676897859879241483267007;
-        a[12][8] = 1.093896392979466849054156189707873746269;
-        a[12][9] = 0.0;
-        a[12][10] = -0.4657483105197485920611507484730881742515e-4;
-        a[12][11] = -0.6432077441802662594433314318598382583251e-1;
-
-        c[13] = 0.3;
-
-        a[13][0] = 0.1919717775847849685111579308732385545808e-1;
-        a[13][1] = 0.0;
-        a[13][2] = 0.0;
-        a[13][3] = -0.4739427173175533989609356107380842465934e-1;
-        a[13][4] = -0.2641522514821109801913974252340613156326e-1;
-        a[13][5] = 0.7054089167217366089095708745682724510887e-7;
-        a[13][6] = 1.551079892101621437320838156104778064997;
-        a[13][7] = -4.428095538025376982599451861908192890036;
-        a[13][8] = 2.883718325674217149748617484261428760772;
-        a[13][9] = 0.0;
-        a[13][10] = -0.4297481934497428346001203166435386609045e-2;
-        a[13][11] = 0.1262644300588589746368092621052046424260;
-        a[13][12] = 0.2259426207057731181296447821560200523865;
-
-        c[14] = 0.549;
-
-        a[14][0] = 0.6125242269008585247666695476683627174961e-1;
-        a[14][1] = 0.0;
-        a[14][2] = 0.0;
-        a[14][3] = 0.2525310401834353816786786047482953692850;
-        a[14][4] = 0.1507054877413180556949606462372746086457;
-        a[14][5] = 0.1024270639019054271328153617559339591462e-1;
-        a[14][6] = -6.438444503804215943508588627336983785574;
-        a[14][7] = 18.44870904095403871125278855846631959613;
-        a[14][8] = -12.03592743273174801641739335902460845726;
-        a[14][9] = 0.0;
-        a[14][10] = 0.1513193748819522199500460159517896993473e-1;
-        a[14][11] = 0.1043447705299915390244100081194588662149;
-        a[14][12] = -0.1954546944129134490980892374736483504656e-1;
-        a[14][13] = 0.0;
-
-        c[15] = 0.757;
-
-        a[15][0] = 0.8959466056806525077749567658988579412839e-1;
-        a[15][1] = 0.0;
-        a[15][2] = 0.0;
-        a[15][3] = 0.6389214946562193863310204803419201536875;
-        a[15][4] = 0.4309120296822662010167723442150653884245;
-        a[15][5] = 0.7695592670277012062550066233926247449166e-1;
-        a[15][6] = -7.189894742007330283923979037259652075040;
-        a[15][7] = 21.03645589658795446951415976277135782026;
-        a[15][8] = -13.86140603343527732092225205611764539171;
-        a[15][9] = 0.0;
-        a[15][10] = 0.8282681779726606982694318905099680481702e-2;
-        a[15][11] = -0.2148456743069278028823832921628374810433;
-        a[15][12] = -0.2579762402274666275190288596224563636723;
-        a[15][13] = 0.0;
-        a[15][14] = 0.0;
-
-        bi7[0][0] = 1.0;
-        bi7[0][1] = -8.503890289208650086322705078936310611299;
-        bi7[0][2] = 31.18234339085710879255131238050971776003;
-        bi7[0][3] = -56.83047976918539469582020937384021018043;
-        bi7[0][4] = 52.79739831471022178930385951784095486118;
-        bi7[0][5] = -22.95183302720539338830128746758333934610;
-        bi7[0][6] = 3.358096581752686280222962532571367204997;
-
-        bi7[1][0] = 0.0;
-        bi7[1][1] = 0.0;
-        bi7[1][2] = 0.0;
-        bi7[1][3] = 0.0;
-        bi7[1][4] = 0.0;
-        bi7[1][5] = 0.0;
-        bi7[1][6] = 0.0;
-
-        bi7[2][0] = 0.0;
-        bi7[2][1] = 0.0;
-        bi7[2][2] = 0.0;
-        bi7[2][3] = 0.0;
-        bi7[2][4] = 0.0;
-        bi7[2][5] = 0.0;
-        bi7[2][6] = 0.0;
-
-        bi7[3][0] = 0.0;
-        bi7[3][1] = 6.267709807177576432406478077069849818602;
-        bi7[3][2] = -68.59145123448141827278305837000589905312;
-        bi7[3][3] = 270.1006378954180870537780616882949819305;
-        bi7[3][4] = -486.2388094162716631231341727978491179719;
-        bi7[3][5] = 407.1399818231500101241727526996656040234;
-        bi7[3][6] = -128.4013516214464273415630971437214234972;
-
-        bi7[4][0] = 0.0;
-        bi7[4][1] = 7.642585077624110415092972164105836819717;
-        bi7[4][2] = -83.63756743442629097888399107640792828691;
-        bi7[4][3] = 329.3495021534571296433374383575409388467;
-        bi7[4][4] = -592.8994135546816065453818037938193613121;
-        bi7[4][5] = 496.4495876982774414703954108911602253524;
-        bi7[4][6] = -156.5672764117220689374781672724308842456;
-
-        bi7[5][0] = 0.0;
-        bi7[5][1] = 4.268409521381499328229656167910411979775;
-        bi7[5][2] = -46.71186327091233753264841481404164445942;
-        bi7[5][3] = 183.9428068612486847316162527859995659975;
-        bi7[5][4] = -331.1363205426106275865463338454994779573;
-        bi7[5][5] = 277.2687677656847319301330472909470886979;
-        bi7[5][6] = -87.44335150801085409043509907941132813885;
-
-        bi7[6][0] = 0.0;
-        bi7[6][1] = 555.8670558189042010639619738936099948255;
-        bi7[6][2] = -6083.199322405529992849514286236321590545;
-        bi7[6][3] = 23954.53059900739546845895162864872763414;
-        bi7[6][4] = -43123.26890207833867031432260116626186577;
-        bi7[6][5] = 36108.19740617653066320546034515369066848;
-        bi7[6][6] = -11387.58549530261298929661952410914291952;
-
-        bi7[7][0] = 0.0;
-        bi7[7][1] = -1558.605518026587403407644982678592484625;
-        bi7[7][2] = 17056.79070544840796438857322691726272923;
-        bi7[7][3] = -67166.53412450698972069889474017010525280;
-        bi7[7][4] = 120914.1001657429967184958527822159285576;
-        bi7[7][5] = -101244.4165095385499516578794640912685653;
-        bi7[7][6] = 31929.85337803603227341616601696482662805;
-
-        bi7[8][0] = 0.0;
-        bi7[8][1] = 1006.040439872400448623697106213084759679;
-        bi7[8][2] = -11009.72698072282062279582145142481879227;
-        bi7[8][3] = 43354.29892541348545701322502088633140394;
-        bi7[8][4] = -78046.99336079503441879814959011469619894;
-        bi7[8][5] = 65350.71006860324690110919871372728749245;
-        bi7[8][6] = -20609.91274955351288136818203906961181690;
-
-        bi7[9][0] = 0.0;
-        bi7[9][1] = 0.0;
-        bi7[9][2] = 0.0;
-        bi7[9][3] = 0.0;
-        bi7[9][4] = 0.0;
-        bi7[9][5] = 0.0;
-        bi7[9][6] = 0.0;
-
-        bi7[10][0] = 0.0;
-        bi7[10][1] = -1.612970387190759595570543259931261611538;
-        bi7[10][2] = 17.72909772735324772041234368381732499671;
-        bi7[10][3] = -70.34937384312194980443642443527448780299;
-        bi7[10][4] = 128.1350088912801760638876471983594315219;
-        bi7[10][5] = -109.0734352266536956182626795265083915641;
-        bi7[10][6] = 35.17167283833298123396965633953738445994;
-
-        bi7[11][0] = 0.0;
-        bi7[11][1] = -2.626599644929271923086786435983631061624;
-        bi7[11][2] = 9.920024380612826013243113812430759529521;
-        bi7[11][3] = -7.644130389085101218500146944756885977802;
-        bi7[11][4] = -14.86430552714213266618722882949033593695;
-        bi7[11][5] = 26.11390292373462455033637542344644699734;
-        bi7[11][6] = -10.89889174319094475580532702564635355049;
-
-        bi7[12][0] = 0.0;
-        bi7[12][1] = 9.561675242924823427014852060688684933761;
-        bi7[12][2] = -36.11210856303548088470642279130799743208;
-        bi7[12][3] = 27.82711572968801973495325529709252051216;
-        bi7[12][4] = 54.11089673926027907217912866089741112920;
-        bi7[12][5] = -95.06308263006681094546659235128381873539;
-        bi7[12][6] = 39.67550348122916959602577912391319959234;
-
-        bi7[13][0] = 0.0;
-        bi7[13][1] = -9.791432054831754863321754176936907826271;
-        bi7[13][2] = 116.6733547596190133707339789196320683056;
-        bi7[13][3] = -427.0178661529826169224279390827040458460;
-        bi7[13][4] = 710.0693817808117189359000209005264369724;
-        bi7[13][5] = -556.8214238669928662708433693467844704976;
-        bi7[13][6] = 166.8879855343765057499590627862669188919;
-
-        bi7[14][0] = 0.0;
-        bi7[14][1] = 0.7372370767192203283333752534265429181322e-1;
-        bi7[14][2] = 2.611540012355412116044967235469929822777;
-        bi7[14][3] = -30.67846770866969010553103505851633192303;
-        bi7[14][4] = 100.4285900318910859666056494362334084911;
-        bi7[14][5] = -119.6365555255543602449647502655108119632;
-        bi7[14][6] = 47.20116948230563023501183112698115128060;
-
-        bi7[15][0] = 0.0;
-        bi7[15][1] = -8.581188645336741447289604471431596612033;
-        bi7[15][2] = 93.07222791200057091279868176378132240393;
-        bi7[15][3] = -360.9951446906583731902511625987898308351;
-        bi7[15][4] = 635.7596704131289187099926426178301572747;
-        bi7[15][5] = -517.9168751756012942639785021372581285996;
-        bi7[15][6] = 158.6613101864669192787279448258680763681;
-
-        Self {
-            c,
-            a,
-            b,
-            bh: Some(bh),
-            bi: Some(bi7),
-        }
-    }
-}
