--- src/tableau/rkv87.rs
@@ -1,909 +1,0 @@
-use crate::tableau::{
-    ButcherTableau,
-};
-
-impl ButcherTableau<13, 17> {
-    /// An efficient Runge-Kutta (13:7) pair method with 7th-order interpolation.
-    ///
-    /// # Overview
-    /// This provides a 13-stage Runge-Kutta method with:
-    /// - Primary order: 7
-    /// - Embedded order: 7 (used for stepsize control)
-    /// - Number of stages: 13 primary + 4 additional stages for interpolation
-    /// - Dominant stage order: 4
-    ///
-    /// # Interpolation
-    /// - The method provides a 7th-order interpolant using the nodes c[14]=1, c[15]-c[17].
-    /// - The interpolant has continuous derivatives.
-    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
-    ///   
-    ///   Ti_82 ~ 0.0000063833
-    ///   
-    /// - The interpolation error has three local maximum values on [0,1].
-    ///
-    /// # Notes
-    /// - This method uses the same initial 13 stages as the 8th-order method (rkv878e).
-    /// - The interpolant requires fewer additional stages than the 8th-order version.
-    ///
-    /// # References
-    /// - J.H. Verner, SIAM NA 1978, "Explicit Runge--Kutta methods with estimates of the Local Truncation Error"
-    /// - J.H. Verner, SIAM NA 30, 1993, "Differentiable Interpolants for high-order Runge--Kutta methods"
-    pub const fn rkv877e() -> Self {
-        let mut c = [0.0; 17];
-        let mut a = [[0.0; 17]; 17];
-        let mut b = [0.0; 13];
-        let mut bh = [0.0; 13];
-        let mut bi7 = [[0.0; 17]; 17];
-
-        c[0] = 0.0;
-        c[1] = 0.92662e-1;
-        c[2] = 0.1312230361754017604780747799402406525075;
-        c[3] = 0.1968345542631026407171121699103609787612;
-        c[4] = 0.427173;
-        c[5] = 0.485972;
-        c[6] = 0.161915;
-        c[7] = 0.985468;
-        c[8] = 0.9626977348604540392115664231947926180050;
-        c[9] = 0.99626;
-        c[10] = 0.997947;
-        c[11] = 1.0;
-        c[12] = 1.0;
-
-        a[1][0] = 0.92662e-1;
-
-        a[2][0] = 0.3830746548250284242039554953085778876548e-1;
-        a[2][1] = 0.9291557069289891805767923040938286374199e-1;
-
-        a[3][0] = 0.4920863856577566017927804247759024469030e-1;
-        a[3][1] = 0.0;
-        a[3][2] = 0.1476259156973269805378341274327707340709;
-
-        a[4][0] = 0.2743076085702486894953667699331084997155;
-        a[4][1] = 0.0;
-        a[4][2] = -0.9319887203102656329703895298805780048911;
-        a[4][3] = 1.084854111740016943475022759947469505176;
-
-        a[5][0] = 0.6461852970939692178473977038055776659235e-1;
-        a[5][1] = 0.0;
-        a[5][2] = 0.0;
-        a[5][3] = 0.2687629213368923356417418162215129378455;
-        a[5][4] = 0.1525905489537107425735184133979292955622;
-
-        a[6][0] = 0.7189155819773216802747111989237192757241e-1;
-        a[6][1] = 0.0;
-        a[6][2] = 0.0;
-        a[6][3] = 0.1221265783362549661095644262553368220549;
-        a[6][4] = -0.7943550859198561207449556225926957458888e-1;
-        a[6][5] = 0.4733237205799847793746001611156082496161e-1;
-
-        a[7][0] = -6.073603893714328779581044654669389532243;
-        a[7][1] = 0.0;
-        a[7][2] = 0.0;
-        a[7][3] = -73.8956;
-        a[7][4] = 11.93985370695273926305714852083567777565;
-        a[7][5] = -3.839251541405054537968455811084326767757;
-        a[7][6] = 72.85406972816664405449235194491803852435;
-
-        a[8][0] = -4.868640079323569115532909137641679122100;
-        a[8][1] = 0.0;
-        a[8][2] = 0.0;
-        a[8][3] = -59.18572799975646020086112235306397792921;
-        a[8][4] = 9.230819319232425236436363261556755728972;
-        a[8][5] = -2.676847914962525780976208768457285096420;
-        a[8][6] = 58.45720009994685754009028962323039746457;
-        a[8][7] = 0.5894309723726360055153797570581572198272e-2;
-
-        a[9][0] = -6.689861899320853351893486767784362717327;
-        a[9][1] = 0.0;
-        a[9][2] = 0.0;
-        a[9][3] = -81.44271004053111286646869388606624375878;
-        a[9][4] = 13.36778825698397107436768361456719138955;
-        a[9][5] = -4.470777638416181156550300441220009175768;
-        a[9][6] = 80.23321392161410397716314034458127569922;
-        a[9][7] = -0.1313638336212181564579824084323466067556e-1;
-        a[9][8] = 0.1174378303219413902745537676538322377075e-1;
-
-        a[10][0] = -6.788841955800464091202279324616261399375;
-        a[10][1] = 0.0;
-        a[10][2] = 0.0;
-        a[10][3] = -82.65639855934828888595908426400236814314;
-        a[10][4] = 13.59973921874899036529372991555792312537;
-        a[10][5] = -4.574464055350503753023585975406886448931;
-        a[10][6] = 81.41943207216075927716035509139094411649;
-        a[10][7] = -0.1416248014826418017596237345824563213708e-1;
-        a[10][8] = 0.1375441580835227405704608471667945126391e-1;
-        a[10][9] = -0.1111656070581006150219154181785069541859e-2;
-
-        a[11][0] = -6.910189846402485729483801895952800612193;
-        a[11][1] = 0.0;
-        a[11][2] = 0.0;
-        a[11][3] = -84.14495154176748682468399818934446849309;
-        a[11][4] = 13.88512122378983816888937012217322612857;
-        a[11][5] = -4.702458788144493296650978281402083070191;
-        a[11][6] = 82.87411451529241610017736751027256505872;
-        a[11][7] = -0.1645498337198780129448623133684652608689e-1;
-        a[11][8] = 0.1644663972162521365153657249183964172873e-1;
-        a[11][9] = 0.4275449370796530995842729680318146579263e-2;
-        a[11][10] = -0.5902668488222361600852336581750274034670e-2;
-
-        a[12][0] = -6.911973921198979615960353484856419992328;
-        a[12][1] = 0.0;
-        a[12][2] = 0.0;
-        a[12][3] = -84.16635595878781036984379688648536865215;
-        a[12][4] = 13.88834627565582007275122755367241993143;
-        a[12][5] = -4.703463178409702575934171709526131267400;
-        a[12][6] = 82.89518622207404915885492275044886922215;
-        a[12][7] = -0.1020345016228260287805622619295788864042e-1;
-        a[12][8] = 0.1427900423230391471526298444108039935137e-1;
-        a[12][9] = -0.5814993403397981705034981501491752405078e-2;
-        a[12][10] = 0.0;
-        a[12][11] = 0.0;
-
-        b[0] = 0.4625543159712467285354070519930680076661e-1;
-        b[1] = 0.0;
-        b[2] = 0.0;
-        b[3] = 0.0;
-        b[4] = 0.0;
-        b[5] = 0.3706666165521011182439275381303388440188;
-        b[6] = 0.2590440824552746577195309846039127860158;
-        b[7] = -679.9841468175039046601229652340421033215;
-        b[8] = 49.89161129042053159104301060910837813887;
-        b[9] = 10271.23522213731241388784467688648118861;
-        b[10] = -14782.19660635689728059010570422698640618;
-        b[11] = 5141.377953616063739322523982737505384322;
-        b[12] = 0.0;
-
-        bh[0] = 0.4638504234365210644214797353760063769606e-1;
-        bh[1] = 0.0;
-        bh[2] = 0.0;
-        bh[3] = 0.0;
-        bh[4] = 0.0;
-        bh[5] = 0.3725767681581196020016675337652919509451;
-        bh[6] = 0.2585685495121687311414935155287094016045;
-        bh[7] = -147.4950767589265301611858882107047892762;
-        bh[8] = 23.84362712644587506762535723842314335392;
-        bh[9] = 347.4264166730550802334952839143352280200;
-        bh[10] = 0.0;
-        bh[11] = 0.0;
-        bh[12] = -223.4524974005883655795200619648851840879;
-
-        c[13] = 1.0;
-
-        a[13][0] = 0.4625543159712467285354070519930680076661e-1;
-        a[13][1] = 0.0;
-        a[13][2] = 0.0;
-        a[13][3] = 0.0;
-        a[13][4] = 0.0;
-        a[13][5] = 0.3706666165521011182439275381303388440188;
-        a[13][6] = 0.2590440824552746577195309846039127860158;
-        a[13][7] = -679.9841468175039046601229652340421033215;
-        a[13][8] = 49.89161129042053159104301060910837813887;
-        a[13][9] = 10271.23522213731241388784467688648118861;
-        a[13][10] = -14782.19660635689728059010570422698640618;
-        a[13][11] = 5141.377953616063739322523982737505384322;
-        a[13][12] = 0.0;
-
-        c[14] = 0.3110177634953863863927417318829099695921;
-
-        a[14][0] = 0.4911138760562453021745624018512939494564e-1;
-        a[14][1] = 0.0;
-        a[14][2] = 0.0;
-        a[14][3] = 0.0;
-        a[14][4] = 0.0;
-        a[14][5] = 0.3117373122649832951079085303758614870942e-1;
-        a[14][6] = 0.2384923679553289916353395945129932253433;
-        a[14][7] = -114.3180256262527071531290136810612300898;
-        a[14][8] = 3.546408122969840402366032921479574656434;
-        a[14][9] = 2460.597003929216132135942626992844507475;
-        a[14][10] = -3738.924986857493946343920875863790949734;
-        a[14][11] = 1389.104435743812477156452794999320497315;
-        a[14][12] = 0.0;
-        a[14][13] = -0.1259503554386166268241032464519842162533e-1;
-
-        c[15] = 0.2;
-
-        a[15][0] = 0.5179349197485644585674743626479955978608e-1;
-        a[15][1] = 0.0;
-        a[15][2] = 0.0;
-        a[15][3] = 0.0;
-        a[15][4] = 0.0;
-        a[15][5] = 0.3491876693023421912147307261172789499006e-1;
-        a[15][6] = 0.1991958559185454797079748618106548367164;
-        a[15][7] = -83.39793280532507836124705826791773425675;
-        a[15][8] = 3.594107171680368155688005352337368373839;
-        a[15][9] = 1608.071558743333488448763391096377128121;
-        a[15][10] = -2405.179666082838573124443824955693898542;
-        a[15][11] = 876.9134237687558313788496904957089183499;
-        a[15][12] = 0.0;
-        a[15][13] = -0.7325540972670704140488857561266969491257e-2;
-        a[15][14] = -0.8007336945700193815591023393769736797741e-1;
-
-        c[16] = 0.6;
-
-        a[16][0] = 0.4436713671328350085533516489655998875304e-1;
-        a[16][1] = 0.0;
-        a[16][2] = 0.0;
-        a[16][3] = 0.0;
-        a[16][4] = 0.0;
-        a[16][5] = 0.2341061441705438467404449942880480215796;
-        a[16][6] = 0.3613912760091486818803425821815508521337;
-        a[16][7] = 148.9884506482496936874751145110012899551;
-        a[16][8] = -9.426031101789146984989508415038462154627;
-        a[16][9] = -2394.447533524453400464298733938611210343;
-        a[16][10] = 3478.192564092670645140476435107181157928;
-        a[16][11] = -1223.319642639490730778566879958817950079;
-        a[16][12] = 0.0;
-        a[16][13] = 0.1180395533992265808766615770156321870980e-1;
-        a[16][14] = 0.1349647037014828275755998285816810134371;
-        a[16][15] = -0.1744406911214421152358160333642284015360;
-
-        bi7[0][0] = 1.0;
-        bi7[0][1] = -8.212110712910061778007623084400813862130;
-        bi7[0][2] = 35.02635320769604791474092034077537606592;
-        bi7[0][3] = -83.00213339876337627241952209202186684333;
-        bi7[0][4] = 108.6409000057226676877405870074292701819;
-        bi7[0][5] = -72.99647106020921661717338874349522718142;
-        bi7[0][6] = 19.58971739006106373797256727691256843986;
-
-        bi7[1][0] = 0.0;
-        bi7[1][1] = 0.0;
-        bi7[1][2] = 0.0;
-        bi7[1][3] = 0.0;
-        bi7[1][4] = 0.0;
-        bi7[1][5] = 0.0;
-        bi7[1][6] = 0.0;
-
-        bi7[2][0] = 0.0;
-        bi7[2][1] = 0.0;
-        bi7[2][2] = 0.0;
-        bi7[2][3] = 0.0;
-        bi7[2][4] = 0.0;
-        bi7[2][5] = 0.0;
-        bi7[2][6] = 0.0;
-
-        bi7[3][0] = 0.0;
-        bi7[3][1] = 0.0;
-        bi7[3][2] = 0.0;
-        bi7[3][3] = 0.0;
-        bi7[3][4] = 0.0;
-        bi7[3][5] = 0.0;
-        bi7[3][6] = 0.0;
-
-        bi7[4][0] = 0.0;
-        bi7[4][1] = 0.0;
-        bi7[4][2] = 0.0;
-        bi7[4][3] = 0.0;
-        bi7[4][4] = 0.0;
-        bi7[4][5] = 0.0;
-        bi7[4][6] = 0.0;
-
-        bi7[5][0] = 0.0;
-        bi7[5][1] = 24.49504632679124657248259074284359937838;
-        bi7[5][2] = -197.6984345684939158045973156075628087774;
-        bi7[5][3] = 648.8167079210014594451537283787671992452;
-        bi7[5][4] = -1029.913614541901893330814502946072472329;
-        bi7[5][5] = 784.2902782766835465098516222387889570492;
-        bi7[5][6] = -229.6193167975283422738321952686341357219;
-
-        bi7[6][0] = 0.0;
-        bi7[6][1] = 31.56918953505704448457573784274510770401;
-        bi7[6][2] = -253.5018651524157970015956512936350906006;
-        bi7[6][3] = 825.6524274319808118183975171217026655762;
-        bi7[6][4] = -1296.622633293770326113106573164742870221;
-        bi7[6][5] = 974.2628058031631049585612278174199570978;
-        bi7[6][6] = -281.1008802415595634891127273388858567704;
-
-        bi7[7][0] = 0.0;
-        bi7[7][1] = 13532.46062486349519418285003949403697565;
-        bi7[7][2] = -103993.6912108123447484699919983753429873;
-        bi7[7][3] = 315790.0517926441798783566243358352274542;
-        bi7[7][4] = -444650.9408635634151265292682380509278972;
-        bi7[7][5] = 285484.2990404036663083335205079890657793;
-        bi7[7][6] = -66842.16353035308541053385761212610142806;
-
-        bi7[8][0] = 0.0;
-        bi7[8][1] = -797.3373937748126309130191533345207238350;
-        bi7[8][2] = 6069.299906337186949139058394441653960311;
-        bi7[8][3] = -18132.72848645950527711755529827748496005;
-        bi7[8][4] = 24818.08337443399123270224838601234691225;
-        bi7[8][5] = -15179.25266693120755490566761402249251945;
-        bi7[8][6] = 3271.826877684767812685978295789605708913;
-
-        bi7[9][0] = 0.0;
-        bi7[9][1] = -223490.9058981754844945580562298428473795;
-        bi7[9][2] = 1723137.281290412037973827810617942131263;
-        bi7[9][3] = -5261550.360871223434658436208344654853597;
-        bi7[9][4] = 7478225.628671211907649685713214284343111;
-        bi7[9][5] = -4874996.123844563053849368849978953045265;
-        bi7[9][6] = 1168945.715874475339792737435398110753057;
-
-        bi7[10][0] = 0.0;
-        bi7[10][1] = 325937.3545980911494561573600538089395441;
-        bi7[10][2] = -2514175.045366638831325535367544931340193;
-        bi7[10][3] = 7682914.240892724251462067375934686533398;
-        bi7[10][4] = -10933909.13214721004597165165677694552484;
-        bi7[10][5] = 7142613.573847848634614325115730923207696;
-        bi7[10][6] = -1718163.188431172055515952933101768802010;
-
-        bi7[11][0] = 0.0;
-        bi7[11][1] = -115180.8935430816768533254678552911538230;
-        bi7[11][2] = 888954.8267433133410033905646433783738063;
-        bi7[11][3] = -2718988.108015552606334524484935846851660;
-        bi7[11][4] = 3875443.744975928173788736413809491152039;
-        bi7[11][5] = -2537848.359487731062145576624229336937517;
-        bi7[11][6] = 612760.1672807398942806221225503429225392;
-
-        bi7[12][0] = 0.0;
-        bi7[12][1] = 0.0;
-        bi7[12][2] = 0.0;
-        bi7[12][3] = 0.0;
-        bi7[12][4] = 0.0;
-        bi7[12][5] = 0.0;
-        bi7[12][6] = 0.0;
-
-        bi7[13][0] = 0.0;
-        bi7[13][1] = 0.8171974528089539184193047902519773567838;
-        bi7[13][2] = -5.929889493933239653903563757522305816160;
-        bi7[13][3] = 16.17047142225921334983931469529060140792;
-        bi7[13][4] = -18.18864073091308954267555489038495985876;
-        bi7[13][5] = 6.499437906736728059350896773727451974473;
-        bi7[13][6] = 0.6314234430414338689696023886372349357410;
-
-        bi7[14][0] = 0.0;
-        bi7[14][1] = -19.58991707760692246575484124057989970426;
-        bi7[14][2] = 128.4420521863155939052637968458080346391;
-        bi7[14][3] = -321.8756286792641832868863295281122163797;
-        bi7[14][4] = 394.5721351206491640005134395005296613000;
-        bi7[14][5] = -239.3360075607335414326488715970553134960;
-        bi7[14][6] = 57.78736601063988927951280601940973364088;
-
-        bi7[15][0] = 0.0;
-        bi7[15][1] = -15.35274693752665932678945181344254585974;
-        bi7[15][2] = 180.1686249551895421740728266391690224117;
-        bi7[15][3] = -701.6476414218982078177499899730825381674;
-        bi7[15][4] = 1211.468823272660855113033560419426852892;
-        bi7[15][5] = -961.9054874127519588351611984090694516295;
-        bi7[15][6] = 287.2684275443264286925942531369986603532;
-
-        bi7[16][0] = 0.0;
-        bi7[16][1] = -14.40504650928427294859257207187201833753;
-        bi7[16][2] = 120.8217962542519161139448743774919222629;
-        bi7[16][3] = -417.2095154082007875820864103451144802558;
-        bi7[16][4] = 702.6590193669410492418586492824251953058;
-        bi7[16][5] = -564.9514449798660354502747046801145472079;
-        bi7[16][6] = 173.0851912761581306251501634371839282326;
-
-        Self {
-            c,
-            a,
-            b,
-            bh: Some(bh),
-            bi: Some(bi7),
-        }
-    }
-}
-
-impl ButcherTableau<13, 21> {
-    /// A highly efficient Runge-Kutta (13:8(7)) pair method.
-    ///
-    /// # Overview
-    /// This is an efficient 13-stage Runge-Kutta method that provides:
-    /// - Primary order: 8
-    /// - Embedded order: 7 (for error estimation)
-    /// - Number of stages: 13 primary + 8 additional stages for interpolation
-    /// - Dominant stage order: 4
-    ///
-    /// # Efficiency
-    /// The method is particularly "efficient" with a local truncation error 2-norm:
-    ///
-    /// T_92 ~ 0.000000011182
-    ///
-    /// which appears to be the minimum possible for this type of method.
-    ///
-    /// # Interpolation
-    /// - The additional stages (c[14] through c[21]) allow computation of approximations
-    ///   at any point in the solution domain with order 8.
-    /// - The interpolant has continuous derivatives.
-    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
-    ///   
-    ///   Ti_92 ~ 0.0000011515
-    ///   
-    /// - The interpolation error has two local maximum values on [0,1].
-    ///
-    /// # References
-    /// Developed based on formulas in:
-    /// - J.H. Verner, SIAM NA 1978, "Explicit Runge--Kutta methods with estimates of the Local Truncation Error"
-    /// - J.H. Verner, SIAM NA 30, 1993, "Differentiable Interpolants for high-order Runge--Kutta methods"
-    pub const fn rkv878e() -> Self {
-        let mut c = [0.0; 21];
-        let mut a = [[0.0; 21]; 21];
-        let mut b = [0.0; 13];
-        let mut bh = [0.0; 13];
-        let mut bi8 = [[0.0; 21]; 21];
-
-        c[0] = 0.0;
-        c[1] = 0.92662e-1;
-        c[2] = 0.1312230361754017604780747799402406525075;
-        c[3] = 0.1968345542631026407171121699103609787612;
-        c[4] = 0.427173;
-        c[5] = 0.485972;
-        c[6] = 0.161915;
-        c[7] = 0.985468;
-        c[8] = 0.9626977348604540392115664231947926180050;
-        c[9] = 0.99626;
-        c[10] = 0.997947;
-        c[11] = 1.0;
-        c[12] = 1.0;
-
-        a[1][0] = 0.92662e-1;
-
-        a[2][0] = 0.3830746548250284242039554953085778876548e-1;
-        a[2][1] = 0.9291557069289891805767923040938286374199e-1;
-
-        a[3][0] = 0.4920863856577566017927804247759024469030e-1;
-        a[3][1] = 0.0;
-        a[3][2] = 0.1476259156973269805378341274327707340709;
-
-        a[4][0] = 0.2743076085702486894953667699331084997155;
-        a[4][1] = 0.0;
-        a[4][2] = -0.9319887203102656329703895298805780048911;
-        a[4][3] = 1.084854111740016943475022759947469505176;
-
-        a[5][0] = 0.6461852970939692178473977038055776659235e-1;
-        a[5][1] = 0.0;
-        a[5][2] = 0.0;
-        a[5][3] = 0.2687629213368923356417418162215129378455;
-        a[5][4] = 0.1525905489537107425735184133979292955622;
-
-        a[6][0] = 0.7189155819773216802747111989237192757241e-1;
-        a[6][1] = 0.0;
-        a[6][2] = 0.0;
-        a[6][3] = 0.1221265783362549661095644262553368220549;
-        a[6][4] = -0.7943550859198561207449556225926957458888e-1;
-        a[6][5] = 0.4733237205799847793746001611156082496161e-1;
-
-        a[7][0] = -6.073603893714328779581044654669389532243;
-        a[7][1] = 0.0;
-        a[7][2] = 0.0;
-        a[7][3] = -73.8956;
-        a[7][4] = 11.93985370695273926305714852083567777565;
-        a[7][5] = -3.839251541405054537968455811084326767757;
-        a[7][6] = 72.85406972816664405449235194491803852435;
-
-        a[8][0] = -4.868640079323569115532909137641679122100;
-        a[8][1] = 0.0;
-        a[8][2] = 0.0;
-        a[8][3] = -59.18572799975646020086112235306397792921;
-        a[8][4] = 9.230819319232425236436363261556755728972;
-        a[8][5] = -2.676847914962525780976208768457285096420;
-        a[8][6] = 58.45720009994685754009028962323039746457;
-        a[8][7] = 0.5894309723726360055153797570581572198272e-2;
-
-        a[9][0] = -6.689861899320853351893486767784362717327;
-        a[9][1] = 0.0;
-        a[9][2] = 0.0;
-        a[9][3] = -81.44271004053111286646869388606624375878;
-        a[9][4] = 13.36778825698397107436768361456719138955;
-        a[9][5] = -4.470777638416181156550300441220009175768;
-        a[9][6] = 80.23321392161410397716314034458127569922;
-        a[9][7] = -0.1313638336212181564579824084323466067556e-1;
-        a[9][8] = 0.1174378303219413902745537676538322377075e-1;
-
-        a[10][0] = -6.788841955800464091202279324616261399375;
-        a[10][1] = 0.0;
-        a[10][2] = 0.0;
-        a[10][3] = -82.65639855934828888595908426400236814314;
-        a[10][4] = 13.59973921874899036529372991555792312537;
-        a[10][5] = -4.574464055350503753023585975406886448931;
-        a[10][6] = 81.41943207216075927716035509139094411649;
-        a[10][7] = -0.1416248014826418017596237345824563213708e-1;
-        a[10][8] = 0.1375441580835227405704608471667945126391e-1;
-        a[10][9] = -0.1111656070581006150219154181785069541859e-2;
-
-        a[11][0] = -6.910189846402485729483801895952800612193;
-        a[11][1] = 0.0;
-        a[11][2] = 0.0;
-        a[11][3] = -84.14495154176748682468399818934446849309;
-        a[11][4] = 13.88512122378983816888937012217322612857;
-        a[11][5] = -4.702458788144493296650978281402083070191;
-        a[11][6] = 82.87411451529241610017736751027256505872;
-        a[11][7] = -0.1645498337198780129448623133684652608689e-1;
-        a[11][8] = 0.1644663972162521365153657249183964172873e-1;
-        a[11][9] = 0.4275449370796530995842729680318146579263e-2;
-        a[11][10] = -0.5902668488222361600852336581750274034670e-2;
-
-        a[12][0] = -6.911973921198979615960353484856419992328;
-        a[12][1] = 0.0;
-        a[12][2] = 0.0;
-        a[12][3] = -84.16635595878781036984379688648536865215;
-        a[12][4] = 13.88834627565582007275122755367241993143;
-        a[12][5] = -4.703463178409702575934171709526131267400;
-        a[12][6] = 82.89518622207404915885492275044886922215;
-        a[12][7] = -0.1020345016228260287805622619295788864042e-1;
-        a[12][8] = 0.1427900423230391471526298444108039935137e-1;
-        a[12][9] = -0.5814993403397981705034981501491752405078e-2;
-        a[12][10] = 0.0;
-        a[12][11] = 0.0;
-
-        b[0] = 0.4625543159712467285354070519930680076661e-1;
-        b[1] = 0.0;
-        b[2] = 0.0;
-        b[3] = 0.0;
-        b[4] = 0.0;
-        b[5] = 0.3706666165521011182439275381303388440188;
-        b[6] = 0.2590440824552746577195309846039127860158;
-        b[7] = -679.9841468175039046601229652340421033215;
-        b[8] = 49.89161129042053159104301060910837813887;
-        b[9] = 10271.23522213731241388784467688648118861;
-        b[10] = -14782.19660635689728059010570422698640618;
-        b[11] = 5141.377953616063739322523982737505384322;
-        b[12] = 0.0;
-
-        bh[0] = 0.4638504234365210644214797353760063769606e-1;
-        bh[1] = 0.0;
-        bh[2] = 0.0;
-        bh[3] = 0.0;
-        bh[4] = 0.0;
-        bh[5] = 0.3725767681581196020016675337652919509451;
-        bh[6] = 0.2585685495121687311414935155287094016045;
-        bh[7] = -147.4950767589265301611858882107047892762;
-        bh[8] = 23.84362712644587506762535723842314335392;
-        bh[9] = 347.4264166730550802334952839143352280200;
-        bh[10] = 0.0;
-        bh[11] = 0.0;
-        bh[12] = -223.4524974005883655795200619648851840879;
-
-        c[13] = 1.0;
-
-        a[13][0] = 0.4625543159712467285354070519930680076661e-1;
-        a[13][1] = 0.0;
-        a[13][2] = 0.0;
-        a[13][3] = 0.0;
-        a[13][4] = 0.0;
-        a[13][5] = 0.3706666165521011182439275381303388440188;
-        a[13][6] = 0.2590440824552746577195309846039127860158;
-        a[13][7] = -679.9841468175039046601229652340421033215;
-        a[13][8] = 49.89161129042053159104301060910837813887;
-        a[13][9] = 10271.23522213731241388784467688648118861;
-        a[13][10] = -14782.19660635689728059010570422698640618;
-        a[13][11] = 5141.377953616063739322523982737505384322;
-        a[13][12] = 0.0;
-
-        c[14] = 0.3110177634953863863927417318829099695921;
-
-        a[14][0] = 0.4911138760562453021745624018512939494564e-1;
-        a[14][1] = 0.0;
-        a[14][2] = 0.0;
-        a[14][3] = 0.0;
-        a[14][4] = 0.0;
-        a[14][5] = 0.3117373122649832951079085303758614870942e-1;
-        a[14][6] = 0.2384923679553289916353395945129932253433;
-        a[14][7] = -114.3180256262527071531290136810612300898;
-        a[14][8] = 3.546408122969840402366032921479574656434;
-        a[14][9] = 2460.597003929216132135942626992844507475;
-        a[14][10] = -3738.924986857493946343920875863790949734;
-        a[14][11] = 1389.104435743812477156452794999320497315;
-        a[14][12] = 0.0;
-        a[14][13] = -0.1259503554386166268241032464519842162533e-1;
-
-        c[15] = 0.2;
-
-        a[15][0] = 0.5179349197485644585674743626479955978608e-1;
-        a[15][1] = 0.0;
-        a[15][2] = 0.0;
-        a[15][3] = 0.0;
-        a[15][4] = 0.0;
-        a[15][5] = 0.3491876693023421912147307261172789499006e-1;
-        a[15][6] = 0.1991958559185454797079748618106548367164;
-        a[15][7] = -83.39793280532507836124705826791773425675;
-        a[15][8] = 3.594107171680368155688005352337368373839;
-        a[15][9] = 1608.071558743333488448763391096377128121;
-        a[15][10] = -2405.179666082838573124443824955693898542;
-        a[15][11] = 876.9134237687558313788496904957089183499;
-        a[15][12] = 0.0;
-        a[15][13] = -0.7325540972670704140488857561266969491257e-2;
-        a[15][14] = -0.8007336945700193815591023393769736797741e-1;
-
-        c[16] = 0.6;
-
-        a[16][0] = 0.4436713671328350085533516489655998875304e-1;
-        a[16][1] = 0.0;
-        a[16][2] = 0.0;
-        a[16][3] = 0.0;
-        a[16][4] = 0.0;
-        a[16][5] = 0.2341061441705438467404449942880480215796;
-        a[16][6] = 0.3613912760091486818803425821815508521337;
-        a[16][7] = 148.9884506482496936874751145110012899551;
-        a[16][8] = -9.426031101789146984989508415038462154627;
-        a[16][9] = -2394.447533524453400464298733938611210343;
-        a[16][10] = 3478.192564092670645140476435107181157928;
-        a[16][11] = -1223.319642639490730778566879958817950079;
-        a[16][12] = 0.0;
-        a[16][13] = 0.1180395533992265808766615770156321870980e-1;
-        a[16][14] = 0.1349647037014828275755998285816810134371;
-        a[16][15] = -0.1744406911214421152358160333642284015360;
-
-        c[17] = 0.251;
-
-        a[17][0] = 0.4927309335160545018837186692162194114006e-1;
-        a[17][1] = 0.0;
-        a[17][2] = 0.0;
-        a[17][3] = 0.0;
-        a[17][4] = 0.0;
-        a[17][5] = 0.1478473900086014712887940970539529915367;
-        a[17][6] = 0.1915391503780640342241981813756999876738;
-        a[17][7] = 85.69949519051240679824481172082685765978;
-        a[17][8] = -5.094095153569584154270332765058351241762;
-        a[17][9] = -1410.985179005200593252911690612297856885;
-        a[17][10] = 2056.873083845176325396897975581141270029;
-        a[17][11] = -726.4906359530084743440935010911966625383;
-        a[17][12] = 0.0;
-        a[17][13] = 0.5440493103670101179960418796640615006313e-2;
-        a[17][14] = -0.1437912346504880100428763097368602714809;
-        a[17][15] = 0.8130926198406332202245160436986436093968e-1;
-        a[17][16] = -0.8328707808559681272816269219617664899487e-1;
-
-        c[18] = 0.451;
-
-        a[18][0] = 0.4696791108973669580037521395572005624943e-1;
-        a[18][1] = 0.0;
-        a[18][2] = 0.0;
-        a[18][3] = 0.0;
-        a[18][4] = 0.0;
-        a[18][5] = 0.2010207606761097437232392903607688964540;
-        a[18][6] = 0.2636820858834097192664348219821625228752;
-        a[18][7] = 129.6946428658473021706281203179680341347;
-        a[18][8] = -7.848674084672579804338628962846012175606;
-        a[18][9] = -2121.729757286215698398545896705499941278;
-        a[18][10] = 3090.157418699412077487400584308334376641;
-        a[18][11] = -1090.274673114349457503609053352481192228;
-        a[18][12] = 0.0;
-        a[18][13] = 0.8963759217746587577580254051724170346208e-2;
-        a[18][14] = 0.4875201082359089303480235638217668769362e-1;
-        a[18][15] = -0.2352913684548179696659019693524338623585e-1;
-        a[18][16] = -0.9381447086675579397096734527257404140647e-1;
-        a[18][17] = 0.0;
-
-        c[19] = 0.705;
-
-        a[19][0] = 0.4619987481300149103923582280519782866590e-1;
-        a[19][1] = 0.0;
-        a[19][2] = 0.0;
-        a[19][3] = 0.0;
-        a[19][4] = 0.0;
-        a[19][5] = 0.2322836783017554180760864761340094394641;
-        a[19][6] = 0.2982051978226913141175018152700660351379;
-        a[19][7] = 123.6261700088200670668986465184100244666;
-        a[19][8] = -7.231890372340911603936034207018277000416;
-        a[19][9] = -2046.801418638051394033916845196316375171;
-        a[19][10] = 2986.087035469917759873466009280884537880;
-        a[19][11] = -1055.671796398942928677813090910742298972;
-        a[19][12] = 0.0;
-        a[19][13] = 0.7930343978313866522846570067876462025829e-2;
-        a[19][14] = 0.8980932823961332375619194670691924357732e-1;
-        a[19][15] = -0.8017153994764179414644520644800061934373e-1;
-        a[19][16] = 0.1026430473896737559358970902463204065417;
-        a[19][17] = 0.0;
-        a[19][18] = 0.0;
-
-        c[20] = 0.8795;
-
-        a[20][0] = 0.5329137345256786599202448371739175627057e-1;
-        a[20][1] = 0.0;
-        a[20][2] = 0.0;
-        a[20][3] = 0.0;
-        a[20][4] = 0.0;
-        a[20][5] = 0.2043185802985482892225157926395741314696;
-        a[20][6] = 0.1358728897801712829783843582299538835128;
-        a[20][7] = -402.8079844608128209694219700870509710185;
-        a[20][8] = 29.46110568417652997397169152745513583508;
-        a[20][9] = 6093.592821606031634755087198810590639032;
-        a[20][10] = -8771.855718519913765428348513348127981921;
-        a[20][11] = 3051.798314820693282143366860169701133691;
-        a[20][12] = 0.0;
-        a[20][13] = -0.3308969275605729485812518877458826604068e-1;
-        a[20][14] = 0.2805382986735395987509175854573185663385e-1;
-        a[20][15] = 0.1335953634219562765350679978518628190610;
-        a[20][16] = 0.1689185257605991455997737252221182006005;
-        a[20][17] = 0.0;
-        a[20][18] = 0.0;
-        a[20][19] = 0.0;
-
-        bi8[0][0] = 1.0;
-        bi8[0][1] = -8.560562317505520965893266232545864348587;
-        bi8[0][2] = 37.91120319790645210673257106393947681809;
-        bi8[0][3] = -95.90997970387049459577928861171245504580;
-        bi8[0][4] = 143.9557989560652588678965218768899263890;
-        bi8[0][5] = -126.8645258830010362955147010306166337461;
-        bi8[0][6] = 60.67897508156613701498205859458556491581;
-        bi8[0][7] = -12.16465389956367145957035495534070818157;
-
-        bi8[1][0] = 0.0;
-        bi8[1][1] = 0.0;
-        bi8[1][2] = 0.0;
-        bi8[1][3] = 0.0;
-        bi8[1][4] = 0.0;
-        bi8[1][5] = 0.0;
-        bi8[1][6] = 0.0;
-        bi8[1][7] = 0.0;
-
-        bi8[2][0] = 0.0;
-        bi8[2][1] = 0.0;
-        bi8[2][2] = 0.0;
-        bi8[2][3] = 0.0;
-        bi8[2][4] = 0.0;
-        bi8[2][5] = 0.0;
-        bi8[2][6] = 0.0;
-        bi8[2][7] = 0.0;
-
-        bi8[3][0] = 0.0;
-        bi8[3][1] = 0.0;
-        bi8[3][2] = 0.0;
-        bi8[3][3] = 0.0;
-        bi8[3][4] = 0.0;
-        bi8[3][5] = 0.0;
-        bi8[3][6] = 0.0;
-        bi8[3][7] = 0.0;
-
-        bi8[4][0] = 0.0;
-        bi8[4][1] = 0.0;
-        bi8[4][2] = 0.0;
-        bi8[4][3] = 0.0;
-        bi8[4][4] = 0.0;
-        bi8[4][5] = 0.0;
-        bi8[4][6] = 0.0;
-        bi8[4][7] = 0.0;
-
-        bi8[5][0] = 0.0;
-        bi8[5][1] = 31.58105876966070755817711181026329343665;
-        bi8[5][2] = -180.5178006497936101340318821069462934370;
-        bi8[5][3] = 371.2428684175585341424791160692889410620;
-        bi8[5][4] = -219.9594432616132922400607592351592840556;
-        bi8[5][5] = -240.0938376157735211521838900097427201378;
-        bi8[5][6] = 371.1625149095733967216817534260019455117;
-        bi8[5][7] = -133.0446939530601137778175224155755435359;
-
-        bi8[6][0] = 0.0;
-        bi8[6][1] = 22.07073965292731709025653809370859921609;
-        bi8[6][2] = -126.1566754274917694565163606891720887307;
-        bi8[6][3] = 259.4468018507764496957034678807673289852;
-        bi8[6][4] = -153.7208629336292887173729422706050971616;
-        bi8[6][5] = -167.7919863592614905102403297532219205119;
-        bi8[6][6] = 259.3906459958421903925844375556299672132;
-        bi8[6][7] = -92.97961869670813383669527983250287622418;
-
-        bi8[7][0] = 0.0;
-        bi8[7][1] = -57935.13185200130004451750696526889385541;
-        bi8[7][2] = 331158.0735325498459302442841924636844465;
-        bi8[7][3] = -681041.2750173130138192297984810369904889;
-        bi8[7][4] = 403513.3666796717724078043933204165043790;
-        bi8[7][5] = 440449.7088136183995582652495030608971324;
-        bi8[7][6] = -680893.8673222815216844218479358007413730;
-        bi8[7][7] = 244069.1410189383137471951034009314976560;
-
-        bi8[8][0] = 0.0;
-        bi8[8][1] = 4250.800686968169496178700937981058363787;
-        bi8[8][2] = -24297.63981660104922968878449326985727293;
-        bi8[8][3] = 49969.17461226540555036109266936358732405;
-        bi8[8][4] = -29606.47264365440829329549714113549060954;
-        bi8[8][5] = -32316.55586083186219547193751752582986592;
-        bi8[8][6] = 49958.35905508492049348605670433958547126;
-        bi8[8][7] = -17907.77442194075528997858814914394503258;
-
-        bi8[9][0] = 0.0;
-        bi8[9][1] = 875116.5297933782663101143489237552187982;
-        bi8[9][2] = -5002179.072677031402026209959119949939645;
-        bi8[9][3] = 10287203.25970261764886749823989225303597;
-        bi8[9][4] = -6095113.722637695183800934414417711613054;
-        bi8[9][5] = -6653041.227387194363013444576409171926027;
-        bi8[9][6] = 10284976.65027898959354116589597477678226;
-        bi8[9][7] = -3686691.181850927247464301690167065077119;
-
-        bi8[10][0] = 0.0;
-        bi8[10][1] = -1259453.640882216588730059128817651829025;
-        bi8[10][2] = 7199055.703947677084639189578415812070854;
-        bi8[10][3] = -14805177.54931099282922608063380775248053;
-        bi8[10][4] = 8771989.681625245671776936305003143572362;
-        bi8[10][5] = 9574949.976948425063370422937920428882522;
-        bi8[10][6] = -14801973.04882455689422764622042624383139;
-        bi8[10][7] = 5305826.679890061595116647056008036628796;
-
-        bi8[11][0] = 0.0;
-        bi8[11][1] = 438049.0501694909633920945145914847818354;
-        bi8[11][2] = -2503894.872242029083881453081815093438434;
-        bi8[11][3] = 5149370.927637711155082091178724257279041;
-        bi8[11][4] = -3050975.139855797649176572488872351115466;
-        bi8[11][5] = -3330251.790676982456969584764119587826344;
-        bi8[11][6] = 5148256.374192641390109800793347658427157;
-        bi8[11][7] = -1845413.171271418254817053627873630602405;
-
-        bi8[12][0] = 0.0;
-        bi8[12][1] = 0.0;
-        bi8[12][2] = 0.0;
-        bi8[12][3] = 0.0;
-        bi8[12][4] = 0.0;
-        bi8[12][5] = 0.0;
-        bi8[12][6] = 0.0;
-        bi8[12][7] = 0.0;
-
-        bi8[13][0] = 0.0;
-        bi8[13][1] = -2.504431470815825952597052601849741871849;
-        bi8[13][2] = 10.02051593444700250034144043591917565446;
-        bi8[13][3] = 3.188314791667844930562547327020218859370;
-        bi8[13][4] = -81.28919087456095909677167863874787591733;
-        bi8[13][5] = 166.5016415475332861729532713048075758880;
-        bi8[13][6] = -137.9649604853951315639665825699498265027;
-        bi8[13][7] = 42.04811055712378300947805474280047389006;
-
-        bi8[14][0] = 0.0;
-        bi8[14][1] = -8.781714078400345890500582254933426660729;
-        bi8[14][2] = 95.92683681317881690739004713651739721760;
-        bi8[14][3] = -437.8684018822059554032097156332107532369;
-        bi8[14][4] = 1022.893757817680688328813133512141573185;
-        bi8[14][5] = -1279.306771651130473378088029252960844151;
-        bi8[14][6] = 814.4619777825506941886287783493535555706;
-        bi8[14][7] = -207.3256848016734247530336318569075019247;
-
-        bi8[15][0] = 0.0;
-        bi8[15][1] = 23.68361088778252813755830682525871572206;
-        bi8[15][2] = -258.7073385100442622636796566042807424535;
-        bi8[15][3] = 1180.897573941780652808280782076758850644;
-        bi8[15][4] = -2758.666192432960487581886381121050830153;
-        bi8[15][5] = 3450.192469874827074026059813408271639164;
-        bi8[15][6] = -2196.541630994369154046535169690574772508;
-        bi8[15][7] = 559.1415072329836489202023051056171395846;
-
-        bi8[16][0] = 0.0;
-        bi8[16][1] = -40.60495590126188439710688778738077499754;
-        bi8[16][2] = 443.5472327808004803148884291518351973556;
-        bi8[16][3] = -2024.619224703986215542471608189157701003;
-        bi8[16][4] = 4729.663885325312770086552887900617076925;
-        bi8[16][5] = -5915.268315880109317019855761487629399742;
-        bi8[16][6] = 3765.915446103794090758138472995147035931;
-        bi8[16][7] = -958.6340677245499242001455325834314344682;
-
-        bi8[17][0] = 0.0;
-        bi8[17][1] = -46.25448592533055837377842531202839839512;
-        bi8[17][2] = 439.1938466041217159655702274089786458299;
-        bi8[17][3] = -1719.438253681139282550356858145413164525;
-        bi8[17][4] = 3520.667063097926217589621564937806803955;
-        bi8[17][5] = -3962.634989140663522153677993129178399151;
-        bi8[17][6] = 2322.579486243480292154738138853866205762;
-        bi8[17][7] = -554.1126671983948626321166546140316934748;
-
-        bi8[18][0] = 0.0;
-        bi8[18][1] = 14.23775221010393402969945815961053497020;
-        bi8[18][2] = -276.2932724089639106808497191835732392038;
-        bi8[18][3] = 1558.151943742051093976689365840541702310;
-        bi8[18][4] = -3984.134381318251874989648181935067360669;
-        bi8[18][5] = 5195.127830395725191831019434328633823950;
-        bi8[18][6] = -3374.420443020703185373799939254029388933;
-        bi8[18][7] = 867.3305704000387512068895820438839275766;
-
-        bi8[19][0] = 0.0;
-        bi8[19][1] = 6.177962769210591542599576397764741591492;
-        bi8[19][2] = -157.4910701175524827231177916800491606577;
-        bi8[19][3] = 978.6380890625254126165941312401420957681;
-        bi8[19][4] = -2692.299557325995918275212259001422935846;
-        bi8[19][5] = 3734.579314715993155267252453495657367624;
-        bi8[19][6] = -2556.424739731601341815253154933956957044;
-        bi8[19][7] = 686.8200006274205833871370444818648485639;
-
-        bi8[20][0] = 0.0;
-        bi8[20][1] = -18.65289021588136658934344739820379525269;
-        bi8[20][2] = 130.3737772179961353812355151043116833235;
-        bi8[20][3] = -398.2673561235244947185709369394563871567;
-        bi8[20][4] = 665.1759551798239720897702016127403918291;
-        bi8[20][5] = -644.5526670389200969746336450782179076782;
-        bi8[20][6] = 346.6953482377737791841235419437045433947;
-        bi8[20][7] = -80.77216725726792837258122924487852845967;
-
-        Self {
-            c,
-            a,
-            b,
-            bh: Some(bh),
-            bi: Some(bi8),
-        }
-    }
-}
