--- src/tableau/rkv98.rs
@@ -1,1267 +1,0 @@
-use crate::tableau::{
-    ButcherTableau,
-};
-
-impl ButcherTableau<16, 21> {
-    /// A better efficient Runge-Kutta (16:8(9)) pair with 8th-order interpolation.
-    ///
-    /// # Overview
-    /// This provides a 16-stage Runge-Kutta method with:
-    /// - Primary order: 8
-    /// - Embedded order: 9 (for error estimation)
-    /// - Number of stages: 16 primary + 5 additional stages for interpolation
-    /// - Dominant stage order: 5
-    ///
-    /// # Efficiency
-    /// This method is particularly efficient with a local truncation error 2-norm:
-    ///
-    /// T_10,2 ~ 0.00000034399
-    ///
-    /// This 2-norm is smaller than previous (16:8,9) pairs and has two local maximum values on [0,1].
-    ///
-    /// # Interpolation
-    /// - Nodes c[17]=1 and c[18]-c[21] were selected to minimize interpolation error.
-    /// - The interpolant has continuous derivatives.
-    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
-    ///
-    ///   Ti_9,2 ~ 0.000003721
-    ///
-    /// - The interpolation error has three local maximum values on [0,1].
-    ///
-    /// # Notes
-    /// - Exact coefficients of the method are possible using surds in terms of sqrt(6).
-    /// - This method shares the same primary stages as the 9th-order method (rkv989e).
-    /// - The interpolant requires fewer additional stages than the 9th-order version.
-    ///
-    /// # References
-    /// - J.H. Verner, SIAM J NA 1978, "Explicit Runge-Kutta methods with estimates of the Local Truncation Error"
-    /// - J.H. Verner, SIAM J NA 1993, "Differentiable Interpolants for high-order Runge-Kutta methods"
-    pub const fn rkv988e() -> Self {
-        let mut c = [0.0; 21];
-        let mut a = [[0.0; 21]; 21];
-        let mut b = [0.0; 16];
-        let mut bh = [0.0; 16];
-        let mut bi8 = [[0.0; 21]; 21];
-
-        c[0] = 0.0;
-        c[1] = 0.3571e-1;
-        c[2] = 0.9906028091267414072062290548990445721420e-1;
-        c[3] = 0.1485904213690112110809343582348566858213;
-        c[4] = 0.6134;
-        c[5] = 0.2327359473605626756631680289030288192566;
-        c[6] = 0.5538640526394373243368319710969711807434;
-        c[7] = 0.6555;
-        c[8] = 0.491625;
-        c[9] = 0.6858e-1;
-        c[10] = 0.253;
-        c[11] = 0.6620641795412045944786226689660879617864;
-        c[12] = 0.8309;
-        c[13] = 0.8998;
-        c[14] = 1.0;
-        c[15] = 1.0;
-
-        a[1][0] = 0.3571e-1;
-
-        a[2][0] = -0.3833735636677017025757228807792426014327e-1;
-        a[2][1] = 0.1373976372794443109781951935678287173575;
-
-        a[3][0] = 0.3714760534225280277023358955871417145532e-1;
-        a[3][1] = 0.0;
-        a[3][2] = 0.1114428160267584083107007686761425143660;
-
-        a[4][0] = 2.674764429871505119194043347886848033362;
-        a[4][1] = 0.0;
-        a[4][2] = -9.982382134885293836441318015951354323506;
-        a[4][3] = 7.921017705013788717247274668064506290145;
-
-        a[5][0] = 0.5242104050577351069841401193318449756747e-1;
-        a[5][1] = 0.0;
-        a[5][2] = 0.0;
-        a[5][3] = 0.1796911189175953081279466649977364298948;
-        a[5][4] = 0.6237879371938568368073519721078917943307e-3;
-
-        a[6][0] = 0.1592492223647632083060991150129732726703;
-        a[6][1] = 0.0;
-        a[6][2] = 0.0;
-        a[6][3] = -0.4298429877241087508189185066909803751595;
-        a[6][4] = 0.6665266542726088051243445942394404211672e-1;
-        a[6][5] = 0.7578051525715219863372169033510342411158;
-
-        a[7][0] = 0.7283333333333333333333333333333333333333e-1;
-        a[7][1] = 0.0;
-        a[7][2] = 0.0;
-        a[7][3] = 0.0;
-        a[7][4] = 0.0;
-        a[7][5] = 0.3359344590665103678713422141936031057620;
-        a[7][6] = 0.2467322076001562987953244524730635609046;
-
-        a[8][0] = 0.7297558593750000000000000000000000000000e-1;
-        a[8][1] = 0.0;
-        a[8][2] = 0.0;
-        a[8][3] = 0.0;
-        a[8][4] = 0.0;
-        a[8][5] = 0.3348009729699333533129043555243262838993;
-        a[8][6] = 0.1184158239050666466870956444756737161007;
-        a[8][7] = -0.3456738281250000000000000000000000000000e-1;
-
-        a[9][0] = 0.4911213663452096382929799914888692571185e-1;
-        a[9][1] = 0.0;
-        a[9][2] = 0.0;
-        a[9][3] = 0.0;
-        a[9][4] = 0.0;
-        a[9][5] = 0.3983857361308652347066808423646954244674e-1;
-        a[9][6] = 0.1069675288939354812077139476183615284408;
-        a[9][7] = -0.2174259165458647598455650055976622663086e-1;
-        a[9][8] = -0.1055956474869564925231235304439517699685;
-
-        a[10][0] = -0.2707988818641280489175862864774930636734e-1;
-        a[10][1] = 0.0;
-        a[10][2] = 0.0;
-        a[10][3] = 0.0;
-        a[10][4] = 0.0;
-        a[10][5] = 0.3330000000000000000000000000000000000000e-1;
-        a[10][6] = -0.1645526070036057075627040519129748090051;
-        a[10][7] = 0.3428266306497389946775103799535438271025e-1;
-        a[10][8] = 0.1585264064439221046855989221358760453749;
-        a[10][9] = 0.2185234256811225083011127204294936872873;
-
-        a[11][0] = 0.5584657769108862727526281006220448941548e-1;
-        a[11][1] = 0.0;
-        a[11][2] = 0.0;
-        a[11][3] = 0.0;
-        a[11][4] = 0.0;
-        a[11][5] = 0.9166533166672539087479545790553151183505e-1;
-        a[11][6] = 0.2392399655523627049569179599100121305027;
-        a[11][7] = 0.1023834712248414879740316672186561113729e-1;
-        a[11][8] = -0.2679331322859542442991625976085644567159e-2;
-        a[11][9] = 0.4235624181474284646680744174197106205439e-1;
-        a[11][10] = 0.2253970470166604185504274586005888014087;
-
-        a[12][0] = -0.4802510512725195893103945790000307739284;
-        a[12][1] = 0.0;
-        a[12][2] = 0.0;
-        a[12][3] = 0.0;
-        a[12][4] = 0.0;
-        a[12][5] = -6.359610162555930097843606206238737368598;
-        a[12][6] = -0.2762313898040841385163666588363199062208;
-        a[12][7] = -6.500796633979846747011407331804389650124;
-        a[12][8] = 0.5734765877040956875980719881692630407295;
-        a[12][9] = 1.347125994868138849716719646646049357523;
-        a[12][10] = 5.936840409706221306933249144503479736029;
-        a[12][11] = 6.590346245333924728433733996560685564589;
-
-        a[13][0] = 0.3307533067671401079975082476263855100019;
-        a[13][1] = 0.0;
-        a[13][2] = 0.0;
-        a[13][3] = 0.0;
-        a[13][4] = 0.0;
-        a[13][5] = 5.956207776829962111958631943398154563177;
-        a[13][6] = -0.4868316400481527952968983900652017328476;
-        a[13][7] = 4.462055288206771191616083778225319083232;
-        a[13][8] = 0.7410258231442071778525406966672954745201;
-        a[13][9] = -0.7118192034575913119318346652024763899638;
-        a[13][10] = -5.454619594516665440649148499095762263275;
-        a[13][11] = -4.140803729244709693305868672684077643030;
-        a[13][12] = 0.2038319723190386517589855611303633981843;
-
-        a[14][0] = -0.5847111122998944389651719860007729221315;
-        a[14][1] = 0.0;
-        a[14][2] = 0.0;
-        a[14][3] = 0.0;
-        a[14][4] = 0.0;
-        a[14][5] = -12.41268417116267068926060969930886386670;
-        a[14][6] = 1.360245445660928146977003413153919520189;
-        a[14][7] = -22.42610531111868229005965344123311160860;
-        a[14][8] = -0.8828857055865458252669099251378431470909;
-        a[14][9] = 1.770155128538230466813900572142546329490;
-        a[14][10] = 12.15809651918533877218739262890752294243;
-        a[14][11] = 22.23037520407760703834156132941550000589;
-        a[14][12] = -0.6634483760201249006317401316871863808959;
-        a[14][13] = 0.4509623787258137198642272397482891274306;
-
-        a[15][0] = 1.940575549810648717395482898478643635949;
-        a[15][1] = 0.0;
-        a[15][2] = 0.0;
-        a[15][3] = 0.0;
-        a[15][4] = 0.0;
-        a[15][5] = 21.97798408114556310164832153356570429956;
-        a[15][6] = 0.8230747326984728145128172720328075817426;
-        a[15][7] = 68.16441683626354817206843139989369741474;
-        a[15][8] = -3.117097463620266666801525529574905518624;
-        a[15][9] = -4.568841021822439620303730003755793384193;
-        a[15][10] = -18.74190987126264955904919353670543412590;
-        a[15][11] = -66.57711839637831878350036856645874492472;
-        a[15][12] = 1.098915553165441824029764532524025021446;
-        a[15][13] = 0.0;
-        a[15][14] = 0.0;
-
-        b[0] = 0.1500669014979724795766288712377040980022e-1;
-        b[1] = 0.0;
-        b[2] = 0.0;
-        b[3] = 0.0;
-        b[4] = 0.0;
-        b[5] = 0.0;
-        b[6] = 0.0;
-        b[7] = -1.055180992746381278594381685080184474710;
-        b[8] = 0.2384947263782183112638140557774851953952;
-        b[9] = 0.1288151774282991354622515887144081172371;
-        b[10] = 0.2276623111046215614614917896665746508574;
-        b[11] = 1.229532587437517443160321815173576421896;
-        b[12] = 0.4624976662810383487397308275256128637703e-1;
-        b[13] = 0.1386196319366293903628308621201843492468;
-        b[14] = 0.3080010168319435405203560375162404389992e-1;
-        b[15] = 0.0;
-
-        bh[0] = 0.1897210532481101330735875918987801428423e-1;
-        bh[1] = 0.0;
-        bh[2] = 0.0;
-        bh[3] = 0.0;
-        bh[4] = 0.0;
-        bh[5] = 0.0;
-        bh[6] = 0.0;
-        bh[7] = 3.408110314549493848404398964228060776034;
-        bh[8] = 0.1260323883820920906560270507988839661845;
-        bh[9] = 0.1188375063451149770930378540709095182997;
-        bh[10] = 0.2491041997838687569190073177760326138188;
-        bh[11] = -3.269966219928978218713853055139116510335;
-        bh[12] = 0.3023798100228882907409723963501699715077;
-        bh[13] = 0.0;
-        bh[14] = 0.0;
-        bh[15] = 0.4652989552070924159305071272518165020637e-1;
-
-        c[16] = 1.0;
-
-        a[16][0] = 0.1500669014979724795766288712377040980022e-1;
-        a[16][1] = 0.0;
-        a[16][2] = 0.0;
-        a[16][3] = 0.0;
-        a[16][4] = 0.0;
-        a[16][5] = 0.0;
-        a[16][6] = 0.0;
-        a[16][7] = -1.055180992746381278594381685080184474710;
-        a[16][8] = 0.2384947263782183112638140557774851953952;
-        a[16][9] = 0.1288151774282991354622515887144081172371;
-        a[16][10] = 0.2276623111046215614614917896665746508574;
-        a[16][11] = 1.229532587437517443160321815173576421896;
-        a[16][12] = 0.4624976662810383487397308275256128637703e-1;
-        a[16][13] = 0.1386196319366293903628308621201843492468;
-        a[16][14] = 0.3080010168319435405203560375162404389992e-1;
-        a[16][15] = 0.0;
-
-        c[17] = 0.7375018139988810429214526568476200466933;
-
-        a[17][0] = 0.1571801061417788021580499748292804743973e-1;
-        a[17][1] = 0.0;
-        a[17][2] = 0.0;
-        a[17][3] = 0.0;
-        a[17][4] = 0.0;
-        a[17][5] = 0.0;
-        a[17][6] = 0.0;
-        a[17][7] = 0.4853403452657363080900487157345670578815;
-        a[17][8] = 0.2107787568904546684574678306538806242458;
-        a[17][9] = 0.1269802413053354152906551816812865612400;
-        a[17][10] = 0.2319687014513919157472314236852033704833;
-        a[17][11] = -0.3620214714069096774945761015915844860748;
-        a[17][12] = 0.5366106712036344076786687672191174826870e-1;
-        a[17][13] = -0.2806066613385549629370667652641771053798e-1;
-        a[17][14] = -0.2378121372710330174923994809623697280793e-1;
-        a[17][15] = 0.0;
-        a[17][16] = 0.2691804261928988988990035710208180655494e-1;
-
-        c[18] = 0.749;
-
-        a[18][0] = 0.1569705832522204237038521016179271074328e-1;
-        a[18][1] = 0.0;
-        a[18][2] = 0.0;
-        a[18][3] = 0.0;
-        a[18][4] = 0.0;
-        a[18][5] = 0.0;
-        a[18][6] = 0.0;
-        a[18][7] = 0.4616075242202111936044636619654233606961;
-        a[18][8] = 0.2113946516698113234089715994366982070261;
-        a[18][9] = 0.1270330917167109232354671760560477406684;
-        a[18][10] = 0.2318540550298708286361100159494550129997;
-        a[18][11] = -0.3385266406688372942274816228586594758312;
-        a[18][12] = 0.5298251972194235542699858758068190007453e-1;
-        a[18][13] = -0.2750461365887187780272257875389040879148e-1;
-        a[18][14] = -0.2361906185395527177468217904834875318962e-1;
-        a[18][15] = 0.0;
-        a[18][16] = 0.2668458089504036186694391521145359328377e-1;
-        a[18][17] = 0.1139683460285541525554621429934611232038e-1;
-
-        c[19] = 0.65;
-
-        a[19][0] = 0.1438964884291216402951708307100553583630e-1;
-        a[19][1] = 0.0;
-        a[19][2] = 0.0;
-        a[19][3] = 0.0;
-        a[19][4] = 0.0;
-        a[19][5] = 0.0;
-        a[19][6] = 0.0;
-        a[19][7] = -1.206901219123788519767229935392578222389;
-        a[19][8] = 0.2505628554639376294459724790398627595052;
-        a[19][9] = 0.1303332915702127271150218433265955255131;
-        a[19][10] = 0.2246717759263522068724895908997148014087;
-        a[19][11] = 1.308419325781946480763221457213122900404;
-        a[19][12] = 0.2589750180376235922034567811002287469857e-2;
-        a[19][13] = 0.8070743254562856892787425005539711000810e-2;
-        a[19][14] = -0.1267568255392829384757055330207124946149e-1;
-        a[19][15] = 0.0;
-        a[19][16] = 0.1129158072373321675992773077865249355325e-1;
-        a[19][17] = 0.3422056680709749787491027385482635813798e-1;
-        a[19][18] = -0.1149726368734142020610819623056729009783;
-
-        c[20] = 0.487;
-
-        a[20][0] = 0.1452348029801041972968851600082942318526e-1;
-        a[20][1] = 0.0;
-        a[20][2] = 0.0;
-        a[20][3] = 0.0;
-        a[20][4] = 0.0;
-        a[20][5] = 0.0;
-        a[20][6] = 0.0;
-        a[20][7] = -0.5214243102465818231380106106018875132310;
-        a[20][8] = 0.1866698844204603947548300448071067931901;
-        a[20][9] = 0.1299316354451273004594203171440036506875;
-        a[20][10] = 0.2262141085765719298250459612902177827497;
-        a[20][11] = 0.6104582639466710747030070634860296980845;
-        a[20][12] = 0.1418715607022412453299982942636029187870e-1;
-        a[20][13] = 0.1480061054412245756808618945399364705863e-1;
-        a[20][14] = -0.3711471609871774475268028540581604884370e-2;
-        a[20][15] = 0.0;
-        a[20][16] = 0.1393256979572558890317889514899316376009e-2;
-        a[20][17] = 1.114731020624073270253135239909350910479;
-        a[20][18] = -1.021208555757145859230933049243060207745;
-        a[20][19] = -0.2795650792912340738723193626472621878286;
-
-        bi8[0][0] = 1.0;
-        bi8[0][1] = -11.00917227309954827048799176229874144363;
-        bi8[0][2] = 53.60376844185445526077321090756616710855;
-        bi8[0][3] = -143.7818601051773394006248841909965394832;
-        bi8[0][4] = 227.7802272724059672847516567630398210649;
-        bi8[0][5] = -212.7667774745397515281313247919614156196;
-        bi8[0][6] = 108.4765585030943501072304151917413023749;
-        bi8[0][7] = -23.28773767438833620555341922996682359208;
-
-        bi8[1][0] = 0.0;
-        bi8[1][1] = 0.0;
-        bi8[1][2] = 0.0;
-        bi8[1][3] = 0.0;
-        bi8[1][4] = 0.0;
-        bi8[1][5] = 0.0;
-        bi8[1][6] = 0.0;
-        bi8[1][7] = 0.0;
-
-        bi8[2][0] = 0.0;
-        bi8[2][1] = 0.0;
-        bi8[2][2] = 0.0;
-        bi8[2][3] = 0.0;
-        bi8[2][4] = 0.0;
-        bi8[2][5] = 0.0;
-        bi8[2][6] = 0.0;
-        bi8[2][7] = 0.0;
-
-        bi8[3][0] = 0.0;
-        bi8[3][1] = 0.0;
-        bi8[3][2] = 0.0;
-        bi8[3][3] = 0.0;
-        bi8[3][4] = 0.0;
-        bi8[3][5] = 0.0;
-        bi8[3][6] = 0.0;
-        bi8[3][7] = 0.0;
-
-        bi8[4][0] = 0.0;
-        bi8[4][1] = 0.0;
-        bi8[4][2] = 0.0;
-        bi8[4][3] = 0.0;
-        bi8[4][4] = 0.0;
-        bi8[4][5] = 0.0;
-        bi8[4][6] = 0.0;
-        bi8[4][7] = 0.0;
-
-        bi8[5][0] = 0.0;
-        bi8[5][1] = 0.0;
-        bi8[5][2] = 0.0;
-        bi8[5][3] = 0.0;
-        bi8[5][4] = 0.0;
-        bi8[5][5] = 0.0;
-        bi8[5][6] = 0.0;
-        bi8[5][7] = 0.0;
-
-        bi8[6][0] = 0.0;
-        bi8[6][1] = 0.0;
-        bi8[6][2] = 0.0;
-        bi8[6][3] = 0.0;
-        bi8[6][4] = 0.0;
-        bi8[6][5] = 0.0;
-        bi8[6][6] = 0.0;
-        bi8[6][7] = 0.0;
-
-        bi8[7][0] = 0.0;
-        bi8[7][1] = 179.0879062178872372197637635178056034051;
-        bi8[7][2] = -1728.403136057631342220019375555018665689;
-        bi8[7][3] = 6564.460940561142416685202971494559376099;
-        bi8[7][4] = -12902.91081584121105690497909548869952323;
-        bi8[7][5] = 13965.17977100265059245172194550171929806;
-        bi8[7][6] = -7920.424061687375443376559247327959298606;
-        bi8[7][7] = 1841.954214811791214866274656172513025486;
-
-        bi8[8][0] = 0.0;
-        bi8[8][1] = -22.88304714767390548521469744522669995060;
-        bi8[8][2] = 238.8040507685294155366494768970788101399;
-        bi8[8][3] = -936.8585782296073752493303117453241744494;
-        bi8[8][4] = 1874.044957762151353867893217234671921697;
-        bi8[8][5] = -2050.511559353009354817713813307430993559;
-        bi8[8][6] = 1171.508545192416250747220534524328950392;
-        bi8[8][7] = -273.8658742664281662882405921023203290746;
-
-        bi8[9][0] = 0.0;
-        bi8[9][1] = 12.17331549628124801002349940562808097269;
-        bi8[9][2] = -82.75076476436870272146201057456216825170;
-        bi8[9][3] = 256.4804438751307755574323340365457074340;
-        bi8[9][4] = -441.0999429758310962594951067937304764418;
-        bi8[9][5] = 434.5079831401213158642449792005352310524;
-        bi8[9][6] = -229.8934605896900265488669050173042421551;
-        bi8[9][7] = 50.71124099578478523358546133160227550674;
-
-        bi8[10][0] = 0.0;
-        bi8[10][1] = 2.613203520133622756551032901538192914958;
-        bi8[10][2] = 16.88002937261184447692017262466190582959;
-        bi8[10][3] = -134.1727921155237789914884572421173077178;
-        bi8[10][4] = 340.1204801692637664465059992362156474712;
-        bi8[10][5] = -420.1258893776879400857436087457229872331;
-        bi8[10][6] = 258.3234372146557103657079225360621734929;
-        bi8[10][7] = -63.41080647234860340699156952097105010691;
-
-        bi8[11][0] = 0.0;
-        bi8[11][1] = -212.3127005913665992194796715806401319940;
-        bi8[11][2] = 2045.352316395154364712249302798824723141;
-        bi8[11][3] = -7762.060413644090906790449120787506530500;
-        bi8[11][4] = 15250.14841459378132543421843792712597249;
-        bi8[11][5] = -16501.08003544517859797013884904039476397;
-        bi8[11][6] = 9356.907363957304758100332957460543520093;
-        bi8[11][7] = -2175.725412678166826823572734962779212842;
-
-        bi8[12][0] = 0.0;
-        bi8[12][1] = -11.50164708567429764283826551320712022931;
-        bi8[12][2] = 107.2769704045725235564526349677807440972;
-        bi8[12][3] = -401.2345332983947627677963563088489537002;
-        bi8[12][4] = 781.8916989572930300323034641740782180816;
-        bi8[12][5] = -841.6593371013254430188454534530483533077;
-        bi8[12][6] = 475.5767391485588457657241425216173583523;
-        bi8[12][7] = -110.3036412584017920901261933056193320075;
-
-        bi8[13][0] = 0.0;
-        bi8[13][1] = -38.77239677340365800498371633092550416428;
-        bi8[13][2] = 358.6391180152970136912934982919377109538;
-        bi8[13][3] = -1336.215732140226920321570086014013189585;
-        bi8[13][4] = 2598.193515313187034150047120315374589407;
-        bi8[13][5] = -2792.878330051016777349083250990468387014;
-        bi8[13][6] = 1576.586790342810048230642938513691709160;
-        bi8[13][7] = -365.4143450747101110059836729234767444075;
-
-        bi8[14][0] = 0.0;
-        bi8[14][1] = -10.00425469998852674211886984302245249578;
-        bi8[14][2] = 91.67754241560498208460382885503842174321;
-        bi8[14][3] = -340.0777600652848208705042811235806913159;
-        bi8[14][4] = 659.6008369976978366087037536255754670281;
-        bi8[14][5] = -707.8826284112975671515477305189138171574;
-        bi8[14][6] = 399.1580030260127128211116842683795971037;
-        bi8[14][7] = -92.44093916106142239619634965972490086213;
-
-        bi8[15][0] = 0.0;
-        bi8[15][1] = 0.0;
-        bi8[15][2] = 0.0;
-        bi8[15][3] = 0.0;
-        bi8[15][4] = 0.0;
-        bi8[15][5] = 0.0;
-        bi8[15][6] = 0.0;
-        bi8[15][7] = 0.0;
-
-        bi8[16][0] = 0.0;
-        bi8[16][1] = 16.21014491337561130525562931390759162785;
-        bi8[16][2] = -150.4492131048662883122727122117594677554;
-        bi8[16][3] = 565.1196631722152490306979317160273112948;
-        bi8[16][4] = -1111.210282651553705321998904060968429058;
-        bi8[16][5] = 1211.236976000843526549958975517785065570;
-        bi8[16][6] = -695.3365606918091595268831805414223001360;
-        bi8[16][7] = 164.4292723617947662752422602664302284566;
-
-        bi8[17][0] = 0.0;
-        bi8[17][1] = -1872.805019318414830108065147723853146433;
-        bi8[17][2] = 16400.66633207048634386587617517712424379;
-        bi8[17][3] = -57270.16260166053287080817377636090729160;
-        bi8[17][4] = 103330.5912960663516254356794429889210991;
-        bi8[17][5] = -102051.1888173727465381067712242944167386;
-        bi8[17][6] = 52424.75260874662694445820923552319700594;
-        bi8[17][7] = -10961.85379853177067473675470531006517221;
-
-        bi8[18][0] = 0.0;
-        bi8[18][1] = 1720.562129057968827268151794567408887461;
-        bi8[18][2] = -15159.62046432254747381864100381352909840;
-        bi8[18][3] = 53314.59561935523492911472686991966671330;
-        bi8[18][4] = -96988.37468691576351426414353587220276161;
-        bi8[18][5] = 96690.39181023687073280170142510495743481;
-        bi8[18][6] = -50199.31248988246623378558547060878127074;
-        bi8[18][7] = 10621.75808247070273268378992070248009519;
-
-        bi8[19][0] = 0.0;
-        bi8[19][1] = 264.2470488396084725751013414957565817480;
-        bi8[19][2] = -2232.338003905684731509741541080134626580;
-        bi8[19][3] = 7453.753802419143998655808817000026688315;
-        bi8[19][4] = -12770.71881400916730561960440987151459948;
-        bi8[19][5] = 11903.25226767446224225175138830095252030;
-        bi8[19][6] = -5733.155576507225740169825158561334353021;
-        bi8[19][7] = 1114.959275488863063816509562716247788712;
-
-        bi8[20][0] = 0.0;
-        bi8[20][1] = -15.60551015563365366165870100287114141984;
-        bi8[20][2] = 40.66145427098759539731834271499129986994;
-        bi8[20][3] = 170.1538018759714061560683496064688819112;
-        bi8[20][4] = -848.0568847386052608898820401778869464852;
-        bi8[20][5] = 1373.524566531853560108596541516407906615;
-        bi8[20][6] = -993.1678967729130171884598684827601522494;
-        bi8[20][7] = 272.4904689883393700780173758256501517585;
-
-        Self {
-            c,
-            a,
-            b,
-            bh: Some(bh),
-            bi: Some(bi8),
-        }
-    }
-}
-
-impl ButcherTableau<16, 26> {
-    /// A better efficient Runge-Kutta (16:9(8)) pair with 9th-order interpolation.
-    ///
-    /// # Overview
-    /// This provides a 16-stage Runge-Kutta method with:
-    /// - Primary order: 9
-    /// - Embedded order: 8 (for error estimation)
-    /// - Number of stages: 16 primary + 10 additional stages for interpolation
-    /// - Dominant stage order: 5
-    ///
-    /// # Efficiency
-    /// This method is particularly efficient with a local truncation error 2-norm:
-    ///
-    /// T_10,2 ~ 0.00000034399
-    ///
-    /// This 2-norm is smaller than previous (16:8,9) pairs and has two local maximum values on [0,1].
-    ///
-    /// # Interpolation
-    /// - The method provides a 9th-order interpolant using additional nodes c[17] through c[25].
-    /// - The interpolant has continuous derivatives.
-    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
-    ///
-    ///   Ti_10,2 ~ 0.0000007830
-    ///
-    /// - The interpolation error has two local maximum values on [0,1].
-    ///
-    /// # Notes
-    /// - Exact coefficients of the method are possible using surds in terms of sqrt(6).
-    /// - Exact coefficients for the interpolants cannot be conveniently represented
-    ///   as they require c[18] to be a zero of a cubic polynomial with coefficients
-    ///   that are surds in terms of sqrt(6).
-    ///
-    /// # References
-    /// - J.H. Verner, SIAM J NA 1978, "Explicit Runge-Kutta methods with estimates of the Local Truncation Error"
-    /// - J.H. Verner, SIAM J NA 1993, "Differentiable Interpolants for high-order Runge-Kutta methods"
-    pub const fn rkv989e() -> Self {
-        let mut c = [0.0; 26];
-        let mut a = [[0.0; 26]; 26];
-        let mut b = [0.0; 16];
-        let mut bh = [0.0; 16];
-        let mut bi9 = [[0.0; 26]; 26];
-
-        c[0] = 0.0;
-        c[1] = 0.3571e-1;
-        c[2] = 0.9906028091267414072062290548990445721420e-1;
-        c[3] = 0.1485904213690112110809343582348566858213;
-        c[4] = 0.6134;
-        c[5] = 0.2327359473605626756631680289030288192566;
-        c[6] = 0.5538640526394373243368319710969711807434;
-        c[7] = 0.6555;
-        c[8] = 0.491625;
-        c[9] = 0.6858e-1;
-        c[10] = 0.253;
-        c[11] = 0.6620641795412045944786226689660879617864;
-        c[12] = 0.8309;
-        c[13] = 0.8998;
-        c[14] = 1.0;
-        c[15] = 1.0;
-
-        a[1][0] = 0.3571e-1;
-
-        a[2][0] = -0.3833735636677017025757228807792426014327e-1;
-        a[2][1] = 0.1373976372794443109781951935678287173575;
-
-        a[3][0] = 0.3714760534225280277023358955871417145532e-1;
-        a[3][1] = 0.0;
-        a[3][2] = 0.1114428160267584083107007686761425143660;
-
-        a[4][0] = 2.674764429871505119194043347886848033362;
-        a[4][1] = 0.0;
-        a[4][2] = -9.982382134885293836441318015951354323506;
-        a[4][3] = 7.921017705013788717247274668064506290145;
-
-        a[5][0] = 0.5242104050577351069841401193318449756747e-1;
-        a[5][1] = 0.0;
-        a[5][2] = 0.0;
-        a[5][3] = 0.1796911189175953081279466649977364298948;
-        a[5][4] = 0.6237879371938568368073519721078917943307e-3;
-
-        a[6][0] = 0.1592492223647632083060991150129732726703;
-        a[6][1] = 0.0;
-        a[6][2] = 0.0;
-        a[6][3] = -0.4298429877241087508189185066909803751595;
-        a[6][4] = 0.6665266542726088051243445942394404211672e-1;
-        a[6][5] = 0.7578051525715219863372169033510342411158;
-
-        a[7][0] = 0.7283333333333333333333333333333333333333e-1;
-        a[7][1] = 0.0;
-        a[7][2] = 0.0;
-        a[7][3] = 0.0;
-        a[7][4] = 0.0;
-        a[7][5] = 0.3359344590665103678713422141936031057620;
-        a[7][6] = 0.2467322076001562987953244524730635609046;
-
-        a[8][0] = 0.7297558593750000000000000000000000000000e-1;
-        a[8][1] = 0.0;
-        a[8][2] = 0.0;
-        a[8][3] = 0.0;
-        a[8][4] = 0.0;
-        a[8][5] = 0.3348009729699333533129043555243262838993;
-        a[8][6] = 0.1184158239050666466870956444756737161007;
-        a[8][7] = -0.3456738281250000000000000000000000000000e-1;
-
-        a[9][0] = 0.4911213663452096382929799914888692571185e-1;
-        a[9][1] = 0.0;
-        a[9][2] = 0.0;
-        a[9][3] = 0.0;
-        a[9][4] = 0.0;
-        a[9][5] = 0.3983857361308652347066808423646954244674e-1;
-        a[9][6] = 0.1069675288939354812077139476183615284408;
-        a[9][7] = -0.2174259165458647598455650055976622663086e-1;
-        a[9][8] = -0.1055956474869564925231235304439517699685;
-
-        a[10][0] = -0.2707988818641280489175862864774930636734e-1;
-        a[10][1] = 0.0;
-        a[10][2] = 0.0;
-        a[10][3] = 0.0;
-        a[10][4] = 0.0;
-        a[10][5] = 0.3330000000000000000000000000000000000000e-1;
-        a[10][6] = -0.1645526070036057075627040519129748090051;
-        a[10][7] = 0.3428266306497389946775103799535438271025e-1;
-        a[10][8] = 0.1585264064439221046855989221358760453749;
-        a[10][9] = 0.2185234256811225083011127204294936872873;
-
-        a[11][0] = 0.5584657769108862727526281006220448941548e-1;
-        a[11][1] = 0.0;
-        a[11][2] = 0.0;
-        a[11][3] = 0.0;
-        a[11][4] = 0.0;
-        a[11][5] = 0.9166533166672539087479545790553151183505e-1;
-        a[11][6] = 0.2392399655523627049569179599100121305027;
-        a[11][7] = 0.1023834712248414879740316672186561113729e-1;
-        a[11][8] = -0.2679331322859542442991625976085644567159e-2;
-        a[11][9] = 0.4235624181474284646680744174197106205439e-1;
-        a[11][10] = 0.2253970470166604185504274586005888014087;
-
-        a[12][0] = -0.4802510512725195893103945790000307739284;
-        a[12][1] = 0.0;
-        a[12][2] = 0.0;
-        a[12][3] = 0.0;
-        a[12][4] = 0.0;
-        a[12][5] = -6.359610162555930097843606206238737368598;
-        a[12][6] = -0.2762313898040841385163666588363199062208;
-        a[12][7] = -6.500796633979846747011407331804389650124;
-        a[12][8] = 0.5734765877040956875980719881692630407295;
-        a[12][9] = 1.347125994868138849716719646646049357523;
-        a[12][10] = 5.936840409706221306933249144503479736029;
-        a[12][11] = 6.590346245333924728433733996560685564589;
-
-        a[13][0] = 0.3307533067671401079975082476263855100019;
-        a[13][1] = 0.0;
-        a[13][2] = 0.0;
-        a[13][3] = 0.0;
-        a[13][4] = 0.0;
-        a[13][5] = 5.956207776829962111958631943398154563177;
-        a[13][6] = -0.4868316400481527952968983900652017328476;
-        a[13][7] = 4.462055288206771191616083778225319083232;
-        a[13][8] = 0.7410258231442071778525406966672954745201;
-        a[13][9] = -0.7118192034575913119318346652024763899638;
-        a[13][10] = -5.454619594516665440649148499095762263275;
-        a[13][11] = -4.140803729244709693305868672684077643030;
-        a[13][12] = 0.2038319723190386517589855611303633981843;
-
-        a[14][0] = -0.5847111122998944389651719860007729221315;
-        a[14][1] = 0.0;
-        a[14][2] = 0.0;
-        a[14][3] = 0.0;
-        a[14][4] = 0.0;
-        a[14][5] = -12.41268417116267068926060969930886386670;
-        a[14][6] = 1.360245445660928146977003413153919520189;
-        a[14][7] = -22.42610531111868229005965344123311160860;
-        a[14][8] = -0.8828857055865458252669099251378431470909;
-        a[14][9] = 1.770155128538230466813900572142546329490;
-        a[14][10] = 12.15809651918533877218739262890752294243;
-        a[14][11] = 22.23037520407760703834156132941550000589;
-        a[14][12] = -0.6634483760201249006317401316871863808959;
-        a[14][13] = 0.4509623787258137198642272397482891274306;
-
-        a[15][0] = 1.940575549810648717395482898478643635949;
-        a[15][1] = 0.0;
-        a[15][2] = 0.0;
-        a[15][3] = 0.0;
-        a[15][4] = 0.0;
-        a[15][5] = 21.97798408114556310164832153356570429956;
-        a[15][6] = 0.8230747326984728145128172720328075817426;
-        a[15][7] = 68.16441683626354817206843139989369741474;
-        a[15][8] = -3.117097463620266666801525529574905518624;
-        a[15][9] = -4.568841021822439620303730003755793384193;
-        a[15][10] = -18.74190987126264955904919353670543412590;
-        a[15][11] = -66.57711839637831878350036856645874492472;
-        a[15][12] = 1.098915553165441824029764532524025021446;
-        a[15][13] = 0.0;
-        a[15][14] = 0.0;
-
-        b[0] = 0.1500669014979724795766288712377040980022e-1;
-        b[1] = 0.0;
-        b[2] = 0.0;
-        b[3] = 0.0;
-        b[4] = 0.0;
-        b[5] = 0.0;
-        b[6] = 0.0;
-        b[7] = -1.055180992746381278594381685080184474710;
-        b[8] = 0.2384947263782183112638140557774851953952;
-        b[9] = 0.1288151774282991354622515887144081172371;
-        b[10] = 0.2276623111046215614614917896665746508574;
-        b[11] = 1.229532587437517443160321815173576421896;
-        b[12] = 0.4624976662810383487397308275256128637703e-1;
-        b[13] = 0.1386196319366293903628308621201843492468;
-        b[14] = 0.3080010168319435405203560375162404389992e-1;
-        b[15] = 0.0;
-
-        bh[0] = 0.1897210532481101330735875918987801428423e-1;
-        bh[1] = 0.0;
-        bh[2] = 0.0;
-        bh[3] = 0.0;
-        bh[4] = 0.0;
-        bh[5] = 0.0;
-        bh[6] = 0.0;
-        bh[7] = 3.408110314549493848404398964228060776034;
-        bh[8] = 0.1260323883820920906560270507988839661845;
-        bh[9] = 0.1188375063451149770930378540709095182997;
-        bh[10] = 0.2491041997838687569190073177760326138188;
-        bh[11] = -3.269966219928978218713853055139116510335;
-        bh[12] = 0.3023798100228882907409723963501699715077;
-        bh[13] = 0.0;
-        bh[14] = 0.0;
-        bh[15] = 0.4652989552070924159305071272518165020637e-1;
-
-        c[16] = 1.0;
-
-        a[16][0] = 0.1500669014979724795766288712377040980022e-1;
-        a[16][1] = 0.0;
-        a[16][2] = 0.0;
-        a[16][3] = 0.0;
-        a[16][4] = 0.0;
-        a[16][5] = 0.0;
-        a[16][6] = 0.0;
-        a[16][7] = -1.055180992746381278594381685080184474710;
-        a[16][8] = 0.2384947263782183112638140557774851953952;
-        a[16][9] = 0.1288151774282991354622515887144081172371;
-        a[16][10] = 0.2276623111046215614614917896665746508574;
-        a[16][11] = 1.229532587437517443160321815173576421896;
-        a[16][12] = 0.4624976662810383487397308275256128637703e-1;
-        a[16][13] = 0.1386196319366293903628308621201843492468;
-        a[16][14] = 0.3080010168319435405203560375162404389992e-1;
-        a[16][15] = 0.0;
-
-        c[17] = 0.7375018139988810429214526568476200466933;
-
-        a[17][0] = 0.1571801061417788021580499748292804743973e-1;
-        a[17][1] = 0.0;
-        a[17][2] = 0.0;
-        a[17][3] = 0.0;
-        a[17][4] = 0.0;
-        a[17][5] = 0.0;
-        a[17][6] = 0.0;
-        a[17][7] = 0.4853403452657363080900487157345670578815;
-        a[17][8] = 0.2107787568904546684574678306538806242458;
-        a[17][9] = 0.1269802413053354152906551816812865612400;
-        a[17][10] = 0.2319687014513919157472314236852033704833;
-        a[17][11] = -0.3620214714069096774945761015915844860748;
-        a[17][12] = 0.5366106712036344076786687672191174826870e-1;
-        a[17][13] = -0.2806066613385549629370667652641771053798e-1;
-        a[17][14] = -0.2378121372710330174923994809623697280793e-1;
-        a[17][15] = 0.0;
-        a[17][16] = 0.2691804261928988988990035710208180655494e-1;
-
-        c[18] = 0.749;
-
-        a[18][0] = 0.1569705832522204237038521016179271074328e-1;
-        a[18][1] = 0.0;
-        a[18][2] = 0.0;
-        a[18][3] = 0.0;
-        a[18][4] = 0.0;
-        a[18][5] = 0.0;
-        a[18][6] = 0.0;
-        a[18][7] = 0.4616075242202111936044636619654233606961;
-        a[18][8] = 0.2113946516698113234089715994366982070261;
-        a[18][9] = 0.1270330917167109232354671760560477406684;
-        a[18][10] = 0.2318540550298708286361100159494550129997;
-        a[18][11] = -0.3385266406688372942274816228586594758312;
-        a[18][12] = 0.5298251972194235542699858758068190007453e-1;
-        a[18][13] = -0.2750461365887187780272257875389040879148e-1;
-        a[18][14] = -0.2361906185395527177468217904834875318962e-1;
-        a[18][15] = 0.0;
-        a[18][16] = 0.2668458089504036186694391521145359328377e-1;
-        a[18][17] = 0.1139683460285541525554621429934611232038e-1;
-
-        c[19] = 0.65;
-
-        a[19][0] = 0.1438964884291216402951708307100553583630e-1;
-        a[19][1] = 0.0;
-        a[19][2] = 0.0;
-        a[19][3] = 0.0;
-        a[19][4] = 0.0;
-        a[19][5] = 0.0;
-        a[19][6] = 0.0;
-        a[19][7] = -1.206901219123788519767229935392578222389;
-        a[19][8] = 0.2505628554639376294459724790398627595052;
-        a[19][9] = 0.1303332915702127271150218433265955255131;
-        a[19][10] = 0.2246717759263522068724895908997148014087;
-        a[19][11] = 1.308419325781946480763221457213122900404;
-        a[19][12] = 0.2589750180376235922034567811002287469857e-2;
-        a[19][13] = 0.8070743254562856892787425005539711000810e-2;
-        a[19][14] = -0.1267568255392829384757055330207124946149e-1;
-        a[19][15] = 0.0;
-        a[19][16] = 0.1129158072373321675992773077865249355325e-1;
-        a[19][17] = 0.3422056680709749787491027385482635813798e-1;
-        a[19][18] = -0.1149726368734142020610819623056729009783;
-
-        c[20] = 0.487;
-
-        a[20][0] = 0.1452348029801041972968851600082942318526e-1;
-        a[20][1] = 0.0;
-        a[20][2] = 0.0;
-        a[20][3] = 0.0;
-        a[20][4] = 0.0;
-        a[20][5] = 0.0;
-        a[20][6] = 0.0;
-        a[20][7] = -0.5214243102465818231380106106018875132310;
-        a[20][8] = 0.1866698844204603947548300448071067931901;
-        a[20][9] = 0.1299316354451273004594203171440036506875;
-        a[20][10] = 0.2262141085765719298250459612902177827497;
-        a[20][11] = 0.6104582639466710747030070634860296980845;
-        a[20][12] = 0.1418715607022412453299982942636029187870e-1;
-        a[20][13] = 0.1480061054412245756808618945399364705863e-1;
-        a[20][14] = -0.3711471609871774475268028540581604884370e-2;
-        a[20][15] = 0.0;
-        a[20][16] = 0.1393256979572558890317889514899316376009e-2;
-        a[20][17] = 1.114731020624073270253135239909350910479;
-        a[20][18] = -1.021208555757145859230933049243060207745;
-        a[20][19] = -0.2795650792912340738723193626472621878286;
-
-        c[21] = 0.97e-2;
-
-        a[21][0] = 0.8711816186418632397551552134391273160887e-2;
-        a[21][1] = 0.0;
-        a[21][2] = 0.0;
-        a[21][3] = 0.0;
-        a[21][4] = 0.0;
-        a[21][5] = 0.0;
-        a[21][6] = 0.0;
-        a[21][7] = 0.1532993247326560354009297246756198757646e-1;
-        a[21][8] = -0.1943250606288015181452523235039064995331e-2;
-        a[21][9] = 0.1072095950570478409673351209542485874475e-2;
-        a[21][10] = 0.2601233036074381124605301648955068308034e-3;
-        a[21][11] = -0.1817718521410219268803888649170562657549e-1;
-        a[21][12] = -0.9877668338996712912794853387601975290298e-3;
-        a[21][13] = -0.3332383192417756458522769371199819990865e-2;
-        a[21][14] = -0.8605833352714280922890371244764817761631e-3;
-        a[21][15] = 0.0;
-        a[21][16] = 0.1392810143886650671629507962655268250424e-2;
-        a[21][17] = -0.1617419987897630058856682050471490927435;
-        a[21][18] = 0.1485156561306044255987175208430691701911;
-        a[21][19] = 0.2289051095253062003982633508518682297336e-1;
-        a[21][20] = -0.1429777169141779172700863258972231247118e-2;
-
-        c[22] = 0.138;
-
-        a[22][0] = 0.2710092628714179537584868171073201523422e-1;
-        a[22][1] = 0.0;
-        a[22][2] = 0.0;
-        a[22][3] = 0.0;
-        a[22][4] = 0.0;
-        a[22][5] = 0.0;
-        a[22][6] = 0.0;
-        a[22][7] = 0.6923104986003885114772983986172611691902;
-        a[22][8] = -0.6725243132164494445073150422370697609664e-1;
-        a[22][9] = 0.8808440659269460280192719417821072306037e-1;
-        a[22][10] = 0.5982566312199630889836790781586188392095e-1;
-        a[22][11] = -0.8251290314815523858109567180972702723517;
-        a[22][12] = -0.4886453387508305661509501250020800230341e-1;
-        a[22][13] = -0.1682608371566511953625070499451460275789;
-        a[22][14] = -0.4443170503743608176862370639456174098455e-1;
-        a[22][15] = 0.0;
-        a[22][16] = 0.7037844639432780238885500484343585742713e-1;
-        a[22][17] = -8.818686397504859095347055843377871115177;
-        a[22][18] = 8.028821919603921347277176943010114087042;
-        a[22][19] = 1.306591406499157862229641342220359676579;
-        a[22][20] = -0.1624883307224014710941456378572112779619;
-        a[22][21] = 0.0;
-
-        c[23] = 0.249;
-
-        a[23][0] = 0.1466442426857961230207869561545353695312e-1;
-        a[23][1] = 0.0;
-        a[23][2] = 0.0;
-        a[23][3] = 0.0;
-        a[23][4] = 0.0;
-        a[23][5] = 0.0;
-        a[23][6] = 0.0;
-        a[23][7] = 0.1898399561399124450873282966092465494738;
-        a[23][8] = 0.3711530651907537053995301753121055816571e-1;
-        a[23][9] = 0.1316247981199510429999180335475037001729;
-        a[23][10] = 0.1466622928822685713088228930531811168096;
-        a[23][11] = -0.2377331162467838377764310139745632904067;
-        a[23][12] = -0.2493071232192949761907468241344412740688e-1;
-        a[23][13] = -0.9427777049221307692559326299260474943801e-1;
-        a[23][14] = -0.2726667627690941070300686079610559177653e-1;
-        a[23][15] = 0.0;
-        a[23][16] = 0.4097659404063362064133671142526181935526e-1;
-        a[23][17] = -5.538725321977494545432278985839620727708;
-        a[23][18] = 4.973442175247713118429917494679987165227;
-        a[23][19] = 0.8626294449189858956229369969895335684221;
-        a[23][20] = -0.2250213948217893084759073334350395278429;
-        a[23][21] = 0.0;
-        a[23][22] = 0.0;
-
-        c[24] = 0.439;
-
-        a[24][0] = 0.1191252689920919922350166270042464848683e-1;
-        a[24][1] = 0.0;
-        a[24][2] = 0.0;
-        a[24][3] = 0.0;
-        a[24][4] = 0.0;
-        a[24][5] = 0.0;
-        a[24][6] = 0.0;
-        a[24][7] = -0.6723137858130876080443132301001699876552;
-        a[24][8] = 0.1801284266825688933352476317750604591900;
-        a[24][9] = 0.1365689447449347163335592887842694005196;
-        a[24][10] = 0.2111038393798906615990851223053694372026;
-        a[24][11] = 0.7775854726244728792738759030556073621755;
-        a[24][12] = 0.2362117948505939633346199470372118351743e-1;
-        a[24][13] = 0.6391325607581230059121991748062989083367e-1;
-        a[24][14] = 0.1197654289246164466950805226357270574198e-1;
-        a[24][15] = 0.0;
-        a[24][16] = -0.2188622145087031801816344581960035242509e-1;
-        a[24][17] = 3.500218060218659348421859254969160434525;
-        a[24][18] = -3.195765244251373581009030415350620096609;
-        a[24][19] = -0.5994401048600148264885439684062323011428;
-        a[24][20] = 0.1137710737227729377873223163880721563872e-1;
-        a[24][21] = 0.0;
-        a[24][22] = 0.0;
-        a[24][23] = 0.0;
-
-        c[25] = 0.794;
-
-        a[25][0] = 0.1184012014074604271397800922251586670558e-1;
-        a[25][1] = 0.0;
-        a[25][2] = 0.0;
-        a[25][3] = 0.0;
-        a[25][4] = 0.0;
-        a[25][5] = 0.0;
-        a[25][6] = 0.0;
-        a[25][7] = -0.6641263289768790103963252230448945748496;
-        a[25][8] = 0.1788914061873194018510863179806909040625;
-        a[25][9] = 0.1367562642287085016908037148859544845819;
-        a[25][10] = 0.2107758119730286112356534765669813731420;
-        a[25][11] = 0.7679184744806335059253231677825531412084;
-        a[25][12] = 0.2313496869580297355994450248598930620113e-1;
-        a[25][13] = 0.6230605705999706953753333719939087802061e-1;
-        a[25][14] = 0.1157099029193866324646273674695267419004e-1;
-        a[25][15] = 0.0;
-        a[25][16] = -0.2129441601042184344575923744216391881547e-1;
-        a[25][17] = 3.163166251075345445930896810245694869610;
-        a[25][18] = -2.819544872276817190650813744596266559171;
-        a[25][19] = -0.3978769046309758754403585402077121882020;
-        a[25][20] = 0.1304821777615737042415746721743137433164;
-        a[25][21] = 0.0;
-        a[25][22] = 0.0;
-        a[25][23] = 0.0;
-        a[25][24] = 0.0;
-
-        bi9[0][0] = 1.0;
-        bi9[0][1] = -60.67156499096275825192551363602396761243;
-        bi9[0][2] = 669.4173339890965298528834018742984105094;
-        bi9[0][3] = -3377.878946225199351520226690513332256243;
-        bi9[0][4] = 9286.468967391046692225695247237821541998;
-        bi9[0][5] = -14780.47713681102054116333940623447596217;
-        bi9[0][6] = 13604.99386328224720102647507806334531178;
-        bi9[0][7] = -6724.705443356114485986782323617396379907;
-        bi9[0][8] = 1381.867933411056511065177869712887072054;
-
-        bi9[1][0] = 0.0;
-        bi9[1][1] = 0.0;
-        bi9[1][2] = 0.0;
-        bi9[1][3] = 0.0;
-        bi9[1][4] = 0.0;
-        bi9[1][5] = 0.0;
-        bi9[1][6] = 0.0;
-        bi9[1][7] = 0.0;
-        bi9[1][8] = 0.0;
-
-        bi9[2][0] = 0.0;
-        bi9[2][1] = 0.0;
-        bi9[2][2] = 0.0;
-        bi9[2][3] = 0.0;
-        bi9[2][4] = 0.0;
-        bi9[2][5] = 0.0;
-        bi9[2][6] = 0.0;
-        bi9[2][7] = 0.0;
-        bi9[2][8] = 0.0;
-
-        bi9[3][0] = 0.0;
-        bi9[3][1] = 0.0;
-        bi9[3][2] = 0.0;
-        bi9[3][3] = 0.0;
-        bi9[3][4] = 0.0;
-        bi9[3][5] = 0.0;
-        bi9[3][6] = 0.0;
-        bi9[3][7] = 0.0;
-        bi9[3][8] = 0.0;
-
-        bi9[4][0] = 0.0;
-        bi9[4][1] = 0.0;
-        bi9[4][2] = 0.0;
-        bi9[4][3] = 0.0;
-        bi9[4][4] = 0.0;
-        bi9[4][5] = 0.0;
-        bi9[4][6] = 0.0;
-        bi9[4][7] = 0.0;
-        bi9[4][8] = 0.0;
-
-        bi9[5][0] = 0.0;
-        bi9[5][1] = 0.0;
-        bi9[5][2] = 0.0;
-        bi9[5][3] = 0.0;
-        bi9[5][4] = 0.0;
-        bi9[5][5] = 0.0;
-        bi9[5][6] = 0.0;
-        bi9[5][7] = 0.0;
-        bi9[5][8] = 0.0;
-
-        bi9[6][0] = 0.0;
-        bi9[6][1] = 0.0;
-        bi9[6][2] = 0.0;
-        bi9[6][3] = 0.0;
-        bi9[6][4] = 0.0;
-        bi9[6][5] = 0.0;
-        bi9[6][6] = 0.0;
-        bi9[6][7] = 0.0;
-        bi9[6][8] = 0.0;
-
-        bi9[7][0] = 0.0;
-        bi9[7][1] = -0.5972167605668447080195299461102185885272;
-        bi9[7][2] = 47.93604836413167338885438264277611366496;
-        bi9[7][3] = -565.7170417002332153647406623941231321861;
-        bi9[7][4] = 2799.622707202858561151696501297251806365;
-        bi9[7][5] = -7032.161794919678731491945450324980521259;
-        bi9[7][6] = 9404.154583498766261417380396057323248088;
-        bi9[7][7] = -6374.661804344304055026346155757703309071;
-        bi9[7][8] = 1720.369337666279969354526136740485828512;
-
-        bi9[8][0] = 0.0;
-        bi9[8][1] = 0.1349844707960069717220566866826298984923;
-        bi9[8][2] = -10.83462914594452681222359026508445675428;
-        bi9[8][3] = 127.8648231869934298266652232809372858609;
-        bi9[8][4] = -632.7779367772187332338007906200698487630;
-        bi9[8][5] = 1589.427325412255464804263544956685619289;
-        bi9[8][6] = -2125.551246305557207529871063524735793374;
-        bi9[8][7] = 1440.817483665754881321490009540733392622;
-        bi9[8][8] = -388.8423097807010970369815759993713435843;
-
-        bi9[9][0] = 0.0;
-        bi9[9][1] = 0.7290747606753265739553888434183906298602e-1;
-        bi9[9][2] = -5.851972901033216313427390997286151384252;
-        bi9[9][3] = 69.06202973876299355875761438134035666890;
-        bi9[9][4] = -341.7744427162097135566927248434513873995;
-        bi9[9][5] = 858.4775271201373000458419246386670445384;
-        bi9[9][6] = -1148.047443579864082962792949259101806969;
-        bi9[9][7] = 778.2107496409214137189987901610599544952;
-        bi9[9][8] = -210.0205396013539280126185513768554408959;
-
-        bi9[10][0] = 0.0;
-        bi9[10][1] = 0.1288534847345785561116203329301814360204;
-        bi9[10][2] = -10.34252098059179829195550043573389974374;
-        bi9[10][3] = 122.0572110662541635697477898873947855333;
-        bi9[10][4] = -604.0372032137003894354892742901494195352;
-        bi9[10][5] = 1517.235637581122757484801301501208740856;
-        bi9[10][6] = -2029.008844152904067945761282194453917620;
-        bi9[10][7] = 1375.375645376320937242441893999796663414;
-        bi9[10][8] = -371.1811168501315596184350570113265596896;
-
-        bi9[11][0] = 0.0;
-        bi9[11][1] = 0.6958971720762387362164791425187492767466;
-        bi9[11][2] = -55.85670513574830649397484217620113261823;
-        bi9[11][3] = 659.1926340795728211557235694809221532847;
-        bi9[11][4] = -3262.215084140846354068888231363556701923;
-        bi9[11][5] = 8194.112807588290125367213373592291507372;
-        bi9[11][6] = -10958.03904467296551539422209426971705612;
-        bi9[11][7] = 7427.971576643475789982515741208313369971;
-        bi9[11][8] = -2004.632548946417281841423673799397312817;
-
-        bi9[12][0] = 0.0;
-        bi9[12][1] = 0.2617668058132623028300450746186815539953e-1;
-        bi9[12][2] = -2.101090775094604504096600471528167656924;
-        bi9[12][3] = 24.79601256660032039131082591571668336326;
-        bi9[12][4] = -122.7106039105787684003939157931166303491;
-        bi9[12][5] = 308.2275402436813943146648421712827254341;
-        bi9[12][6] = -412.1946451000669582597347477727378421190;
-        bi9[12][7] = 279.4085780645544831469608316472195929468;
-        bi9[12][8] = -75.40571800304908908412026712154566848821;
-
-        bi9[13][0] = 0.0;
-        bi9[13][1] = 0.7845665161262074618502665073713581595690e-1;
-        bi9[13][2] = -6.297381611696195354835186970953877598132;
-        bi9[13][3] = 74.31851847203782644803150125553425852543;
-        bi9[13][4] = -367.7877747056500197543410562506219227864;
-        bi9[13][5] = 923.8184599908639672968037927810710633244;
-        bi9[13][6] = -1235.428287659738944177317308989226766106;
-        bi9[13][7] = 837.4423716056154246604113713897017147785;
-        bi9[13][8] = -226.0057431111080504745753090041214216040;
-
-        bi9[14][0] = 0.0;
-        bi9[14][1] = 0.1743239982411997083542666099618776565927e-1;
-        bi9[14][2] = -1.399224563421080747566112426047905797818;
-        bi9[14][3] = 16.51294188206731169216889160389405043672;
-        bi9[14][4] = -81.71931133065192111822155020120302944168;
-        bi9[14][5] = 205.2645942498128099152915051189419910813;
-        bi9[14][6] = -274.5016441798805281613128379066882052764;
-        bi9[14][7] = 186.0725630195001464648099116049710397144;
-        bi9[14][8] = -50.21655137556766366195319885111250443823;
-
-        bi9[15][0] = 0.0;
-        bi9[15][1] = 0.0;
-        bi9[15][2] = 0.0;
-        bi9[15][3] = 0.0;
-        bi9[15][4] = 0.0;
-        bi9[15][5] = 0.0;
-        bi9[15][6] = 0.0;
-        bi9[15][7] = 0.0;
-        bi9[15][8] = 0.0;
-
-        bi9[16][0] = 0.0;
-        bi9[16][1] = -0.3120237034739356102149090910458972991542e-1;
-        bi9[16][2] = 2.505323513858665425821148146925028556689;
-        bi9[16][3] = -29.63209944182268891094187050958569485080;
-        bi9[16][4] = 147.1735748991918299269418428659506735081;
-        bi9[16][5] = -371.6682626297095949736614911358789582154;
-        bi9[16][6] = 500.8032537625967383990239072447982187459;
-        bi9[16][7] = -342.9490465144388046578978125156518258548;
-        bi9[16][8] = 93.79845878067124835173576681254714784020;
-
-        bi9[17][0] = 0.0;
-        bi9[17][1] = 1.005189184600557566129408870772027118073;
-        bi9[17][2] = -80.14524762849856177105235169139266835049;
-        bi9[17][3] = 904.2862871618540482735522275321299612604;
-        bi9[17][4] = -4159.377900797260916077366144135436208057;
-        bi9[17][5] = 9382.948440638739069789096069230867579132;
-        bi9[17][6] = -10865.07436232189535132287509765058944820;
-        bi9[17][7] = 6171.218731586134383883573674542623005216;
-        bi9[17][8] = -1354.861137823673230341057786698974248115;
-
-        bi9[18][0] = 0.0;
-        bi9[18][1] = -1.018790304892149008339208622085912341214;
-        bi9[18][2] = 81.22968543433086703732306275594490877087;
-        bi9[18][3] = -916.5220998408498750879913972119183817869;
-        bi9[18][4] = 4215.658051870923057302442161815205855483;
-        bi9[18][5] = -9509.908233268860695396133182984225322485;
-        bi9[18][6] = 11012.08866136426918043466675789153630399;
-        bi9[18][7] = -6254.720911673139287616309544212696219769;
-        bi9[18][8] = 1373.193636418218902334341350568238768134;
-
-        bi9[19][0] = 0.0;
-        bi9[19][1] = 0.1486756383162591888438322851466844348297e-1;
-        bi9[19][2] = -1.185413256701197518560901363041446866220;
-        bi9[19][3] = 13.37512808773948803650703726266929826638;
-        bi9[19][4] = -61.52057482048160294775557914196592133996;
-        bi9[19][5] = 138.7814224498311940264420678029949672145;
-        bi9[19][6] = -160.7032676951998445023342766109082536156;
-        bi9[19][7] = 91.27733347752088221300420564599480170832;
-        bi9[19][8] = -20.03949580654054522618693682425811381100;
-
-        bi9[20][0] = 0.0;
-        bi9[20][1] = -0.2611963106691748849299256055037070183077e-2;
-        bi9[20][2] = 0.2082557524387797755360510166314825074305;
-        bi9[20][3] = -2.349769034664475310156364256215305171863;
-        bi9[20][4] = 10.80805662268303734097853101791700152378;
-        bi9[20][5] = -24.38139559637047359365680950022456660620;
-        bi9[20][6] = 28.23266885539013812966436802211627788468;
-        bi9[20][7] = -16.03578301196446470276182164155397173990;
-        bi9[20][8] = 3.520578375594150109245344597384118672243;
-
-        bi9[21][0] = 0.0;
-        bi9[21][1] = 61.75136568508868110286254101938221664703;
-        bi9[21][2] = -750.1268910552884473493229930939554325311;
-        bi9[21][3] = 3898.685586244841930988256009725848784326;
-        bi9[21][4] = -10841.82272864745777154320543067614124209;
-        bi9[21][5] = 17321.57566462516617366130887449349951271;
-        bi9[21][6] = -15937.70492491813691151403407581204545506;
-        bi9[21][7] = 7853.047625862506649651583373646970497362;
-        bi9[21][8] = -1605.405697796720304997448299303558881367;
-
-        bi9[22][0] = 0.0;
-        bi9[22][1] = -1.934298042224096808013043975920800289243;
-        bi9[22][2] = 147.5886431665603647558155476856964112175;
-        bi9[22][3] = -1181.453462187681932247747922750006635641;
-        bi9[22][4] = 4124.816268317402451398433861948686547907;
-        bi9[22][5] = -7708.555106592447263782269817638341609752;
-        bi9[22][6] = 8052.349419898650085369881466158633077027;
-        bi9[22][7] = -4443.023055054952034626751291729687203499;
-        bi9[22][8] = 1010.211590494692425940651200300940213030;
-
-        bi9[23][0] = 0.0;
-        bi9[23][1] = 0.9126428307869666507744889095249555815836;
-        bi9[23][2] = -71.31710609019478576924959388866173624547;
-        bi9[23][3] = 695.4561446075446391059957659052646164863;
-        bi9[23][4] = -2620.882193436484363083577561241318740147;
-        bi9[23][5] = 4957.008141154350233563895859921266117993;
-        bi9[23][6] = -5038.298085544127560098020485475941711703;
-        bi9[23][7] = 2633.333935059078624368760947592332675986;
-        bi9[23][8] = -556.2134785809537547385794217224661779516;
-
-        bi9[24][0] = 0.0;
-        bi9[24][1] = -0.4654439970747991651050782556914121007009;
-        bi9[24][2] = 37.09099810444154184865658933003385399191;
-        bi9[24][3] = -417.1382228339854991229014922810663023285;
-        bi9[24][4] = 1920.758799061188217073321625257641207277;
-        bi9[24][5] = -4409.960156149581696033720218281819632770;
-        bi9[24][6] = 5383.383503749768100112081288811655341483;
-        bi9[24][7] = -3353.518501772742141737984950593448341369;
-        bi9[24][8] = 839.8490238379862770256522360126952858160;
-
-        bi9[25][0] = 0.0;
-        bi9[25][1] = -0.1176451708255218561268102928705214692156;
-        bi9[25][2] = 9.481894819354298841374880327580666327841;
-        bi9[25][3] = -114.9156758298319354820100563154045258044;
-        bi9[25][4] = 591.3193291312467068002224871165564177657;
-        bi9[25][5] = -1559.765475086581493834896780108830295683;
-        bi9[25][6] = 2198.545841718649266979102957216738477165;
-        bi9[25][7] = -1564.562048273728342299716850911579457005;
-        bi9[25][8] = 440.0137786917170208520501729678092387026;
-
-        Self {
-            c,
-            a,
-            b,
-            bh: Some(bh),
-            bi: Some(bi9),
-        }
-    }
-}
