+++ src/tableau/verner.rs
@@ -1,0 +1,2802 @@
+//! Verner's Runge-Kutta methods with high-order interpolation. source: https://www.sfu.ca/~jverner/
+
+use crate::tableau::ButcherTableau;
+
+impl ButcherTableau<10,13> {
+    /// A 'most efficient' Runge-Kutta (10:6(7)) pair with 6th-order interpolation.
+    ///
+    /// # Overview
+    /// This provides a 10-stage Runge-Kutta method with:
+    /// - Primary order: 6
+    /// - Embedded order: 7 (for error estimation)
+    /// - Number of stages: 10 primary + 3 additional stages for interpolation
+    /// - Dominant stage order: 3
+    ///
+    /// # Efficiency
+    /// This method is considered "most efficient" in the sense that for a specified
+    /// maximum coefficient from `b` and `A`, it almost minimizes the local truncation
+    /// error 2-norm:
+    ///
+    /// T_82 ~ 0.000003389
+    ///
+    /// (Formulas with slightly different nodes may have a slightly smaller error norm,
+    /// perhaps achieved by having a larger maximum coefficient.)
+    ///
+    /// # Interpolation
+    /// - Node c[11]=1 was chosen to make the interpolants differentiable
+    /// - Node c[12] was chosen as a zero of a quadratic polynomial to enable efficient interpolation
+    /// - The interpolant has continuous derivatives
+    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
+    ///
+    ///   Ti_72 ~ 0.0000165
+    ///
+    /// - The interpolation error has four local maximum values on [0,1]
+    ///
+    /// # References
+    /// - J.H. Verner, SIAM NA 1978, "Explicit Runge-Kutta methods with estimates of the Local Truncation Error"
+    /// - J.H. Verner, SIAM NA 30, 1993, "Differentiable Interpolants for high-order Runge-Kutta methods"
+    pub const fn rkv766e() -> Self {
+        let mut c = [0.0; 13];
+        let mut a = [[0.0; 13]; 13];
+        let mut b = [0.0; 10];
+        let mut bh = [0.0; 10];
+        let mut bi6 = [[0.0; 13]; 13];
+
+        c[0] = 0.0;
+        c[1] = 0.69e-1;
+        c[2] = 0.118;
+        c[3] = 0.177;
+        c[4] = 0.501;
+        c[5] = 0.7737799115305331003715765296862487670813;
+        c[6] = 0.994;
+        c[7] = 0.998;
+        c[8] = 1.0;
+        c[9] = 1.0;
+
+        a[1][0] = 0.69e-1;
+
+        a[2][0] = 0.1710144927536231884057971014492753623188e-1;
+        a[2][1] = 0.1008985507246376811594202898550724637681;
+
+        a[3][0] = 0.4425e-1;
+        a[3][1] = 0.0;
+        a[3][2] = 0.13275;
+
+        a[4][0] = 0.7353445130709566216604424016087331226659;
+        a[4][1] = 0.0;
+        a[4][2] = -2.830160657856937661591496696351623096811;
+        a[4][3] = 2.595816144785981039931054294742889974145;
+
+        a[5][0] = -12.21580485360407974005910916471598682362;
+        a[5][1] = 0.0;
+        a[5][2] = 48.82665485823736062335980699373053427134;
+        a[5][3] = -38.55615592319928364666616600329792491404;
+        a[5][4] = 2.719085830096535863737044703969626233400;
+
+        a[6][0] = 108.8614188704176574066699618897203578466;
+        a[6][1] = 0.0;
+        a[6][2] = -432.4521181775777896358931629332707752654;
+        a[6][3] = 343.9115281800118289547200158889409233641;
+        a[6][4] = -20.55041135925273709189369488701721016265;
+        a[6][5] = 1.223582486401040366396880041626704217305;
+
+        a[7][0] = 113.4755131883738522204615568160304033854;
+        a[7][1] = 0.0;
+        a[7][2] = -450.8122021555997002820400438087344405365;
+        a[7][3] = 358.5132765190089889943579090008312808216;
+        a[7][4] = -21.45046667648445540174055882443151176550;
+        a[7][5] = 1.274053318605952891766776667539031508649;
+        a[7][6] = -0.2174193904638422805639851234763413667602e-2;
+
+        a[8][0] = 115.6996223324232534824963925993127275021;
+        a[8][1] = 0.0;
+        a[8][2] = -459.6635446100248030478961869239726305957;
+        a[8][3] = 365.5534717131745930309149378867953890507;
+        a[8][4] = -21.88511586349784824146225495848432937529;
+        a[8][5] = 1.298718109698721459187976480852777474315;
+        a[8][6] = -0.5318700918481883515898878747322241917739e-4;
+        a[8][7] = -0.3098494764731864405706095716460833640254e-2;
+
+        a[9][0] = 124.1543935612464600014576130437603883332;
+        a[9][1] = 0.0;
+        a[9][2] = -493.2318713314597046194663569971348299332;
+        a[9][3] = 392.2086219315800762927575562172365337929;
+        a[9][4] = -23.48641564290853341361596821616234280392;
+        a[9][5] = 1.362322948908907509911149920532561575254;
+        a[9][6] = -0.7051467367205771043993968232310964220061e-2;
+        a[9][7] = 0.0;
+        a[9][8] = 0.0;
+
+        b[0] = 0.5163520172057869163393251056217968836723e-1;
+        b[1] = 0.0;
+        b[2] = 0.0;
+        b[3] = 0.2767172535461648728769641534539952501983;
+        b[4] = 0.3374175285287150670818592701488271741753;
+        b[5] = 0.1884488267810967803491085059046161195540;
+        b[6] = 24.54134121634868026791753618430192161716;
+        b[7] = -68.81190284469011946382716084194838780382;
+        b[8] = 44.41634281776488378396776021757684795437;
+        b[9] = 0.0;
+
+        bh[0] = 0.5089676583692947576073561095512200263213e-1;
+        bh[1] = 0.0;
+        bh[2] = 0.0;
+        bh[3] = 0.2793777374763233901369432426263934138476;
+        bh[4] = 0.3281330142746535239936396881369403928344;
+        bh[5] = 0.2241721218186151033581794837350130009230;
+        bh[6] = 0.7874574778015076584344903106189416715189;
+        bh[7] = 0.0;
+        bh[8] = 0.0;
+        bh[9] = -0.6700371172080291516839883360724104817561;
+
+        c[10] = 1.0;
+
+        a[10][0] = 0.5163520172057869163393251056217968836723e-1;
+        a[10][1] = 0.0;
+        a[10][2] = 0.0;
+        a[10][3] = 0.2767172535461648728769641534539952501983;
+        a[10][4] = 0.3374175285287150670818592701488271741753;
+        a[10][5] = 0.1884488267810967803491085059046161195540;
+        a[10][6] = 24.54134121634868026791753618430192161716;
+        a[10][7] = -68.81190284469011946382716084194838780382;
+        a[10][8] = 44.41634281776488378396776021757684795437;
+        a[10][9] = 0.0;
+
+        c[11] = 0.3206919742864321488153177349612198724445;
+
+        a[11][0] = 0.5595947882055415742583956081637901452552e-1;
+        a[11][1] = 0.0;
+        a[11][2] = 0.0;
+        a[11][3] = 0.2481262830322509327139022005325437452965;
+        a[11][4] = 0.2103052990591148013587987543073879445482e-1;
+        a[11][5] = -0.1086652415454008449407261022406674546105e-1;
+        a[11][6] = 5.710311335454033086222886871347348898943;
+        a[11][7] = -16.59749151785386284126327966975768025349;
+        a[11][8] = 10.90440686603898964031684903021468235733;
+        a[11][9] = 0.0;
+        a[11][10] = -0.1078447695690422224268752339872593914865e-1;
+
+        c[12] = 0.107;
+
+        a[12][0] = 0.5365176199729764566758655497507692759771e-1;
+        a[12][1] = 0.0;
+        a[12][2] = 0.0;
+        a[12][3] = 0.9828513377552059696782507697799478414148e-1;
+        a[12][4] = 0.2236190536978546688920574502539925812364e-1;
+        a[12][5] = -0.3196168594283279436559562148268690316884e-2;
+        a[12][6] = 0.5451003257307239985591851329642383758416;
+        a[12][7] = -1.638732002009432676897859879241483267007;
+        a[12][8] = 1.093896392979466849054156189707873746269;
+        a[12][9] = 0.0;
+        a[12][10] = -0.4657483105197485920611507484730881742515e-4;
+        a[12][11] = -0.6432077441802662594433314318598382583251e-1;
+
+        bi6[0][0] = 1.0;
+        bi6[0][1] = -7.582446684249578716313054661512721933240;
+        bi6[0][2] = 21.65339697937461679177171822544604176620;
+        bi6[0][3] = -26.50313170111562393346845843813196066333;
+        bi6[0][4] = 13.68567041142918450667757590934976189120;
+        bi6[0][5] = -2.201853803718019957033848524588941372473;
+
+        bi6[1][0] = 0.0;
+        bi6[1][1] = 0.0;
+        bi6[1][2] = 0.0;
+        bi6[1][3] = 0.0;
+        bi6[1][4] = 0.0;
+        bi6[1][5] = 0.0;
+
+        bi6[2][0] = 0.0;
+        bi6[2][1] = 0.0;
+        bi6[2][2] = 0.0;
+        bi6[2][3] = 0.0;
+        bi6[2][4] = 0.0;
+        bi6[2][5] = 0.0;
+
+        bi6[3][0] = 0.0;
+        bi6[3][1] = 5.889034278272535871631390961135086381406;
+        bi6[3][2] = -57.96960154935775491821473273509196795388;
+        bi6[3][3] = 174.1066668861731135303452708623511580529;
+        bi6[3][4] = -196.2003627160861165553101224432427862687;
+        bi6[3][5] = 74.45098035454438694442515750830250503840;
+
+        bi6[4][0] = 0.0;
+        bi6[4][1] = 3.196839349796654162957352828525358112148;
+        bi6[4][2] = -31.32397572700528310044209947396113560366;
+        bi6[4][3] = 93.05298102970384392438258309708366195791;
+        bi6[4][4] = -102.8968871064061647967771234654923865084;
+        bi6[4][5] = 38.30845998243966487696114628399332921615;
+
+        bi6[5][0] = 0.0;
+        bi6[5][1] = -0.8787886596190119861699566896289502028596e-1;
+        bi6[5][2] = 1.013797983077119427760544302767092045959;
+        bi6[5][3] = -4.100173002382458810769471031390844351476;
+        bi6[5][4] = 6.641160480067744814819942865759689363544;
+        bi6[5][5] = -3.278457768019407452844911962268425918186;
+
+        bi6[6][0] = 0.0;
+        bi6[6][1] = -208.3972099204096382331570396018818793788;
+        bi6[6][2] = 2077.909119755834934676745809403064182251;
+        bi6[6][3] = -6428.966291404445362174964426975619327549;
+        bi6[6][4] = 7605.042110521116554859824801255385155563;
+        bi6[6][5] = -3021.046387735747808860531607896646209269;
+
+        bi6[7][0] = 0.0;
+        bi6[7][1] = 594.3593292228751204027648218620251141015;
+        bi6[7][2] = -5925.388971699224459341618710912833599773;
+        bi6[7][3] = 18326.52193738665917733083306878451733081;
+        bi6[7][4] = -21667.18569363528617503083225733032464553;
+        bi6[7][5] = 8602.881495880286217175025916754667412586;
+
+        bi6[8][0] = 0.0;
+        bi6[8][1] = -386.8811915634934821635422144581019035900;
+        bi6[8][2] = 3856.672838888456624774287551141874741487;
+        bi6[8][3] = -11926.20151472972403478718897872831668331;
+        bi6[8][4] = 14096.40733595464142660949072316887784424;
+        bi6[8][5] = -5595.581125732115650649079320906757150875;
+
+        bi6[9][0] = 0.0;
+        bi6[9][1] = 0.0;
+        bi6[9][2] = 0.0;
+        bi6[9][3] = 0.0;
+        bi6[9][4] = 0.0;
+        bi6[9][5] = 0.0;
+
+        bi6[10][0] = 0.0;
+        bi6[10][1] = 0.6111099397023196267050121119527364301482;
+        bi6[10][2] = -6.151956371109289894189681549817017604566;
+        bi6[10][3] = 19.45873699463664083616118469228645171564;
+        bi6[10][4] = -23.90604463475469049657337318293279633818;
+        bi6[10][5] = 9.988154071525019927896857928510625796953;
+
+        bi6[11][0] = 0.0;
+        bi6[11][1] = -8.007634773786193306062361780345514409318;
+        bi6[11][2] = 74.60031392403404792779391526165329391946;
+        bi6[11][3] = -199.6748294772657466686035644277447081048;
+        bi6[11][4] = 207.5792562775741227780748301919115920885;
+        bi6[11][5] = -74.49710595055623073120281924547466349383;
+
+        bi6[12][0] = 0.0;
+        bi6[12][1] = 6.900049017254163553633088407166619306472;
+        bi6[12][2] = -11.01496218408055634389431366310163053450;
+        bi6[12][3] = -27.69438198223954924672720783503507856034;
+        bi6[12][4] = 60.83345444770411331060500303070857149829;
+        bi6[12][5] = -29.02415929863817127361656993973848170993;
+
+        Self {
+            c,
+            a,
+            b,
+            bh: Some(bh),
+            bi: Some(bi6),
+        }
+    }
+}
+
+impl ButcherTableau<10, 16> {
+    /// A 'most efficient' Runge-Kutta (10:7(6)) pair with 7th-order interpolation.
+    ///
+    /// # Overview
+    /// This provides a 10-stage Runge-Kutta method with:
+    /// - Primary order: 7
+    /// - Embedded order: 6 (for error estimation)
+    /// - Number of stages: 10 primary + 6 additional stages for interpolation
+    /// - Dominant stage order: 3
+    ///
+    /// # Efficiency
+    /// This method is considered "most efficient" in the sense that for a specified
+    /// maximum coefficient from `b` and `A`, it almost minimizes the local truncation
+    /// error 2-norm:
+    ///
+    /// T_82 ~ 0.000003389
+    ///
+    /// # Interpolation
+    /// - Additional stages allow computation of approximations at any point with order up to 7
+    /// - The interpolant has continuous derivatives
+    /// - Maximum 2-norm of the local truncation error for the 7th-order interpolant:
+    ///
+    ///   Ti_82 ~ 0.000003389
+    ///
+    /// - The 2-norm is bounded by this value on [0,1] and is nearly monotone increasing
+    /// - The 7th-order interpolant requires 6 additional stages (16 total)
+    ///
+    /// # Notes
+    /// - This method uses the same primary stages as the 6th-order method (rkv766e)
+    /// - The higher-order interpolant requires more additional stages
+    ///
+    /// # References
+    /// - J.H. Verner, SIAM NA 1978, "Explicit Runge-Kutta methods with estimates of the Local Truncation Error"
+    /// - J.H. Verner, SIAM NA 30, 1993, "Differentiable Interpolants for high-order Runge-Kutta methods"
+    pub const fn rkv767e() -> Self {
+        let mut c = [0.0; 16];
+        let mut a = [[0.0; 16]; 16];
+        let mut b = [0.0; 10];
+        let mut bh = [0.0; 10];
+        let mut bi7 = [[0.0; 16]; 16];
+
+        c[0] = 0.0;
+        c[1] = 0.69e-1;
+        c[2] = 0.118;
+        c[3] = 0.177;
+        c[4] = 0.501;
+        c[5] = 0.7737799115305331003715765296862487670813;
+        c[6] = 0.994;
+        c[7] = 0.998;
+        c[8] = 1.0;
+        c[9] = 1.0;
+
+        a[1][0] = 0.69e-1;
+
+        a[2][0] = 0.1710144927536231884057971014492753623188e-1;
+        a[2][1] = 0.1008985507246376811594202898550724637681;
+
+        a[3][0] = 0.4425e-1;
+        a[3][1] = 0.0;
+        a[3][2] = 0.13275;
+
+        a[4][0] = 0.7353445130709566216604424016087331226659;
+        a[4][1] = 0.0;
+        a[4][2] = -2.830160657856937661591496696351623096811;
+        a[4][3] = 2.595816144785981039931054294742889974145;
+
+        a[5][0] = -12.21580485360407974005910916471598682362;
+        a[5][1] = 0.0;
+        a[5][2] = 48.82665485823736062335980699373053427134;
+        a[5][3] = -38.55615592319928364666616600329792491404;
+        a[5][4] = 2.719085830096535863737044703969626233400;
+
+        a[6][0] = 108.8614188704176574066699618897203578466;
+        a[6][1] = 0.0;
+        a[6][2] = -432.4521181775777896358931629332707752654;
+        a[6][3] = 343.9115281800118289547200158889409233641;
+        a[6][4] = -20.55041135925273709189369488701721016265;
+        a[6][5] = 1.223582486401040366396880041626704217305;
+
+        a[7][0] = 113.4755131883738522204615568160304033854;
+        a[7][1] = 0.0;
+        a[7][2] = -450.8122021555997002820400438087344405365;
+        a[7][3] = 358.5132765190089889943579090008312808216;
+        a[7][4] = -21.45046667648445540174055882443151176550;
+        a[7][5] = 1.274053318605952891766776667539031508649;
+        a[7][6] = -0.2174193904638422805639851234763413667602e-2;
+
+        a[8][0] = 115.6996223324232534824963925993127275021;
+        a[8][1] = 0.0;
+        a[8][2] = -459.6635446100248030478961869239726305957;
+        a[8][3] = 365.5534717131745930309149378867953890507;
+        a[8][4] = -21.88511586349784824146225495848432937529;
+        a[8][5] = 1.298718109698721459187976480852777474315;
+        a[8][6] = -0.5318700918481883515898878747322241917739e-4;
+        a[8][7] = -0.3098494764731864405706095716460833640254e-2;
+
+        a[9][0] = 124.1543935612464600014576130437603883332;
+        a[9][1] = 0.0;
+        a[9][2] = -493.2318713314597046194663569971348299332;
+        a[9][3] = 392.2086219315800762927575562172365337929;
+        a[9][4] = -23.48641564290853341361596821616234280392;
+        a[9][5] = 1.362322948908907509911149920532561575254;
+        a[9][6] = -0.7051467367205771043993968232310964220061e-2;
+        a[9][7] = 0.0;
+        a[9][8] = 0.0;
+
+        b[0] = 0.5163520172057869163393251056217968836723e-1;
+        b[1] = 0.0;
+        b[2] = 0.0;
+        b[3] = 0.2767172535461648728769641534539952501983;
+        b[4] = 0.3374175285287150670818592701488271741753;
+        b[5] = 0.1884488267810967803491085059046161195540;
+        b[6] = 24.54134121634868026791753618430192161716;
+        b[7] = -68.81190284469011946382716084194838780382;
+        b[8] = 44.41634281776488378396776021757684795437;
+        b[9] = 0.0;
+
+        bh[0] = 0.5089676583692947576073561095512200263213e-1;
+        bh[1] = 0.0;
+        bh[2] = 0.0;
+        bh[3] = 0.2793777374763233901369432426263934138476;
+        bh[4] = 0.3281330142746535239936396881369403928344;
+        bh[5] = 0.2241721218186151033581794837350130009230;
+        bh[6] = 0.7874574778015076584344903106189416715189;
+        bh[7] = 0.0;
+        bh[8] = 0.0;
+        bh[9] = -0.6700371172080291516839883360724104817561;
+
+        c[10] = 1.0;
+
+        a[10][0] = 0.5163520172057869163393251056217968836723e-1;
+        a[10][1] = 0.0;
+        a[10][2] = 0.0;
+        a[10][3] = 0.2767172535461648728769641534539952501983;
+        a[10][4] = 0.3374175285287150670818592701488271741753;
+        a[10][5] = 0.1884488267810967803491085059046161195540;
+        a[10][6] = 24.54134121634868026791753618430192161716;
+        a[10][7] = -68.81190284469011946382716084194838780382;
+        a[10][8] = 44.41634281776488378396776021757684795437;
+        a[10][9] = 0.0;
+
+        c[11] = 0.3206919742864321488153177349612198724445;
+
+        a[11][0] = 0.5595947882055415742583956081637901452552e-1;
+        a[11][1] = 0.0;
+        a[11][2] = 0.0;
+        a[11][3] = 0.2481262830322509327139022005325437452965;
+        a[11][4] = 0.2103052990591148013587987543073879445482e-1;
+        a[11][5] = -0.1086652415454008449407261022406674546105e-1;
+        a[11][6] = 5.710311335454033086222886871347348898943;
+        a[11][7] = -16.59749151785386284126327966975768025349;
+        a[11][8] = 10.90440686603898964031684903021468235733;
+        a[11][9] = 0.0;
+        a[11][10] = -0.1078447695690422224268752339872593914865e-1;
+
+        c[12] = 0.107;
+
+        a[12][0] = 0.5365176199729764566758655497507692759771e-1;
+        a[12][1] = 0.0;
+        a[12][2] = 0.0;
+        a[12][3] = 0.9828513377552059696782507697799478414148e-1;
+        a[12][4] = 0.2236190536978546688920574502539925812364e-1;
+        a[12][5] = -0.3196168594283279436559562148268690316884e-2;
+        a[12][6] = 0.5451003257307239985591851329642383758416;
+        a[12][7] = -1.638732002009432676897859879241483267007;
+        a[12][8] = 1.093896392979466849054156189707873746269;
+        a[12][9] = 0.0;
+        a[12][10] = -0.4657483105197485920611507484730881742515e-4;
+        a[12][11] = -0.6432077441802662594433314318598382583251e-1;
+
+        c[13] = 0.3;
+
+        a[13][0] = 0.1919717775847849685111579308732385545808e-1;
+        a[13][1] = 0.0;
+        a[13][2] = 0.0;
+        a[13][3] = -0.4739427173175533989609356107380842465934e-1;
+        a[13][4] = -0.2641522514821109801913974252340613156326e-1;
+        a[13][5] = 0.7054089167217366089095708745682724510887e-7;
+        a[13][6] = 1.551079892101621437320838156104778064997;
+        a[13][7] = -4.428095538025376982599451861908192890036;
+        a[13][8] = 2.883718325674217149748617484261428760772;
+        a[13][9] = 0.0;
+        a[13][10] = -0.4297481934497428346001203166435386609045e-2;
+        a[13][11] = 0.1262644300588589746368092621052046424260;
+        a[13][12] = 0.2259426207057731181296447821560200523865;
+
+        c[14] = 0.549;
+
+        a[14][0] = 0.6125242269008585247666695476683627174961e-1;
+        a[14][1] = 0.0;
+        a[14][2] = 0.0;
+        a[14][3] = 0.2525310401834353816786786047482953692850;
+        a[14][4] = 0.1507054877413180556949606462372746086457;
+        a[14][5] = 0.1024270639019054271328153617559339591462e-1;
+        a[14][6] = -6.438444503804215943508588627336983785574;
+        a[14][7] = 18.44870904095403871125278855846631959613;
+        a[14][8] = -12.03592743273174801641739335902460845726;
+        a[14][9] = 0.0;
+        a[14][10] = 0.1513193748819522199500460159517896993473e-1;
+        a[14][11] = 0.1043447705299915390244100081194588662149;
+        a[14][12] = -0.1954546944129134490980892374736483504656e-1;
+        a[14][13] = 0.0;
+
+        c[15] = 0.757;
+
+        a[15][0] = 0.8959466056806525077749567658988579412839e-1;
+        a[15][1] = 0.0;
+        a[15][2] = 0.0;
+        a[15][3] = 0.6389214946562193863310204803419201536875;
+        a[15][4] = 0.4309120296822662010167723442150653884245;
+        a[15][5] = 0.7695592670277012062550066233926247449166e-1;
+        a[15][6] = -7.189894742007330283923979037259652075040;
+        a[15][7] = 21.03645589658795446951415976277135782026;
+        a[15][8] = -13.86140603343527732092225205611764539171;
+        a[15][9] = 0.0;
+        a[15][10] = 0.8282681779726606982694318905099680481702e-2;
+        a[15][11] = -0.2148456743069278028823832921628374810433;
+        a[15][12] = -0.2579762402274666275190288596224563636723;
+        a[15][13] = 0.0;
+        a[15][14] = 0.0;
+
+        bi7[0][0] = 1.0;
+        bi7[0][1] = -8.503890289208650086322705078936310611299;
+        bi7[0][2] = 31.18234339085710879255131238050971776003;
+        bi7[0][3] = -56.83047976918539469582020937384021018043;
+        bi7[0][4] = 52.79739831471022178930385951784095486118;
+        bi7[0][5] = -22.95183302720539338830128746758333934610;
+        bi7[0][6] = 3.358096581752686280222962532571367204997;
+
+        bi7[1][0] = 0.0;
+        bi7[1][1] = 0.0;
+        bi7[1][2] = 0.0;
+        bi7[1][3] = 0.0;
+        bi7[1][4] = 0.0;
+        bi7[1][5] = 0.0;
+        bi7[1][6] = 0.0;
+
+        bi7[2][0] = 0.0;
+        bi7[2][1] = 0.0;
+        bi7[2][2] = 0.0;
+        bi7[2][3] = 0.0;
+        bi7[2][4] = 0.0;
+        bi7[2][5] = 0.0;
+        bi7[2][6] = 0.0;
+
+        bi7[3][0] = 0.0;
+        bi7[3][1] = 6.267709807177576432406478077069849818602;
+        bi7[3][2] = -68.59145123448141827278305837000589905312;
+        bi7[3][3] = 270.1006378954180870537780616882949819305;
+        bi7[3][4] = -486.2388094162716631231341727978491179719;
+        bi7[3][5] = 407.1399818231500101241727526996656040234;
+        bi7[3][6] = -128.4013516214464273415630971437214234972;
+
+        bi7[4][0] = 0.0;
+        bi7[4][1] = 7.642585077624110415092972164105836819717;
+        bi7[4][2] = -83.63756743442629097888399107640792828691;
+        bi7[4][3] = 329.3495021534571296433374383575409388467;
+        bi7[4][4] = -592.8994135546816065453818037938193613121;
+        bi7[4][5] = 496.4495876982774414703954108911602253524;
+        bi7[4][6] = -156.5672764117220689374781672724308842456;
+
+        bi7[5][0] = 0.0;
+        bi7[5][1] = 4.268409521381499328229656167910411979775;
+        bi7[5][2] = -46.71186327091233753264841481404164445942;
+        bi7[5][3] = 183.9428068612486847316162527859995659975;
+        bi7[5][4] = -331.1363205426106275865463338454994779573;
+        bi7[5][5] = 277.2687677656847319301330472909470886979;
+        bi7[5][6] = -87.44335150801085409043509907941132813885;
+
+        bi7[6][0] = 0.0;
+        bi7[6][1] = 555.8670558189042010639619738936099948255;
+        bi7[6][2] = -6083.199322405529992849514286236321590545;
+        bi7[6][3] = 23954.53059900739546845895162864872763414;
+        bi7[6][4] = -43123.26890207833867031432260116626186577;
+        bi7[6][5] = 36108.19740617653066320546034515369066848;
+        bi7[6][6] = -11387.58549530261298929661952410914291952;
+
+        bi7[7][0] = 0.0;
+        bi7[7][1] = -1558.605518026587403407644982678592484625;
+        bi7[7][2] = 17056.79070544840796438857322691726272923;
+        bi7[7][3] = -67166.53412450698972069889474017010525280;
+        bi7[7][4] = 120914.1001657429967184958527822159285576;
+        bi7[7][5] = -101244.4165095385499516578794640912685653;
+        bi7[7][6] = 31929.85337803603227341616601696482662805;
+
+        bi7[8][0] = 0.0;
+        bi7[8][1] = 1006.040439872400448623697106213084759679;
+        bi7[8][2] = -11009.72698072282062279582145142481879227;
+        bi7[8][3] = 43354.29892541348545701322502088633140394;
+        bi7[8][4] = -78046.99336079503441879814959011469619894;
+        bi7[8][5] = 65350.71006860324690110919871372728749245;
+        bi7[8][6] = -20609.91274955351288136818203906961181690;
+
+        bi7[9][0] = 0.0;
+        bi7[9][1] = 0.0;
+        bi7[9][2] = 0.0;
+        bi7[9][3] = 0.0;
+        bi7[9][4] = 0.0;
+        bi7[9][5] = 0.0;
+        bi7[9][6] = 0.0;
+
+        bi7[10][0] = 0.0;
+        bi7[10][1] = -1.612970387190759595570543259931261611538;
+        bi7[10][2] = 17.72909772735324772041234368381732499671;
+        bi7[10][3] = -70.34937384312194980443642443527448780299;
+        bi7[10][4] = 128.1350088912801760638876471983594315219;
+        bi7[10][5] = -109.0734352266536956182626795265083915641;
+        bi7[10][6] = 35.17167283833298123396965633953738445994;
+
+        bi7[11][0] = 0.0;
+        bi7[11][1] = -2.626599644929271923086786435983631061624;
+        bi7[11][2] = 9.920024380612826013243113812430759529521;
+        bi7[11][3] = -7.644130389085101218500146944756885977802;
+        bi7[11][4] = -14.86430552714213266618722882949033593695;
+        bi7[11][5] = 26.11390292373462455033637542344644699734;
+        bi7[11][6] = -10.89889174319094475580532702564635355049;
+
+        bi7[12][0] = 0.0;
+        bi7[12][1] = 9.561675242924823427014852060688684933761;
+        bi7[12][2] = -36.11210856303548088470642279130799743208;
+        bi7[12][3] = 27.82711572968801973495325529709252051216;
+        bi7[12][4] = 54.11089673926027907217912866089741112920;
+        bi7[12][5] = -95.06308263006681094546659235128381873539;
+        bi7[12][6] = 39.67550348122916959602577912391319959234;
+
+        bi7[13][0] = 0.0;
+        bi7[13][1] = -9.791432054831754863321754176936907826271;
+        bi7[13][2] = 116.6733547596190133707339789196320683056;
+        bi7[13][3] = -427.0178661529826169224279390827040458460;
+        bi7[13][4] = 710.0693817808117189359000209005264369724;
+        bi7[13][5] = -556.8214238669928662708433693467844704976;
+        bi7[13][6] = 166.8879855343765057499590627862669188919;
+
+        bi7[14][0] = 0.0;
+        bi7[14][1] = 0.7372370767192203283333752534265429181322e-1;
+        bi7[14][2] = 2.611540012355412116044967235469929822777;
+        bi7[14][3] = -30.67846770866969010553103505851633192303;
+        bi7[14][4] = 100.4285900318910859666056494362334084911;
+        bi7[14][5] = -119.6365555255543602449647502655108119632;
+        bi7[14][6] = 47.20116948230563023501183112698115128060;
+
+        bi7[15][0] = 0.0;
+        bi7[15][1] = -8.581188645336741447289604471431596612033;
+        bi7[15][2] = 93.07222791200057091279868176378132240393;
+        bi7[15][3] = -360.9951446906583731902511625987898308351;
+        bi7[15][4] = 635.7596704131289187099926426178301572747;
+        bi7[15][5] = -517.9168751756012942639785021372581285996;
+        bi7[15][6] = 158.6613101864669192787279448258680763681;
+
+        Self {
+            c,
+            a,
+            b,
+            bh: Some(bh),
+            bi: Some(bi7),
+        }
+    }
+}
+
+impl ButcherTableau<13, 17> {
+    /// An efficient Runge-Kutta (13:7) pair method with 7th-order interpolation.
+    ///
+    /// # Overview
+    /// This provides a 13-stage Runge-Kutta method with:
+    /// - Primary order: 7
+    /// - Embedded order: 7 (used for stepsize control)
+    /// - Number of stages: 13 primary + 4 additional stages for interpolation
+    /// - Dominant stage order: 4
+    ///
+    /// # Interpolation
+    /// - The method provides a 7th-order interpolant using the nodes c[14]=1, c[15]-c[17].
+    /// - The interpolant has continuous derivatives.
+    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
+    ///   
+    ///   Ti_82 ~ 0.0000063833
+    ///   
+    /// - The interpolation error has three local maximum values on [0,1].
+    ///
+    /// # Notes
+    /// - This method uses the same initial 13 stages as the 8th-order method (rkv878e).
+    /// - The interpolant requires fewer additional stages than the 8th-order version.
+    ///
+    /// # References
+    /// - J.H. Verner, SIAM NA 1978, "Explicit Runge--Kutta methods with estimates of the Local Truncation Error"
+    /// - J.H. Verner, SIAM NA 30, 1993, "Differentiable Interpolants for high-order Runge--Kutta methods"
+    pub const fn rkv877e() -> Self {
+        let mut c = [0.0; 17];
+        let mut a = [[0.0; 17]; 17];
+        let mut b = [0.0; 13];
+        let mut bh = [0.0; 13];
+        let mut bi7 = [[0.0; 17]; 17];
+
+        c[0] = 0.0;
+        c[1] = 0.92662e-1;
+        c[2] = 0.1312230361754017604780747799402406525075;
+        c[3] = 0.1968345542631026407171121699103609787612;
+        c[4] = 0.427173;
+        c[5] = 0.485972;
+        c[6] = 0.161915;
+        c[7] = 0.985468;
+        c[8] = 0.9626977348604540392115664231947926180050;
+        c[9] = 0.99626;
+        c[10] = 0.997947;
+        c[11] = 1.0;
+        c[12] = 1.0;
+
+        a[1][0] = 0.92662e-1;
+
+        a[2][0] = 0.3830746548250284242039554953085778876548e-1;
+        a[2][1] = 0.9291557069289891805767923040938286374199e-1;
+
+        a[3][0] = 0.4920863856577566017927804247759024469030e-1;
+        a[3][1] = 0.0;
+        a[3][2] = 0.1476259156973269805378341274327707340709;
+
+        a[4][0] = 0.2743076085702486894953667699331084997155;
+        a[4][1] = 0.0;
+        a[4][2] = -0.9319887203102656329703895298805780048911;
+        a[4][3] = 1.084854111740016943475022759947469505176;
+
+        a[5][0] = 0.6461852970939692178473977038055776659235e-1;
+        a[5][1] = 0.0;
+        a[5][2] = 0.0;
+        a[5][3] = 0.2687629213368923356417418162215129378455;
+        a[5][4] = 0.1525905489537107425735184133979292955622;
+
+        a[6][0] = 0.7189155819773216802747111989237192757241e-1;
+        a[6][1] = 0.0;
+        a[6][2] = 0.0;
+        a[6][3] = 0.1221265783362549661095644262553368220549;
+        a[6][4] = -0.7943550859198561207449556225926957458888e-1;
+        a[6][5] = 0.4733237205799847793746001611156082496161e-1;
+
+        a[7][0] = -6.073603893714328779581044654669389532243;
+        a[7][1] = 0.0;
+        a[7][2] = 0.0;
+        a[7][3] = -73.8956;
+        a[7][4] = 11.93985370695273926305714852083567777565;
+        a[7][5] = -3.839251541405054537968455811084326767757;
+        a[7][6] = 72.85406972816664405449235194491803852435;
+
+        a[8][0] = -4.868640079323569115532909137641679122100;
+        a[8][1] = 0.0;
+        a[8][2] = 0.0;
+        a[8][3] = -59.18572799975646020086112235306397792921;
+        a[8][4] = 9.230819319232425236436363261556755728972;
+        a[8][5] = -2.676847914962525780976208768457285096420;
+        a[8][6] = 58.45720009994685754009028962323039746457;
+        a[8][7] = 0.5894309723726360055153797570581572198272e-2;
+
+        a[9][0] = -6.689861899320853351893486767784362717327;
+        a[9][1] = 0.0;
+        a[9][2] = 0.0;
+        a[9][3] = -81.44271004053111286646869388606624375878;
+        a[9][4] = 13.36778825698397107436768361456719138955;
+        a[9][5] = -4.470777638416181156550300441220009175768;
+        a[9][6] = 80.23321392161410397716314034458127569922;
+        a[9][7] = -0.1313638336212181564579824084323466067556e-1;
+        a[9][8] = 0.1174378303219413902745537676538322377075e-1;
+
+        a[10][0] = -6.788841955800464091202279324616261399375;
+        a[10][1] = 0.0;
+        a[10][2] = 0.0;
+        a[10][3] = -82.65639855934828888595908426400236814314;
+        a[10][4] = 13.59973921874899036529372991555792312537;
+        a[10][5] = -4.574464055350503753023585975406886448931;
+        a[10][6] = 81.41943207216075927716035509139094411649;
+        a[10][7] = -0.1416248014826418017596237345824563213708e-1;
+        a[10][8] = 0.1375441580835227405704608471667945126391e-1;
+        a[10][9] = -0.1111656070581006150219154181785069541859e-2;
+
+        a[11][0] = -6.910189846402485729483801895952800612193;
+        a[11][1] = 0.0;
+        a[11][2] = 0.0;
+        a[11][3] = -84.14495154176748682468399818934446849309;
+        a[11][4] = 13.88512122378983816888937012217322612857;
+        a[11][5] = -4.702458788144493296650978281402083070191;
+        a[11][6] = 82.87411451529241610017736751027256505872;
+        a[11][7] = -0.1645498337198780129448623133684652608689e-1;
+        a[11][8] = 0.1644663972162521365153657249183964172873e-1;
+        a[11][9] = 0.4275449370796530995842729680318146579263e-2;
+        a[11][10] = -0.5902668488222361600852336581750274034670e-2;
+
+        a[12][0] = -6.911973921198979615960353484856419992328;
+        a[12][1] = 0.0;
+        a[12][2] = 0.0;
+        a[12][3] = -84.16635595878781036984379688648536865215;
+        a[12][4] = 13.88834627565582007275122755367241993143;
+        a[12][5] = -4.703463178409702575934171709526131267400;
+        a[12][6] = 82.89518622207404915885492275044886922215;
+        a[12][7] = -0.1020345016228260287805622619295788864042e-1;
+        a[12][8] = 0.1427900423230391471526298444108039935137e-1;
+        a[12][9] = -0.5814993403397981705034981501491752405078e-2;
+        a[12][10] = 0.0;
+        a[12][11] = 0.0;
+
+        b[0] = 0.4625543159712467285354070519930680076661e-1;
+        b[1] = 0.0;
+        b[2] = 0.0;
+        b[3] = 0.0;
+        b[4] = 0.0;
+        b[5] = 0.3706666165521011182439275381303388440188;
+        b[6] = 0.2590440824552746577195309846039127860158;
+        b[7] = -679.9841468175039046601229652340421033215;
+        b[8] = 49.89161129042053159104301060910837813887;
+        b[9] = 10271.23522213731241388784467688648118861;
+        b[10] = -14782.19660635689728059010570422698640618;
+        b[11] = 5141.377953616063739322523982737505384322;
+        b[12] = 0.0;
+
+        bh[0] = 0.4638504234365210644214797353760063769606e-1;
+        bh[1] = 0.0;
+        bh[2] = 0.0;
+        bh[3] = 0.0;
+        bh[4] = 0.0;
+        bh[5] = 0.3725767681581196020016675337652919509451;
+        bh[6] = 0.2585685495121687311414935155287094016045;
+        bh[7] = -147.4950767589265301611858882107047892762;
+        bh[8] = 23.84362712644587506762535723842314335392;
+        bh[9] = 347.4264166730550802334952839143352280200;
+        bh[10] = 0.0;
+        bh[11] = 0.0;
+        bh[12] = -223.4524974005883655795200619648851840879;
+
+        c[13] = 1.0;
+
+        a[13][0] = 0.4625543159712467285354070519930680076661e-1;
+        a[13][1] = 0.0;
+        a[13][2] = 0.0;
+        a[13][3] = 0.0;
+        a[13][4] = 0.0;
+        a[13][5] = 0.3706666165521011182439275381303388440188;
+        a[13][6] = 0.2590440824552746577195309846039127860158;
+        a[13][7] = -679.9841468175039046601229652340421033215;
+        a[13][8] = 49.89161129042053159104301060910837813887;
+        a[13][9] = 10271.23522213731241388784467688648118861;
+        a[13][10] = -14782.19660635689728059010570422698640618;
+        a[13][11] = 5141.377953616063739322523982737505384322;
+        a[13][12] = 0.0;
+
+        c[14] = 0.3110177634953863863927417318829099695921;
+
+        a[14][0] = 0.4911138760562453021745624018512939494564e-1;
+        a[14][1] = 0.0;
+        a[14][2] = 0.0;
+        a[14][3] = 0.0;
+        a[14][4] = 0.0;
+        a[14][5] = 0.3117373122649832951079085303758614870942e-1;
+        a[14][6] = 0.2384923679553289916353395945129932253433;
+        a[14][7] = -114.3180256262527071531290136810612300898;
+        a[14][8] = 3.546408122969840402366032921479574656434;
+        a[14][9] = 2460.597003929216132135942626992844507475;
+        a[14][10] = -3738.924986857493946343920875863790949734;
+        a[14][11] = 1389.104435743812477156452794999320497315;
+        a[14][12] = 0.0;
+        a[14][13] = -0.1259503554386166268241032464519842162533e-1;
+
+        c[15] = 0.2;
+
+        a[15][0] = 0.5179349197485644585674743626479955978608e-1;
+        a[15][1] = 0.0;
+        a[15][2] = 0.0;
+        a[15][3] = 0.0;
+        a[15][4] = 0.0;
+        a[15][5] = 0.3491876693023421912147307261172789499006e-1;
+        a[15][6] = 0.1991958559185454797079748618106548367164;
+        a[15][7] = -83.39793280532507836124705826791773425675;
+        a[15][8] = 3.594107171680368155688005352337368373839;
+        a[15][9] = 1608.071558743333488448763391096377128121;
+        a[15][10] = -2405.179666082838573124443824955693898542;
+        a[15][11] = 876.9134237687558313788496904957089183499;
+        a[15][12] = 0.0;
+        a[15][13] = -0.7325540972670704140488857561266969491257e-2;
+        a[15][14] = -0.8007336945700193815591023393769736797741e-1;
+
+        c[16] = 0.6;
+
+        a[16][0] = 0.4436713671328350085533516489655998875304e-1;
+        a[16][1] = 0.0;
+        a[16][2] = 0.0;
+        a[16][3] = 0.0;
+        a[16][4] = 0.0;
+        a[16][5] = 0.2341061441705438467404449942880480215796;
+        a[16][6] = 0.3613912760091486818803425821815508521337;
+        a[16][7] = 148.9884506482496936874751145110012899551;
+        a[16][8] = -9.426031101789146984989508415038462154627;
+        a[16][9] = -2394.447533524453400464298733938611210343;
+        a[16][10] = 3478.192564092670645140476435107181157928;
+        a[16][11] = -1223.319642639490730778566879958817950079;
+        a[16][12] = 0.0;
+        a[16][13] = 0.1180395533992265808766615770156321870980e-1;
+        a[16][14] = 0.1349647037014828275755998285816810134371;
+        a[16][15] = -0.1744406911214421152358160333642284015360;
+
+        bi7[0][0] = 1.0;
+        bi7[0][1] = -8.212110712910061778007623084400813862130;
+        bi7[0][2] = 35.02635320769604791474092034077537606592;
+        bi7[0][3] = -83.00213339876337627241952209202186684333;
+        bi7[0][4] = 108.6409000057226676877405870074292701819;
+        bi7[0][5] = -72.99647106020921661717338874349522718142;
+        bi7[0][6] = 19.58971739006106373797256727691256843986;
+
+        bi7[1][0] = 0.0;
+        bi7[1][1] = 0.0;
+        bi7[1][2] = 0.0;
+        bi7[1][3] = 0.0;
+        bi7[1][4] = 0.0;
+        bi7[1][5] = 0.0;
+        bi7[1][6] = 0.0;
+
+        bi7[2][0] = 0.0;
+        bi7[2][1] = 0.0;
+        bi7[2][2] = 0.0;
+        bi7[2][3] = 0.0;
+        bi7[2][4] = 0.0;
+        bi7[2][5] = 0.0;
+        bi7[2][6] = 0.0;
+
+        bi7[3][0] = 0.0;
+        bi7[3][1] = 0.0;
+        bi7[3][2] = 0.0;
+        bi7[3][3] = 0.0;
+        bi7[3][4] = 0.0;
+        bi7[3][5] = 0.0;
+        bi7[3][6] = 0.0;
+
+        bi7[4][0] = 0.0;
+        bi7[4][1] = 0.0;
+        bi7[4][2] = 0.0;
+        bi7[4][3] = 0.0;
+        bi7[4][4] = 0.0;
+        bi7[4][5] = 0.0;
+        bi7[4][6] = 0.0;
+
+        bi7[5][0] = 0.0;
+        bi7[5][1] = 24.49504632679124657248259074284359937838;
+        bi7[5][2] = -197.6984345684939158045973156075628087774;
+        bi7[5][3] = 648.8167079210014594451537283787671992452;
+        bi7[5][4] = -1029.913614541901893330814502946072472329;
+        bi7[5][5] = 784.2902782766835465098516222387889570492;
+        bi7[5][6] = -229.6193167975283422738321952686341357219;
+
+        bi7[6][0] = 0.0;
+        bi7[6][1] = 31.56918953505704448457573784274510770401;
+        bi7[6][2] = -253.5018651524157970015956512936350906006;
+        bi7[6][3] = 825.6524274319808118183975171217026655762;
+        bi7[6][4] = -1296.622633293770326113106573164742870221;
+        bi7[6][5] = 974.2628058031631049585612278174199570978;
+        bi7[6][6] = -281.1008802415595634891127273388858567704;
+
+        bi7[7][0] = 0.0;
+        bi7[7][1] = 13532.46062486349519418285003949403697565;
+        bi7[7][2] = -103993.6912108123447484699919983753429873;
+        bi7[7][3] = 315790.0517926441798783566243358352274542;
+        bi7[7][4] = -444650.9408635634151265292682380509278972;
+        bi7[7][5] = 285484.2990404036663083335205079890657793;
+        bi7[7][6] = -66842.16353035308541053385761212610142806;
+
+        bi7[8][0] = 0.0;
+        bi7[8][1] = -797.3373937748126309130191533345207238350;
+        bi7[8][2] = 6069.299906337186949139058394441653960311;
+        bi7[8][3] = -18132.72848645950527711755529827748496005;
+        bi7[8][4] = 24818.08337443399123270224838601234691225;
+        bi7[8][5] = -15179.25266693120755490566761402249251945;
+        bi7[8][6] = 3271.826877684767812685978295789605708913;
+
+        bi7[9][0] = 0.0;
+        bi7[9][1] = -223490.9058981754844945580562298428473795;
+        bi7[9][2] = 1723137.281290412037973827810617942131263;
+        bi7[9][3] = -5261550.360871223434658436208344654853597;
+        bi7[9][4] = 7478225.628671211907649685713214284343111;
+        bi7[9][5] = -4874996.123844563053849368849978953045265;
+        bi7[9][6] = 1168945.715874475339792737435398110753057;
+
+        bi7[10][0] = 0.0;
+        bi7[10][1] = 325937.3545980911494561573600538089395441;
+        bi7[10][2] = -2514175.045366638831325535367544931340193;
+        bi7[10][3] = 7682914.240892724251462067375934686533398;
+        bi7[10][4] = -10933909.13214721004597165165677694552484;
+        bi7[10][5] = 7142613.573847848634614325115730923207696;
+        bi7[10][6] = -1718163.188431172055515952933101768802010;
+
+        bi7[11][0] = 0.0;
+        bi7[11][1] = -115180.8935430816768533254678552911538230;
+        bi7[11][2] = 888954.8267433133410033905646433783738063;
+        bi7[11][3] = -2718988.108015552606334524484935846851660;
+        bi7[11][4] = 3875443.744975928173788736413809491152039;
+        bi7[11][5] = -2537848.359487731062145576624229336937517;
+        bi7[11][6] = 612760.1672807398942806221225503429225392;
+
+        bi7[12][0] = 0.0;
+        bi7[12][1] = 0.0;
+        bi7[12][2] = 0.0;
+        bi7[12][3] = 0.0;
+        bi7[12][4] = 0.0;
+        bi7[12][5] = 0.0;
+        bi7[12][6] = 0.0;
+
+        bi7[13][0] = 0.0;
+        bi7[13][1] = 0.8171974528089539184193047902519773567838;
+        bi7[13][2] = -5.929889493933239653903563757522305816160;
+        bi7[13][3] = 16.17047142225921334983931469529060140792;
+        bi7[13][4] = -18.18864073091308954267555489038495985876;
+        bi7[13][5] = 6.499437906736728059350896773727451974473;
+        bi7[13][6] = 0.6314234430414338689696023886372349357410;
+
+        bi7[14][0] = 0.0;
+        bi7[14][1] = -19.58991707760692246575484124057989970426;
+        bi7[14][2] = 128.4420521863155939052637968458080346391;
+        bi7[14][3] = -321.8756286792641832868863295281122163797;
+        bi7[14][4] = 394.5721351206491640005134395005296613000;
+        bi7[14][5] = -239.3360075607335414326488715970553134960;
+        bi7[14][6] = 57.78736601063988927951280601940973364088;
+
+        bi7[15][0] = 0.0;
+        bi7[15][1] = -15.35274693752665932678945181344254585974;
+        bi7[15][2] = 180.1686249551895421740728266391690224117;
+        bi7[15][3] = -701.6476414218982078177499899730825381674;
+        bi7[15][4] = 1211.468823272660855113033560419426852892;
+        bi7[15][5] = -961.9054874127519588351611984090694516295;
+        bi7[15][6] = 287.2684275443264286925942531369986603532;
+
+        bi7[16][0] = 0.0;
+        bi7[16][1] = -14.40504650928427294859257207187201833753;
+        bi7[16][2] = 120.8217962542519161139448743774919222629;
+        bi7[16][3] = -417.2095154082007875820864103451144802558;
+        bi7[16][4] = 702.6590193669410492418586492824251953058;
+        bi7[16][5] = -564.9514449798660354502747046801145472079;
+        bi7[16][6] = 173.0851912761581306251501634371839282326;
+
+        Self {
+            c,
+            a,
+            b,
+            bh: Some(bh),
+            bi: Some(bi7),
+        }
+    }
+}
+
+impl ButcherTableau<13, 21> {
+    /// A highly efficient Runge-Kutta (13:8(7)) pair method.
+    ///
+    /// # Overview
+    /// This is an efficient 13-stage Runge-Kutta method that provides:
+    /// - Primary order: 8
+    /// - Embedded order: 7 (for error estimation)
+    /// - Number of stages: 13 primary + 8 additional stages for interpolation
+    /// - Dominant stage order: 4
+    ///
+    /// # Efficiency
+    /// The method is particularly "efficient" with a local truncation error 2-norm:
+    ///
+    /// T_92 ~ 0.000000011182
+    ///
+    /// which appears to be the minimum possible for this type of method.
+    ///
+    /// # Interpolation
+    /// - The additional stages (c[14] through c[21]) allow computation of approximations
+    ///   at any point in the solution domain with order 8.
+    /// - The interpolant has continuous derivatives.
+    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
+    ///   
+    ///   Ti_92 ~ 0.0000011515
+    ///   
+    /// - The interpolation error has two local maximum values on [0,1].
+    ///
+    /// # References
+    /// Developed based on formulas in:
+    /// - J.H. Verner, SIAM NA 1978, "Explicit Runge--Kutta methods with estimates of the Local Truncation Error"
+    /// - J.H. Verner, SIAM NA 30, 1993, "Differentiable Interpolants for high-order Runge--Kutta methods"
+    pub const fn rkv878e() -> Self {
+        let mut c = [0.0; 21];
+        let mut a = [[0.0; 21]; 21];
+        let mut b = [0.0; 13];
+        let mut bh = [0.0; 13];
+        let mut bi8 = [[0.0; 21]; 21];
+
+        c[0] = 0.0;
+        c[1] = 0.92662e-1;
+        c[2] = 0.1312230361754017604780747799402406525075;
+        c[3] = 0.1968345542631026407171121699103609787612;
+        c[4] = 0.427173;
+        c[5] = 0.485972;
+        c[6] = 0.161915;
+        c[7] = 0.985468;
+        c[8] = 0.9626977348604540392115664231947926180050;
+        c[9] = 0.99626;
+        c[10] = 0.997947;
+        c[11] = 1.0;
+        c[12] = 1.0;
+
+        a[1][0] = 0.92662e-1;
+
+        a[2][0] = 0.3830746548250284242039554953085778876548e-1;
+        a[2][1] = 0.9291557069289891805767923040938286374199e-1;
+
+        a[3][0] = 0.4920863856577566017927804247759024469030e-1;
+        a[3][1] = 0.0;
+        a[3][2] = 0.1476259156973269805378341274327707340709;
+
+        a[4][0] = 0.2743076085702486894953667699331084997155;
+        a[4][1] = 0.0;
+        a[4][2] = -0.9319887203102656329703895298805780048911;
+        a[4][3] = 1.084854111740016943475022759947469505176;
+
+        a[5][0] = 0.6461852970939692178473977038055776659235e-1;
+        a[5][1] = 0.0;
+        a[5][2] = 0.0;
+        a[5][3] = 0.2687629213368923356417418162215129378455;
+        a[5][4] = 0.1525905489537107425735184133979292955622;
+
+        a[6][0] = 0.7189155819773216802747111989237192757241e-1;
+        a[6][1] = 0.0;
+        a[6][2] = 0.0;
+        a[6][3] = 0.1221265783362549661095644262553368220549;
+        a[6][4] = -0.7943550859198561207449556225926957458888e-1;
+        a[6][5] = 0.4733237205799847793746001611156082496161e-1;
+
+        a[7][0] = -6.073603893714328779581044654669389532243;
+        a[7][1] = 0.0;
+        a[7][2] = 0.0;
+        a[7][3] = -73.8956;
+        a[7][4] = 11.93985370695273926305714852083567777565;
+        a[7][5] = -3.839251541405054537968455811084326767757;
+        a[7][6] = 72.85406972816664405449235194491803852435;
+
+        a[8][0] = -4.868640079323569115532909137641679122100;
+        a[8][1] = 0.0;
+        a[8][2] = 0.0;
+        a[8][3] = -59.18572799975646020086112235306397792921;
+        a[8][4] = 9.230819319232425236436363261556755728972;
+        a[8][5] = -2.676847914962525780976208768457285096420;
+        a[8][6] = 58.45720009994685754009028962323039746457;
+        a[8][7] = 0.5894309723726360055153797570581572198272e-2;
+
+        a[9][0] = -6.689861899320853351893486767784362717327;
+        a[9][1] = 0.0;
+        a[9][2] = 0.0;
+        a[9][3] = -81.44271004053111286646869388606624375878;
+        a[9][4] = 13.36778825698397107436768361456719138955;
+        a[9][5] = -4.470777638416181156550300441220009175768;
+        a[9][6] = 80.23321392161410397716314034458127569922;
+        a[9][7] = -0.1313638336212181564579824084323466067556e-1;
+        a[9][8] = 0.1174378303219413902745537676538322377075e-1;
+
+        a[10][0] = -6.788841955800464091202279324616261399375;
+        a[10][1] = 0.0;
+        a[10][2] = 0.0;
+        a[10][3] = -82.65639855934828888595908426400236814314;
+        a[10][4] = 13.59973921874899036529372991555792312537;
+        a[10][5] = -4.574464055350503753023585975406886448931;
+        a[10][6] = 81.41943207216075927716035509139094411649;
+        a[10][7] = -0.1416248014826418017596237345824563213708e-1;
+        a[10][8] = 0.1375441580835227405704608471667945126391e-1;
+        a[10][9] = -0.1111656070581006150219154181785069541859e-2;
+
+        a[11][0] = -6.910189846402485729483801895952800612193;
+        a[11][1] = 0.0;
+        a[11][2] = 0.0;
+        a[11][3] = -84.14495154176748682468399818934446849309;
+        a[11][4] = 13.88512122378983816888937012217322612857;
+        a[11][5] = -4.702458788144493296650978281402083070191;
+        a[11][6] = 82.87411451529241610017736751027256505872;
+        a[11][7] = -0.1645498337198780129448623133684652608689e-1;
+        a[11][8] = 0.1644663972162521365153657249183964172873e-1;
+        a[11][9] = 0.4275449370796530995842729680318146579263e-2;
+        a[11][10] = -0.5902668488222361600852336581750274034670e-2;
+
+        a[12][0] = -6.911973921198979615960353484856419992328;
+        a[12][1] = 0.0;
+        a[12][2] = 0.0;
+        a[12][3] = -84.16635595878781036984379688648536865215;
+        a[12][4] = 13.88834627565582007275122755367241993143;
+        a[12][5] = -4.703463178409702575934171709526131267400;
+        a[12][6] = 82.89518622207404915885492275044886922215;
+        a[12][7] = -0.1020345016228260287805622619295788864042e-1;
+        a[12][8] = 0.1427900423230391471526298444108039935137e-1;
+        a[12][9] = -0.5814993403397981705034981501491752405078e-2;
+        a[12][10] = 0.0;
+        a[12][11] = 0.0;
+
+        b[0] = 0.4625543159712467285354070519930680076661e-1;
+        b[1] = 0.0;
+        b[2] = 0.0;
+        b[3] = 0.0;
+        b[4] = 0.0;
+        b[5] = 0.3706666165521011182439275381303388440188;
+        b[6] = 0.2590440824552746577195309846039127860158;
+        b[7] = -679.9841468175039046601229652340421033215;
+        b[8] = 49.89161129042053159104301060910837813887;
+        b[9] = 10271.23522213731241388784467688648118861;
+        b[10] = -14782.19660635689728059010570422698640618;
+        b[11] = 5141.377953616063739322523982737505384322;
+        b[12] = 0.0;
+
+        bh[0] = 0.4638504234365210644214797353760063769606e-1;
+        bh[1] = 0.0;
+        bh[2] = 0.0;
+        bh[3] = 0.0;
+        bh[4] = 0.0;
+        bh[5] = 0.3725767681581196020016675337652919509451;
+        bh[6] = 0.2585685495121687311414935155287094016045;
+        bh[7] = -147.4950767589265301611858882107047892762;
+        bh[8] = 23.84362712644587506762535723842314335392;
+        bh[9] = 347.4264166730550802334952839143352280200;
+        bh[10] = 0.0;
+        bh[11] = 0.0;
+        bh[12] = -223.4524974005883655795200619648851840879;
+
+        c[13] = 1.0;
+
+        a[13][0] = 0.4625543159712467285354070519930680076661e-1;
+        a[13][1] = 0.0;
+        a[13][2] = 0.0;
+        a[13][3] = 0.0;
+        a[13][4] = 0.0;
+        a[13][5] = 0.3706666165521011182439275381303388440188;
+        a[13][6] = 0.2590440824552746577195309846039127860158;
+        a[13][7] = -679.9841468175039046601229652340421033215;
+        a[13][8] = 49.89161129042053159104301060910837813887;
+        a[13][9] = 10271.23522213731241388784467688648118861;
+        a[13][10] = -14782.19660635689728059010570422698640618;
+        a[13][11] = 5141.377953616063739322523982737505384322;
+        a[13][12] = 0.0;
+
+        c[14] = 0.3110177634953863863927417318829099695921;
+
+        a[14][0] = 0.4911138760562453021745624018512939494564e-1;
+        a[14][1] = 0.0;
+        a[14][2] = 0.0;
+        a[14][3] = 0.0;
+        a[14][4] = 0.0;
+        a[14][5] = 0.3117373122649832951079085303758614870942e-1;
+        a[14][6] = 0.2384923679553289916353395945129932253433;
+        a[14][7] = -114.3180256262527071531290136810612300898;
+        a[14][8] = 3.546408122969840402366032921479574656434;
+        a[14][9] = 2460.597003929216132135942626992844507475;
+        a[14][10] = -3738.924986857493946343920875863790949734;
+        a[14][11] = 1389.104435743812477156452794999320497315;
+        a[14][12] = 0.0;
+        a[14][13] = -0.1259503554386166268241032464519842162533e-1;
+
+        c[15] = 0.2;
+
+        a[15][0] = 0.5179349197485644585674743626479955978608e-1;
+        a[15][1] = 0.0;
+        a[15][2] = 0.0;
+        a[15][3] = 0.0;
+        a[15][4] = 0.0;
+        a[15][5] = 0.3491876693023421912147307261172789499006e-1;
+        a[15][6] = 0.1991958559185454797079748618106548367164;
+        a[15][7] = -83.39793280532507836124705826791773425675;
+        a[15][8] = 3.594107171680368155688005352337368373839;
+        a[15][9] = 1608.071558743333488448763391096377128121;
+        a[15][10] = -2405.179666082838573124443824955693898542;
+        a[15][11] = 876.9134237687558313788496904957089183499;
+        a[15][12] = 0.0;
+        a[15][13] = -0.7325540972670704140488857561266969491257e-2;
+        a[15][14] = -0.8007336945700193815591023393769736797741e-1;
+
+        c[16] = 0.6;
+
+        a[16][0] = 0.4436713671328350085533516489655998875304e-1;
+        a[16][1] = 0.0;
+        a[16][2] = 0.0;
+        a[16][3] = 0.0;
+        a[16][4] = 0.0;
+        a[16][5] = 0.2341061441705438467404449942880480215796;
+        a[16][6] = 0.3613912760091486818803425821815508521337;
+        a[16][7] = 148.9884506482496936874751145110012899551;
+        a[16][8] = -9.426031101789146984989508415038462154627;
+        a[16][9] = -2394.447533524453400464298733938611210343;
+        a[16][10] = 3478.192564092670645140476435107181157928;
+        a[16][11] = -1223.319642639490730778566879958817950079;
+        a[16][12] = 0.0;
+        a[16][13] = 0.1180395533992265808766615770156321870980e-1;
+        a[16][14] = 0.1349647037014828275755998285816810134371;
+        a[16][15] = -0.1744406911214421152358160333642284015360;
+
+        c[17] = 0.251;
+
+        a[17][0] = 0.4927309335160545018837186692162194114006e-1;
+        a[17][1] = 0.0;
+        a[17][2] = 0.0;
+        a[17][3] = 0.0;
+        a[17][4] = 0.0;
+        a[17][5] = 0.1478473900086014712887940970539529915367;
+        a[17][6] = 0.1915391503780640342241981813756999876738;
+        a[17][7] = 85.69949519051240679824481172082685765978;
+        a[17][8] = -5.094095153569584154270332765058351241762;
+        a[17][9] = -1410.985179005200593252911690612297856885;
+        a[17][10] = 2056.873083845176325396897975581141270029;
+        a[17][11] = -726.4906359530084743440935010911966625383;
+        a[17][12] = 0.0;
+        a[17][13] = 0.5440493103670101179960418796640615006313e-2;
+        a[17][14] = -0.1437912346504880100428763097368602714809;
+        a[17][15] = 0.8130926198406332202245160436986436093968e-1;
+        a[17][16] = -0.8328707808559681272816269219617664899487e-1;
+
+        c[18] = 0.451;
+
+        a[18][0] = 0.4696791108973669580037521395572005624943e-1;
+        a[18][1] = 0.0;
+        a[18][2] = 0.0;
+        a[18][3] = 0.0;
+        a[18][4] = 0.0;
+        a[18][5] = 0.2010207606761097437232392903607688964540;
+        a[18][6] = 0.2636820858834097192664348219821625228752;
+        a[18][7] = 129.6946428658473021706281203179680341347;
+        a[18][8] = -7.848674084672579804338628962846012175606;
+        a[18][9] = -2121.729757286215698398545896705499941278;
+        a[18][10] = 3090.157418699412077487400584308334376641;
+        a[18][11] = -1090.274673114349457503609053352481192228;
+        a[18][12] = 0.0;
+        a[18][13] = 0.8963759217746587577580254051724170346208e-2;
+        a[18][14] = 0.4875201082359089303480235638217668769362e-1;
+        a[18][15] = -0.2352913684548179696659019693524338623585e-1;
+        a[18][16] = -0.9381447086675579397096734527257404140647e-1;
+        a[18][17] = 0.0;
+
+        c[19] = 0.705;
+
+        a[19][0] = 0.4619987481300149103923582280519782866590e-1;
+        a[19][1] = 0.0;
+        a[19][2] = 0.0;
+        a[19][3] = 0.0;
+        a[19][4] = 0.0;
+        a[19][5] = 0.2322836783017554180760864761340094394641;
+        a[19][6] = 0.2982051978226913141175018152700660351379;
+        a[19][7] = 123.6261700088200670668986465184100244666;
+        a[19][8] = -7.231890372340911603936034207018277000416;
+        a[19][9] = -2046.801418638051394033916845196316375171;
+        a[19][10] = 2986.087035469917759873466009280884537880;
+        a[19][11] = -1055.671796398942928677813090910742298972;
+        a[19][12] = 0.0;
+        a[19][13] = 0.7930343978313866522846570067876462025829e-2;
+        a[19][14] = 0.8980932823961332375619194670691924357732e-1;
+        a[19][15] = -0.8017153994764179414644520644800061934373e-1;
+        a[19][16] = 0.1026430473896737559358970902463204065417;
+        a[19][17] = 0.0;
+        a[19][18] = 0.0;
+
+        c[20] = 0.8795;
+
+        a[20][0] = 0.5329137345256786599202448371739175627057e-1;
+        a[20][1] = 0.0;
+        a[20][2] = 0.0;
+        a[20][3] = 0.0;
+        a[20][4] = 0.0;
+        a[20][5] = 0.2043185802985482892225157926395741314696;
+        a[20][6] = 0.1358728897801712829783843582299538835128;
+        a[20][7] = -402.8079844608128209694219700870509710185;
+        a[20][8] = 29.46110568417652997397169152745513583508;
+        a[20][9] = 6093.592821606031634755087198810590639032;
+        a[20][10] = -8771.855718519913765428348513348127981921;
+        a[20][11] = 3051.798314820693282143366860169701133691;
+        a[20][12] = 0.0;
+        a[20][13] = -0.3308969275605729485812518877458826604068e-1;
+        a[20][14] = 0.2805382986735395987509175854573185663385e-1;
+        a[20][15] = 0.1335953634219562765350679978518628190610;
+        a[20][16] = 0.1689185257605991455997737252221182006005;
+        a[20][17] = 0.0;
+        a[20][18] = 0.0;
+        a[20][19] = 0.0;
+
+        bi8[0][0] = 1.0;
+        bi8[0][1] = -8.560562317505520965893266232545864348587;
+        bi8[0][2] = 37.91120319790645210673257106393947681809;
+        bi8[0][3] = -95.90997970387049459577928861171245504580;
+        bi8[0][4] = 143.9557989560652588678965218768899263890;
+        bi8[0][5] = -126.8645258830010362955147010306166337461;
+        bi8[0][6] = 60.67897508156613701498205859458556491581;
+        bi8[0][7] = -12.16465389956367145957035495534070818157;
+
+        bi8[1][0] = 0.0;
+        bi8[1][1] = 0.0;
+        bi8[1][2] = 0.0;
+        bi8[1][3] = 0.0;
+        bi8[1][4] = 0.0;
+        bi8[1][5] = 0.0;
+        bi8[1][6] = 0.0;
+        bi8[1][7] = 0.0;
+
+        bi8[2][0] = 0.0;
+        bi8[2][1] = 0.0;
+        bi8[2][2] = 0.0;
+        bi8[2][3] = 0.0;
+        bi8[2][4] = 0.0;
+        bi8[2][5] = 0.0;
+        bi8[2][6] = 0.0;
+        bi8[2][7] = 0.0;
+
+        bi8[3][0] = 0.0;
+        bi8[3][1] = 0.0;
+        bi8[3][2] = 0.0;
+        bi8[3][3] = 0.0;
+        bi8[3][4] = 0.0;
+        bi8[3][5] = 0.0;
+        bi8[3][6] = 0.0;
+        bi8[3][7] = 0.0;
+
+        bi8[4][0] = 0.0;
+        bi8[4][1] = 0.0;
+        bi8[4][2] = 0.0;
+        bi8[4][3] = 0.0;
+        bi8[4][4] = 0.0;
+        bi8[4][5] = 0.0;
+        bi8[4][6] = 0.0;
+        bi8[4][7] = 0.0;
+
+        bi8[5][0] = 0.0;
+        bi8[5][1] = 31.58105876966070755817711181026329343665;
+        bi8[5][2] = -180.5178006497936101340318821069462934370;
+        bi8[5][3] = 371.2428684175585341424791160692889410620;
+        bi8[5][4] = -219.9594432616132922400607592351592840556;
+        bi8[5][5] = -240.0938376157735211521838900097427201378;
+        bi8[5][6] = 371.1625149095733967216817534260019455117;
+        bi8[5][7] = -133.0446939530601137778175224155755435359;
+
+        bi8[6][0] = 0.0;
+        bi8[6][1] = 22.07073965292731709025653809370859921609;
+        bi8[6][2] = -126.1566754274917694565163606891720887307;
+        bi8[6][3] = 259.4468018507764496957034678807673289852;
+        bi8[6][4] = -153.7208629336292887173729422706050971616;
+        bi8[6][5] = -167.7919863592614905102403297532219205119;
+        bi8[6][6] = 259.3906459958421903925844375556299672132;
+        bi8[6][7] = -92.97961869670813383669527983250287622418;
+
+        bi8[7][0] = 0.0;
+        bi8[7][1] = -57935.13185200130004451750696526889385541;
+        bi8[7][2] = 331158.0735325498459302442841924636844465;
+        bi8[7][3] = -681041.2750173130138192297984810369904889;
+        bi8[7][4] = 403513.3666796717724078043933204165043790;
+        bi8[7][5] = 440449.7088136183995582652495030608971324;
+        bi8[7][6] = -680893.8673222815216844218479358007413730;
+        bi8[7][7] = 244069.1410189383137471951034009314976560;
+
+        bi8[8][0] = 0.0;
+        bi8[8][1] = 4250.800686968169496178700937981058363787;
+        bi8[8][2] = -24297.63981660104922968878449326985727293;
+        bi8[8][3] = 49969.17461226540555036109266936358732405;
+        bi8[8][4] = -29606.47264365440829329549714113549060954;
+        bi8[8][5] = -32316.55586083186219547193751752582986592;
+        bi8[8][6] = 49958.35905508492049348605670433958547126;
+        bi8[8][7] = -17907.77442194075528997858814914394503258;
+
+        bi8[9][0] = 0.0;
+        bi8[9][1] = 875116.5297933782663101143489237552187982;
+        bi8[9][2] = -5002179.072677031402026209959119949939645;
+        bi8[9][3] = 10287203.25970261764886749823989225303597;
+        bi8[9][4] = -6095113.722637695183800934414417711613054;
+        bi8[9][5] = -6653041.227387194363013444576409171926027;
+        bi8[9][6] = 10284976.65027898959354116589597477678226;
+        bi8[9][7] = -3686691.181850927247464301690167065077119;
+
+        bi8[10][0] = 0.0;
+        bi8[10][1] = -1259453.640882216588730059128817651829025;
+        bi8[10][2] = 7199055.703947677084639189578415812070854;
+        bi8[10][3] = -14805177.54931099282922608063380775248053;
+        bi8[10][4] = 8771989.681625245671776936305003143572362;
+        bi8[10][5] = 9574949.976948425063370422937920428882522;
+        bi8[10][6] = -14801973.04882455689422764622042624383139;
+        bi8[10][7] = 5305826.679890061595116647056008036628796;
+
+        bi8[11][0] = 0.0;
+        bi8[11][1] = 438049.0501694909633920945145914847818354;
+        bi8[11][2] = -2503894.872242029083881453081815093438434;
+        bi8[11][3] = 5149370.927637711155082091178724257279041;
+        bi8[11][4] = -3050975.139855797649176572488872351115466;
+        bi8[11][5] = -3330251.790676982456969584764119587826344;
+        bi8[11][6] = 5148256.374192641390109800793347658427157;
+        bi8[11][7] = -1845413.171271418254817053627873630602405;
+
+        bi8[12][0] = 0.0;
+        bi8[12][1] = 0.0;
+        bi8[12][2] = 0.0;
+        bi8[12][3] = 0.0;
+        bi8[12][4] = 0.0;
+        bi8[12][5] = 0.0;
+        bi8[12][6] = 0.0;
+        bi8[12][7] = 0.0;
+
+        bi8[13][0] = 0.0;
+        bi8[13][1] = -2.504431470815825952597052601849741871849;
+        bi8[13][2] = 10.02051593444700250034144043591917565446;
+        bi8[13][3] = 3.188314791667844930562547327020218859370;
+        bi8[13][4] = -81.28919087456095909677167863874787591733;
+        bi8[13][5] = 166.5016415475332861729532713048075758880;
+        bi8[13][6] = -137.9649604853951315639665825699498265027;
+        bi8[13][7] = 42.04811055712378300947805474280047389006;
+
+        bi8[14][0] = 0.0;
+        bi8[14][1] = -8.781714078400345890500582254933426660729;
+        bi8[14][2] = 95.92683681317881690739004713651739721760;
+        bi8[14][3] = -437.8684018822059554032097156332107532369;
+        bi8[14][4] = 1022.893757817680688328813133512141573185;
+        bi8[14][5] = -1279.306771651130473378088029252960844151;
+        bi8[14][6] = 814.4619777825506941886287783493535555706;
+        bi8[14][7] = -207.3256848016734247530336318569075019247;
+
+        bi8[15][0] = 0.0;
+        bi8[15][1] = 23.68361088778252813755830682525871572206;
+        bi8[15][2] = -258.7073385100442622636796566042807424535;
+        bi8[15][3] = 1180.897573941780652808280782076758850644;
+        bi8[15][4] = -2758.666192432960487581886381121050830153;
+        bi8[15][5] = 3450.192469874827074026059813408271639164;
+        bi8[15][6] = -2196.541630994369154046535169690574772508;
+        bi8[15][7] = 559.1415072329836489202023051056171395846;
+
+        bi8[16][0] = 0.0;
+        bi8[16][1] = -40.60495590126188439710688778738077499754;
+        bi8[16][2] = 443.5472327808004803148884291518351973556;
+        bi8[16][3] = -2024.619224703986215542471608189157701003;
+        bi8[16][4] = 4729.663885325312770086552887900617076925;
+        bi8[16][5] = -5915.268315880109317019855761487629399742;
+        bi8[16][6] = 3765.915446103794090758138472995147035931;
+        bi8[16][7] = -958.6340677245499242001455325834314344682;
+
+        bi8[17][0] = 0.0;
+        bi8[17][1] = -46.25448592533055837377842531202839839512;
+        bi8[17][2] = 439.1938466041217159655702274089786458299;
+        bi8[17][3] = -1719.438253681139282550356858145413164525;
+        bi8[17][4] = 3520.667063097926217589621564937806803955;
+        bi8[17][5] = -3962.634989140663522153677993129178399151;
+        bi8[17][6] = 2322.579486243480292154738138853866205762;
+        bi8[17][7] = -554.1126671983948626321166546140316934748;
+
+        bi8[18][0] = 0.0;
+        bi8[18][1] = 14.23775221010393402969945815961053497020;
+        bi8[18][2] = -276.2932724089639106808497191835732392038;
+        bi8[18][3] = 1558.151943742051093976689365840541702310;
+        bi8[18][4] = -3984.134381318251874989648181935067360669;
+        bi8[18][5] = 5195.127830395725191831019434328633823950;
+        bi8[18][6] = -3374.420443020703185373799939254029388933;
+        bi8[18][7] = 867.3305704000387512068895820438839275766;
+
+        bi8[19][0] = 0.0;
+        bi8[19][1] = 6.177962769210591542599576397764741591492;
+        bi8[19][2] = -157.4910701175524827231177916800491606577;
+        bi8[19][3] = 978.6380890625254126165941312401420957681;
+        bi8[19][4] = -2692.299557325995918275212259001422935846;
+        bi8[19][5] = 3734.579314715993155267252453495657367624;
+        bi8[19][6] = -2556.424739731601341815253154933956957044;
+        bi8[19][7] = 686.8200006274205833871370444818648485639;
+
+        bi8[20][0] = 0.0;
+        bi8[20][1] = -18.65289021588136658934344739820379525269;
+        bi8[20][2] = 130.3737772179961353812355151043116833235;
+        bi8[20][3] = -398.2673561235244947185709369394563871567;
+        bi8[20][4] = 665.1759551798239720897702016127403918291;
+        bi8[20][5] = -644.5526670389200969746336450782179076782;
+        bi8[20][6] = 346.6953482377737791841235419437045433947;
+        bi8[20][7] = -80.77216725726792837258122924487852845967;
+
+        Self {
+            c,
+            a,
+            b,
+            bh: Some(bh),
+            bi: Some(bi8),
+        }
+    }
+}
+
+impl ButcherTableau<16, 21> {
+    /// A better efficient Runge-Kutta (16:8(9)) pair with 8th-order interpolation.
+    ///
+    /// # Overview
+    /// This provides a 16-stage Runge-Kutta method with:
+    /// - Primary order: 8
+    /// - Embedded order: 9 (for error estimation)
+    /// - Number of stages: 16 primary + 5 additional stages for interpolation
+    /// - Dominant stage order: 5
+    ///
+    /// # Efficiency
+    /// This method is particularly efficient with a local truncation error 2-norm:
+    ///
+    /// T_10,2 ~ 0.00000034399
+    ///
+    /// This 2-norm is smaller than previous (16:8,9) pairs and has two local maximum values on [0,1].
+    ///
+    /// # Interpolation
+    /// - Nodes c[17]=1 and c[18]-c[21] were selected to minimize interpolation error.
+    /// - The interpolant has continuous derivatives.
+    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
+    ///
+    ///   Ti_9,2 ~ 0.000003721
+    ///
+    /// - The interpolation error has three local maximum values on [0,1].
+    ///
+    /// # Notes
+    /// - Exact coefficients of the method are possible using surds in terms of sqrt(6).
+    /// - This method shares the same primary stages as the 9th-order method (rkv989e).
+    /// - The interpolant requires fewer additional stages than the 9th-order version.
+    ///
+    /// # References
+    /// - J.H. Verner, SIAM J NA 1978, "Explicit Runge-Kutta methods with estimates of the Local Truncation Error"
+    /// - J.H. Verner, SIAM J NA 1993, "Differentiable Interpolants for high-order Runge-Kutta methods"
+    pub const fn rkv988e() -> Self {
+        let mut c = [0.0; 21];
+        let mut a = [[0.0; 21]; 21];
+        let mut b = [0.0; 16];
+        let mut bh = [0.0; 16];
+        let mut bi8 = [[0.0; 21]; 21];
+
+        c[0] = 0.0;
+        c[1] = 0.3571e-1;
+        c[2] = 0.9906028091267414072062290548990445721420e-1;
+        c[3] = 0.1485904213690112110809343582348566858213;
+        c[4] = 0.6134;
+        c[5] = 0.2327359473605626756631680289030288192566;
+        c[6] = 0.5538640526394373243368319710969711807434;
+        c[7] = 0.6555;
+        c[8] = 0.491625;
+        c[9] = 0.6858e-1;
+        c[10] = 0.253;
+        c[11] = 0.6620641795412045944786226689660879617864;
+        c[12] = 0.8309;
+        c[13] = 0.8998;
+        c[14] = 1.0;
+        c[15] = 1.0;
+
+        a[1][0] = 0.3571e-1;
+
+        a[2][0] = -0.3833735636677017025757228807792426014327e-1;
+        a[2][1] = 0.1373976372794443109781951935678287173575;
+
+        a[3][0] = 0.3714760534225280277023358955871417145532e-1;
+        a[3][1] = 0.0;
+        a[3][2] = 0.1114428160267584083107007686761425143660;
+
+        a[4][0] = 2.674764429871505119194043347886848033362;
+        a[4][1] = 0.0;
+        a[4][2] = -9.982382134885293836441318015951354323506;
+        a[4][3] = 7.921017705013788717247274668064506290145;
+
+        a[5][0] = 0.5242104050577351069841401193318449756747e-1;
+        a[5][1] = 0.0;
+        a[5][2] = 0.0;
+        a[5][3] = 0.1796911189175953081279466649977364298948;
+        a[5][4] = 0.6237879371938568368073519721078917943307e-3;
+
+        a[6][0] = 0.1592492223647632083060991150129732726703;
+        a[6][1] = 0.0;
+        a[6][2] = 0.0;
+        a[6][3] = -0.4298429877241087508189185066909803751595;
+        a[6][4] = 0.6665266542726088051243445942394404211672e-1;
+        a[6][5] = 0.7578051525715219863372169033510342411158;
+
+        a[7][0] = 0.7283333333333333333333333333333333333333e-1;
+        a[7][1] = 0.0;
+        a[7][2] = 0.0;
+        a[7][3] = 0.0;
+        a[7][4] = 0.0;
+        a[7][5] = 0.3359344590665103678713422141936031057620;
+        a[7][6] = 0.2467322076001562987953244524730635609046;
+
+        a[8][0] = 0.7297558593750000000000000000000000000000e-1;
+        a[8][1] = 0.0;
+        a[8][2] = 0.0;
+        a[8][3] = 0.0;
+        a[8][4] = 0.0;
+        a[8][5] = 0.3348009729699333533129043555243262838993;
+        a[8][6] = 0.1184158239050666466870956444756737161007;
+        a[8][7] = -0.3456738281250000000000000000000000000000e-1;
+
+        a[9][0] = 0.4911213663452096382929799914888692571185e-1;
+        a[9][1] = 0.0;
+        a[9][2] = 0.0;
+        a[9][3] = 0.0;
+        a[9][4] = 0.0;
+        a[9][5] = 0.3983857361308652347066808423646954244674e-1;
+        a[9][6] = 0.1069675288939354812077139476183615284408;
+        a[9][7] = -0.2174259165458647598455650055976622663086e-1;
+        a[9][8] = -0.1055956474869564925231235304439517699685;
+
+        a[10][0] = -0.2707988818641280489175862864774930636734e-1;
+        a[10][1] = 0.0;
+        a[10][2] = 0.0;
+        a[10][3] = 0.0;
+        a[10][4] = 0.0;
+        a[10][5] = 0.3330000000000000000000000000000000000000e-1;
+        a[10][6] = -0.1645526070036057075627040519129748090051;
+        a[10][7] = 0.3428266306497389946775103799535438271025e-1;
+        a[10][8] = 0.1585264064439221046855989221358760453749;
+        a[10][9] = 0.2185234256811225083011127204294936872873;
+
+        a[11][0] = 0.5584657769108862727526281006220448941548e-1;
+        a[11][1] = 0.0;
+        a[11][2] = 0.0;
+        a[11][3] = 0.0;
+        a[11][4] = 0.0;
+        a[11][5] = 0.9166533166672539087479545790553151183505e-1;
+        a[11][6] = 0.2392399655523627049569179599100121305027;
+        a[11][7] = 0.1023834712248414879740316672186561113729e-1;
+        a[11][8] = -0.2679331322859542442991625976085644567159e-2;
+        a[11][9] = 0.4235624181474284646680744174197106205439e-1;
+        a[11][10] = 0.2253970470166604185504274586005888014087;
+
+        a[12][0] = -0.4802510512725195893103945790000307739284;
+        a[12][1] = 0.0;
+        a[12][2] = 0.0;
+        a[12][3] = 0.0;
+        a[12][4] = 0.0;
+        a[12][5] = -6.359610162555930097843606206238737368598;
+        a[12][6] = -0.2762313898040841385163666588363199062208;
+        a[12][7] = -6.500796633979846747011407331804389650124;
+        a[12][8] = 0.5734765877040956875980719881692630407295;
+        a[12][9] = 1.347125994868138849716719646646049357523;
+        a[12][10] = 5.936840409706221306933249144503479736029;
+        a[12][11] = 6.590346245333924728433733996560685564589;
+
+        a[13][0] = 0.3307533067671401079975082476263855100019;
+        a[13][1] = 0.0;
+        a[13][2] = 0.0;
+        a[13][3] = 0.0;
+        a[13][4] = 0.0;
+        a[13][5] = 5.956207776829962111958631943398154563177;
+        a[13][6] = -0.4868316400481527952968983900652017328476;
+        a[13][7] = 4.462055288206771191616083778225319083232;
+        a[13][8] = 0.7410258231442071778525406966672954745201;
+        a[13][9] = -0.7118192034575913119318346652024763899638;
+        a[13][10] = -5.454619594516665440649148499095762263275;
+        a[13][11] = -4.140803729244709693305868672684077643030;
+        a[13][12] = 0.2038319723190386517589855611303633981843;
+
+        a[14][0] = -0.5847111122998944389651719860007729221315;
+        a[14][1] = 0.0;
+        a[14][2] = 0.0;
+        a[14][3] = 0.0;
+        a[14][4] = 0.0;
+        a[14][5] = -12.41268417116267068926060969930886386670;
+        a[14][6] = 1.360245445660928146977003413153919520189;
+        a[14][7] = -22.42610531111868229005965344123311160860;
+        a[14][8] = -0.8828857055865458252669099251378431470909;
+        a[14][9] = 1.770155128538230466813900572142546329490;
+        a[14][10] = 12.15809651918533877218739262890752294243;
+        a[14][11] = 22.23037520407760703834156132941550000589;
+        a[14][12] = -0.6634483760201249006317401316871863808959;
+        a[14][13] = 0.4509623787258137198642272397482891274306;
+
+        a[15][0] = 1.940575549810648717395482898478643635949;
+        a[15][1] = 0.0;
+        a[15][2] = 0.0;
+        a[15][3] = 0.0;
+        a[15][4] = 0.0;
+        a[15][5] = 21.97798408114556310164832153356570429956;
+        a[15][6] = 0.8230747326984728145128172720328075817426;
+        a[15][7] = 68.16441683626354817206843139989369741474;
+        a[15][8] = -3.117097463620266666801525529574905518624;
+        a[15][9] = -4.568841021822439620303730003755793384193;
+        a[15][10] = -18.74190987126264955904919353670543412590;
+        a[15][11] = -66.57711839637831878350036856645874492472;
+        a[15][12] = 1.098915553165441824029764532524025021446;
+        a[15][13] = 0.0;
+        a[15][14] = 0.0;
+
+        b[0] = 0.1500669014979724795766288712377040980022e-1;
+        b[1] = 0.0;
+        b[2] = 0.0;
+        b[3] = 0.0;
+        b[4] = 0.0;
+        b[5] = 0.0;
+        b[6] = 0.0;
+        b[7] = -1.055180992746381278594381685080184474710;
+        b[8] = 0.2384947263782183112638140557774851953952;
+        b[9] = 0.1288151774282991354622515887144081172371;
+        b[10] = 0.2276623111046215614614917896665746508574;
+        b[11] = 1.229532587437517443160321815173576421896;
+        b[12] = 0.4624976662810383487397308275256128637703e-1;
+        b[13] = 0.1386196319366293903628308621201843492468;
+        b[14] = 0.3080010168319435405203560375162404389992e-1;
+        b[15] = 0.0;
+
+        bh[0] = 0.1897210532481101330735875918987801428423e-1;
+        bh[1] = 0.0;
+        bh[2] = 0.0;
+        bh[3] = 0.0;
+        bh[4] = 0.0;
+        bh[5] = 0.0;
+        bh[6] = 0.0;
+        bh[7] = 3.408110314549493848404398964228060776034;
+        bh[8] = 0.1260323883820920906560270507988839661845;
+        bh[9] = 0.1188375063451149770930378540709095182997;
+        bh[10] = 0.2491041997838687569190073177760326138188;
+        bh[11] = -3.269966219928978218713853055139116510335;
+        bh[12] = 0.3023798100228882907409723963501699715077;
+        bh[13] = 0.0;
+        bh[14] = 0.0;
+        bh[15] = 0.4652989552070924159305071272518165020637e-1;
+
+        c[16] = 1.0;
+
+        a[16][0] = 0.1500669014979724795766288712377040980022e-1;
+        a[16][1] = 0.0;
+        a[16][2] = 0.0;
+        a[16][3] = 0.0;
+        a[16][4] = 0.0;
+        a[16][5] = 0.0;
+        a[16][6] = 0.0;
+        a[16][7] = -1.055180992746381278594381685080184474710;
+        a[16][8] = 0.2384947263782183112638140557774851953952;
+        a[16][9] = 0.1288151774282991354622515887144081172371;
+        a[16][10] = 0.2276623111046215614614917896665746508574;
+        a[16][11] = 1.229532587437517443160321815173576421896;
+        a[16][12] = 0.4624976662810383487397308275256128637703e-1;
+        a[16][13] = 0.1386196319366293903628308621201843492468;
+        a[16][14] = 0.3080010168319435405203560375162404389992e-1;
+        a[16][15] = 0.0;
+
+        c[17] = 0.7375018139988810429214526568476200466933;
+
+        a[17][0] = 0.1571801061417788021580499748292804743973e-1;
+        a[17][1] = 0.0;
+        a[17][2] = 0.0;
+        a[17][3] = 0.0;
+        a[17][4] = 0.0;
+        a[17][5] = 0.0;
+        a[17][6] = 0.0;
+        a[17][7] = 0.4853403452657363080900487157345670578815;
+        a[17][8] = 0.2107787568904546684574678306538806242458;
+        a[17][9] = 0.1269802413053354152906551816812865612400;
+        a[17][10] = 0.2319687014513919157472314236852033704833;
+        a[17][11] = -0.3620214714069096774945761015915844860748;
+        a[17][12] = 0.5366106712036344076786687672191174826870e-1;
+        a[17][13] = -0.2806066613385549629370667652641771053798e-1;
+        a[17][14] = -0.2378121372710330174923994809623697280793e-1;
+        a[17][15] = 0.0;
+        a[17][16] = 0.2691804261928988988990035710208180655494e-1;
+
+        c[18] = 0.749;
+
+        a[18][0] = 0.1569705832522204237038521016179271074328e-1;
+        a[18][1] = 0.0;
+        a[18][2] = 0.0;
+        a[18][3] = 0.0;
+        a[18][4] = 0.0;
+        a[18][5] = 0.0;
+        a[18][6] = 0.0;
+        a[18][7] = 0.4616075242202111936044636619654233606961;
+        a[18][8] = 0.2113946516698113234089715994366982070261;
+        a[18][9] = 0.1270330917167109232354671760560477406684;
+        a[18][10] = 0.2318540550298708286361100159494550129997;
+        a[18][11] = -0.3385266406688372942274816228586594758312;
+        a[18][12] = 0.5298251972194235542699858758068190007453e-1;
+        a[18][13] = -0.2750461365887187780272257875389040879148e-1;
+        a[18][14] = -0.2361906185395527177468217904834875318962e-1;
+        a[18][15] = 0.0;
+        a[18][16] = 0.2668458089504036186694391521145359328377e-1;
+        a[18][17] = 0.1139683460285541525554621429934611232038e-1;
+
+        c[19] = 0.65;
+
+        a[19][0] = 0.1438964884291216402951708307100553583630e-1;
+        a[19][1] = 0.0;
+        a[19][2] = 0.0;
+        a[19][3] = 0.0;
+        a[19][4] = 0.0;
+        a[19][5] = 0.0;
+        a[19][6] = 0.0;
+        a[19][7] = -1.206901219123788519767229935392578222389;
+        a[19][8] = 0.2505628554639376294459724790398627595052;
+        a[19][9] = 0.1303332915702127271150218433265955255131;
+        a[19][10] = 0.2246717759263522068724895908997148014087;
+        a[19][11] = 1.308419325781946480763221457213122900404;
+        a[19][12] = 0.2589750180376235922034567811002287469857e-2;
+        a[19][13] = 0.8070743254562856892787425005539711000810e-2;
+        a[19][14] = -0.1267568255392829384757055330207124946149e-1;
+        a[19][15] = 0.0;
+        a[19][16] = 0.1129158072373321675992773077865249355325e-1;
+        a[19][17] = 0.3422056680709749787491027385482635813798e-1;
+        a[19][18] = -0.1149726368734142020610819623056729009783;
+
+        c[20] = 0.487;
+
+        a[20][0] = 0.1452348029801041972968851600082942318526e-1;
+        a[20][1] = 0.0;
+        a[20][2] = 0.0;
+        a[20][3] = 0.0;
+        a[20][4] = 0.0;
+        a[20][5] = 0.0;
+        a[20][6] = 0.0;
+        a[20][7] = -0.5214243102465818231380106106018875132310;
+        a[20][8] = 0.1866698844204603947548300448071067931901;
+        a[20][9] = 0.1299316354451273004594203171440036506875;
+        a[20][10] = 0.2262141085765719298250459612902177827497;
+        a[20][11] = 0.6104582639466710747030070634860296980845;
+        a[20][12] = 0.1418715607022412453299982942636029187870e-1;
+        a[20][13] = 0.1480061054412245756808618945399364705863e-1;
+        a[20][14] = -0.3711471609871774475268028540581604884370e-2;
+        a[20][15] = 0.0;
+        a[20][16] = 0.1393256979572558890317889514899316376009e-2;
+        a[20][17] = 1.114731020624073270253135239909350910479;
+        a[20][18] = -1.021208555757145859230933049243060207745;
+        a[20][19] = -0.2795650792912340738723193626472621878286;
+
+        bi8[0][0] = 1.0;
+        bi8[0][1] = -11.00917227309954827048799176229874144363;
+        bi8[0][2] = 53.60376844185445526077321090756616710855;
+        bi8[0][3] = -143.7818601051773394006248841909965394832;
+        bi8[0][4] = 227.7802272724059672847516567630398210649;
+        bi8[0][5] = -212.7667774745397515281313247919614156196;
+        bi8[0][6] = 108.4765585030943501072304151917413023749;
+        bi8[0][7] = -23.28773767438833620555341922996682359208;
+
+        bi8[1][0] = 0.0;
+        bi8[1][1] = 0.0;
+        bi8[1][2] = 0.0;
+        bi8[1][3] = 0.0;
+        bi8[1][4] = 0.0;
+        bi8[1][5] = 0.0;
+        bi8[1][6] = 0.0;
+        bi8[1][7] = 0.0;
+
+        bi8[2][0] = 0.0;
+        bi8[2][1] = 0.0;
+        bi8[2][2] = 0.0;
+        bi8[2][3] = 0.0;
+        bi8[2][4] = 0.0;
+        bi8[2][5] = 0.0;
+        bi8[2][6] = 0.0;
+        bi8[2][7] = 0.0;
+
+        bi8[3][0] = 0.0;
+        bi8[3][1] = 0.0;
+        bi8[3][2] = 0.0;
+        bi8[3][3] = 0.0;
+        bi8[3][4] = 0.0;
+        bi8[3][5] = 0.0;
+        bi8[3][6] = 0.0;
+        bi8[3][7] = 0.0;
+
+        bi8[4][0] = 0.0;
+        bi8[4][1] = 0.0;
+        bi8[4][2] = 0.0;
+        bi8[4][3] = 0.0;
+        bi8[4][4] = 0.0;
+        bi8[4][5] = 0.0;
+        bi8[4][6] = 0.0;
+        bi8[4][7] = 0.0;
+
+        bi8[5][0] = 0.0;
+        bi8[5][1] = 0.0;
+        bi8[5][2] = 0.0;
+        bi8[5][3] = 0.0;
+        bi8[5][4] = 0.0;
+        bi8[5][5] = 0.0;
+        bi8[5][6] = 0.0;
+        bi8[5][7] = 0.0;
+
+        bi8[6][0] = 0.0;
+        bi8[6][1] = 0.0;
+        bi8[6][2] = 0.0;
+        bi8[6][3] = 0.0;
+        bi8[6][4] = 0.0;
+        bi8[6][5] = 0.0;
+        bi8[6][6] = 0.0;
+        bi8[6][7] = 0.0;
+
+        bi8[7][0] = 0.0;
+        bi8[7][1] = 179.0879062178872372197637635178056034051;
+        bi8[7][2] = -1728.403136057631342220019375555018665689;
+        bi8[7][3] = 6564.460940561142416685202971494559376099;
+        bi8[7][4] = -12902.91081584121105690497909548869952323;
+        bi8[7][5] = 13965.17977100265059245172194550171929806;
+        bi8[7][6] = -7920.424061687375443376559247327959298606;
+        bi8[7][7] = 1841.954214811791214866274656172513025486;
+
+        bi8[8][0] = 0.0;
+        bi8[8][1] = -22.88304714767390548521469744522669995060;
+        bi8[8][2] = 238.8040507685294155366494768970788101399;
+        bi8[8][3] = -936.8585782296073752493303117453241744494;
+        bi8[8][4] = 1874.044957762151353867893217234671921697;
+        bi8[8][5] = -2050.511559353009354817713813307430993559;
+        bi8[8][6] = 1171.508545192416250747220534524328950392;
+        bi8[8][7] = -273.8658742664281662882405921023203290746;
+
+        bi8[9][0] = 0.0;
+        bi8[9][1] = 12.17331549628124801002349940562808097269;
+        bi8[9][2] = -82.75076476436870272146201057456216825170;
+        bi8[9][3] = 256.4804438751307755574323340365457074340;
+        bi8[9][4] = -441.0999429758310962594951067937304764418;
+        bi8[9][5] = 434.5079831401213158642449792005352310524;
+        bi8[9][6] = -229.8934605896900265488669050173042421551;
+        bi8[9][7] = 50.71124099578478523358546133160227550674;
+
+        bi8[10][0] = 0.0;
+        bi8[10][1] = 2.613203520133622756551032901538192914958;
+        bi8[10][2] = 16.88002937261184447692017262466190582959;
+        bi8[10][3] = -134.1727921155237789914884572421173077178;
+        bi8[10][4] = 340.1204801692637664465059992362156474712;
+        bi8[10][5] = -420.1258893776879400857436087457229872331;
+        bi8[10][6] = 258.3234372146557103657079225360621734929;
+        bi8[10][7] = -63.41080647234860340699156952097105010691;
+
+        bi8[11][0] = 0.0;
+        bi8[11][1] = -212.3127005913665992194796715806401319940;
+        bi8[11][2] = 2045.352316395154364712249302798824723141;
+        bi8[11][3] = -7762.060413644090906790449120787506530500;
+        bi8[11][4] = 15250.14841459378132543421843792712597249;
+        bi8[11][5] = -16501.08003544517859797013884904039476397;
+        bi8[11][6] = 9356.907363957304758100332957460543520093;
+        bi8[11][7] = -2175.725412678166826823572734962779212842;
+
+        bi8[12][0] = 0.0;
+        bi8[12][1] = -11.50164708567429764283826551320712022931;
+        bi8[12][2] = 107.2769704045725235564526349677807440972;
+        bi8[12][3] = -401.2345332983947627677963563088489537002;
+        bi8[12][4] = 781.8916989572930300323034641740782180816;
+        bi8[12][5] = -841.6593371013254430188454534530483533077;
+        bi8[12][6] = 475.5767391485588457657241425216173583523;
+        bi8[12][7] = -110.3036412584017920901261933056193320075;
+
+        bi8[13][0] = 0.0;
+        bi8[13][1] = -38.77239677340365800498371633092550416428;
+        bi8[13][2] = 358.6391180152970136912934982919377109538;
+        bi8[13][3] = -1336.215732140226920321570086014013189585;
+        bi8[13][4] = 2598.193515313187034150047120315374589407;
+        bi8[13][5] = -2792.878330051016777349083250990468387014;
+        bi8[13][6] = 1576.586790342810048230642938513691709160;
+        bi8[13][7] = -365.4143450747101110059836729234767444075;
+
+        bi8[14][0] = 0.0;
+        bi8[14][1] = -10.00425469998852674211886984302245249578;
+        bi8[14][2] = 91.67754241560498208460382885503842174321;
+        bi8[14][3] = -340.0777600652848208705042811235806913159;
+        bi8[14][4] = 659.6008369976978366087037536255754670281;
+        bi8[14][5] = -707.8826284112975671515477305189138171574;
+        bi8[14][6] = 399.1580030260127128211116842683795971037;
+        bi8[14][7] = -92.44093916106142239619634965972490086213;
+
+        bi8[15][0] = 0.0;
+        bi8[15][1] = 0.0;
+        bi8[15][2] = 0.0;
+        bi8[15][3] = 0.0;
+        bi8[15][4] = 0.0;
+        bi8[15][5] = 0.0;
+        bi8[15][6] = 0.0;
+        bi8[15][7] = 0.0;
+
+        bi8[16][0] = 0.0;
+        bi8[16][1] = 16.21014491337561130525562931390759162785;
+        bi8[16][2] = -150.4492131048662883122727122117594677554;
+        bi8[16][3] = 565.1196631722152490306979317160273112948;
+        bi8[16][4] = -1111.210282651553705321998904060968429058;
+        bi8[16][5] = 1211.236976000843526549958975517785065570;
+        bi8[16][6] = -695.3365606918091595268831805414223001360;
+        bi8[16][7] = 164.4292723617947662752422602664302284566;
+
+        bi8[17][0] = 0.0;
+        bi8[17][1] = -1872.805019318414830108065147723853146433;
+        bi8[17][2] = 16400.66633207048634386587617517712424379;
+        bi8[17][3] = -57270.16260166053287080817377636090729160;
+        bi8[17][4] = 103330.5912960663516254356794429889210991;
+        bi8[17][5] = -102051.1888173727465381067712242944167386;
+        bi8[17][6] = 52424.75260874662694445820923552319700594;
+        bi8[17][7] = -10961.85379853177067473675470531006517221;
+
+        bi8[18][0] = 0.0;
+        bi8[18][1] = 1720.562129057968827268151794567408887461;
+        bi8[18][2] = -15159.62046432254747381864100381352909840;
+        bi8[18][3] = 53314.59561935523492911472686991966671330;
+        bi8[18][4] = -96988.37468691576351426414353587220276161;
+        bi8[18][5] = 96690.39181023687073280170142510495743481;
+        bi8[18][6] = -50199.31248988246623378558547060878127074;
+        bi8[18][7] = 10621.75808247070273268378992070248009519;
+
+        bi8[19][0] = 0.0;
+        bi8[19][1] = 264.2470488396084725751013414957565817480;
+        bi8[19][2] = -2232.338003905684731509741541080134626580;
+        bi8[19][3] = 7453.753802419143998655808817000026688315;
+        bi8[19][4] = -12770.71881400916730561960440987151459948;
+        bi8[19][5] = 11903.25226767446224225175138830095252030;
+        bi8[19][6] = -5733.155576507225740169825158561334353021;
+        bi8[19][7] = 1114.959275488863063816509562716247788712;
+
+        bi8[20][0] = 0.0;
+        bi8[20][1] = -15.60551015563365366165870100287114141984;
+        bi8[20][2] = 40.66145427098759539731834271499129986994;
+        bi8[20][3] = 170.1538018759714061560683496064688819112;
+        bi8[20][4] = -848.0568847386052608898820401778869464852;
+        bi8[20][5] = 1373.524566531853560108596541516407906615;
+        bi8[20][6] = -993.1678967729130171884598684827601522494;
+        bi8[20][7] = 272.4904689883393700780173758256501517585;
+
+        Self {
+            c,
+            a,
+            b,
+            bh: Some(bh),
+            bi: Some(bi8),
+        }
+    }
+}
+
+impl ButcherTableau<16, 26> {
+    /// A better efficient Runge-Kutta (16:9(8)) pair with 9th-order interpolation.
+    ///
+    /// # Overview
+    /// This provides a 16-stage Runge-Kutta method with:
+    /// - Primary order: 9
+    /// - Embedded order: 8 (for error estimation)
+    /// - Number of stages: 16 primary + 10 additional stages for interpolation
+    /// - Dominant stage order: 5
+    ///
+    /// # Efficiency
+    /// This method is particularly efficient with a local truncation error 2-norm:
+    ///
+    /// T_10,2 ~ 0.00000034399
+    ///
+    /// This 2-norm is smaller than previous (16:8,9) pairs and has two local maximum values on [0,1].
+    ///
+    /// # Interpolation
+    /// - The method provides a 9th-order interpolant using additional nodes c[17] through c[25].
+    /// - The interpolant has continuous derivatives.
+    /// - Maximum 2-norm of the local truncation error over [0,1] for the interpolant:
+    ///
+    ///   Ti_10,2 ~ 0.0000007830
+    ///
+    /// - The interpolation error has two local maximum values on [0,1].
+    ///
+    /// # Notes
+    /// - Exact coefficients of the method are possible using surds in terms of sqrt(6).
+    /// - Exact coefficients for the interpolants cannot be conveniently represented
+    ///   as they require c[18] to be a zero of a cubic polynomial with coefficients
+    ///   that are surds in terms of sqrt(6).
+    ///
+    /// # References
+    /// - J.H. Verner, SIAM J NA 1978, "Explicit Runge-Kutta methods with estimates of the Local Truncation Error"
+    /// - J.H. Verner, SIAM J NA 1993, "Differentiable Interpolants for high-order Runge-Kutta methods"
+    pub const fn rkv989e() -> Self {
+        let mut c = [0.0; 26];
+        let mut a = [[0.0; 26]; 26];
+        let mut b = [0.0; 16];
+        let mut bh = [0.0; 16];
+        let mut bi9 = [[0.0; 26]; 26];
+
+        c[0] = 0.0;
+        c[1] = 0.3571e-1;
+        c[2] = 0.9906028091267414072062290548990445721420e-1;
+        c[3] = 0.1485904213690112110809343582348566858213;
+        c[4] = 0.6134;
+        c[5] = 0.2327359473605626756631680289030288192566;
+        c[6] = 0.5538640526394373243368319710969711807434;
+        c[7] = 0.6555;
+        c[8] = 0.491625;
+        c[9] = 0.6858e-1;
+        c[10] = 0.253;
+        c[11] = 0.6620641795412045944786226689660879617864;
+        c[12] = 0.8309;
+        c[13] = 0.8998;
+        c[14] = 1.0;
+        c[15] = 1.0;
+
+        a[1][0] = 0.3571e-1;
+
+        a[2][0] = -0.3833735636677017025757228807792426014327e-1;
+        a[2][1] = 0.1373976372794443109781951935678287173575;
+
+        a[3][0] = 0.3714760534225280277023358955871417145532e-1;
+        a[3][1] = 0.0;
+        a[3][2] = 0.1114428160267584083107007686761425143660;
+
+        a[4][0] = 2.674764429871505119194043347886848033362;
+        a[4][1] = 0.0;
+        a[4][2] = -9.982382134885293836441318015951354323506;
+        a[4][3] = 7.921017705013788717247274668064506290145;
+
+        a[5][0] = 0.5242104050577351069841401193318449756747e-1;
+        a[5][1] = 0.0;
+        a[5][2] = 0.0;
+        a[5][3] = 0.1796911189175953081279466649977364298948;
+        a[5][4] = 0.6237879371938568368073519721078917943307e-3;
+
+        a[6][0] = 0.1592492223647632083060991150129732726703;
+        a[6][1] = 0.0;
+        a[6][2] = 0.0;
+        a[6][3] = -0.4298429877241087508189185066909803751595;
+        a[6][4] = 0.6665266542726088051243445942394404211672e-1;
+        a[6][5] = 0.7578051525715219863372169033510342411158;
+
+        a[7][0] = 0.7283333333333333333333333333333333333333e-1;
+        a[7][1] = 0.0;
+        a[7][2] = 0.0;
+        a[7][3] = 0.0;
+        a[7][4] = 0.0;
+        a[7][5] = 0.3359344590665103678713422141936031057620;
+        a[7][6] = 0.2467322076001562987953244524730635609046;
+
+        a[8][0] = 0.7297558593750000000000000000000000000000e-1;
+        a[8][1] = 0.0;
+        a[8][2] = 0.0;
+        a[8][3] = 0.0;
+        a[8][4] = 0.0;
+        a[8][5] = 0.3348009729699333533129043555243262838993;
+        a[8][6] = 0.1184158239050666466870956444756737161007;
+        a[8][7] = -0.3456738281250000000000000000000000000000e-1;
+
+        a[9][0] = 0.4911213663452096382929799914888692571185e-1;
+        a[9][1] = 0.0;
+        a[9][2] = 0.0;
+        a[9][3] = 0.0;
+        a[9][4] = 0.0;
+        a[9][5] = 0.3983857361308652347066808423646954244674e-1;
+        a[9][6] = 0.1069675288939354812077139476183615284408;
+        a[9][7] = -0.2174259165458647598455650055976622663086e-1;
+        a[9][8] = -0.1055956474869564925231235304439517699685;
+
+        a[10][0] = -0.2707988818641280489175862864774930636734e-1;
+        a[10][1] = 0.0;
+        a[10][2] = 0.0;
+        a[10][3] = 0.0;
+        a[10][4] = 0.0;
+        a[10][5] = 0.3330000000000000000000000000000000000000e-1;
+        a[10][6] = -0.1645526070036057075627040519129748090051;
+        a[10][7] = 0.3428266306497389946775103799535438271025e-1;
+        a[10][8] = 0.1585264064439221046855989221358760453749;
+        a[10][9] = 0.2185234256811225083011127204294936872873;
+
+        a[11][0] = 0.5584657769108862727526281006220448941548e-1;
+        a[11][1] = 0.0;
+        a[11][2] = 0.0;
+        a[11][3] = 0.0;
+        a[11][4] = 0.0;
+        a[11][5] = 0.9166533166672539087479545790553151183505e-1;
+        a[11][6] = 0.2392399655523627049569179599100121305027;
+        a[11][7] = 0.1023834712248414879740316672186561113729e-1;
+        a[11][8] = -0.2679331322859542442991625976085644567159e-2;
+        a[11][9] = 0.4235624181474284646680744174197106205439e-1;
+        a[11][10] = 0.2253970470166604185504274586005888014087;
+
+        a[12][0] = -0.4802510512725195893103945790000307739284;
+        a[12][1] = 0.0;
+        a[12][2] = 0.0;
+        a[12][3] = 0.0;
+        a[12][4] = 0.0;
+        a[12][5] = -6.359610162555930097843606206238737368598;
+        a[12][6] = -0.2762313898040841385163666588363199062208;
+        a[12][7] = -6.500796633979846747011407331804389650124;
+        a[12][8] = 0.5734765877040956875980719881692630407295;
+        a[12][9] = 1.347125994868138849716719646646049357523;
+        a[12][10] = 5.936840409706221306933249144503479736029;
+        a[12][11] = 6.590346245333924728433733996560685564589;
+
+        a[13][0] = 0.3307533067671401079975082476263855100019;
+        a[13][1] = 0.0;
+        a[13][2] = 0.0;
+        a[13][3] = 0.0;
+        a[13][4] = 0.0;
+        a[13][5] = 5.956207776829962111958631943398154563177;
+        a[13][6] = -0.4868316400481527952968983900652017328476;
+        a[13][7] = 4.462055288206771191616083778225319083232;
+        a[13][8] = 0.7410258231442071778525406966672954745201;
+        a[13][9] = -0.7118192034575913119318346652024763899638;
+        a[13][10] = -5.454619594516665440649148499095762263275;
+        a[13][11] = -4.140803729244709693305868672684077643030;
+        a[13][12] = 0.2038319723190386517589855611303633981843;
+
+        a[14][0] = -0.5847111122998944389651719860007729221315;
+        a[14][1] = 0.0;
+        a[14][2] = 0.0;
+        a[14][3] = 0.0;
+        a[14][4] = 0.0;
+        a[14][5] = -12.41268417116267068926060969930886386670;
+        a[14][6] = 1.360245445660928146977003413153919520189;
+        a[14][7] = -22.42610531111868229005965344123311160860;
+        a[14][8] = -0.8828857055865458252669099251378431470909;
+        a[14][9] = 1.770155128538230466813900572142546329490;
+        a[14][10] = 12.15809651918533877218739262890752294243;
+        a[14][11] = 22.23037520407760703834156132941550000589;
+        a[14][12] = -0.6634483760201249006317401316871863808959;
+        a[14][13] = 0.4509623787258137198642272397482891274306;
+
+        a[15][0] = 1.940575549810648717395482898478643635949;
+        a[15][1] = 0.0;
+        a[15][2] = 0.0;
+        a[15][3] = 0.0;
+        a[15][4] = 0.0;
+        a[15][5] = 21.97798408114556310164832153356570429956;
+        a[15][6] = 0.8230747326984728145128172720328075817426;
+        a[15][7] = 68.16441683626354817206843139989369741474;
+        a[15][8] = -3.117097463620266666801525529574905518624;
+        a[15][9] = -4.568841021822439620303730003755793384193;
+        a[15][10] = -18.74190987126264955904919353670543412590;
+        a[15][11] = -66.57711839637831878350036856645874492472;
+        a[15][12] = 1.098915553165441824029764532524025021446;
+        a[15][13] = 0.0;
+        a[15][14] = 0.0;
+
+        b[0] = 0.1500669014979724795766288712377040980022e-1;
+        b[1] = 0.0;
+        b[2] = 0.0;
+        b[3] = 0.0;
+        b[4] = 0.0;
+        b[5] = 0.0;
+        b[6] = 0.0;
+        b[7] = -1.055180992746381278594381685080184474710;
+        b[8] = 0.2384947263782183112638140557774851953952;
+        b[9] = 0.1288151774282991354622515887144081172371;
+        b[10] = 0.2276623111046215614614917896665746508574;
+        b[11] = 1.229532587437517443160321815173576421896;
+        b[12] = 0.4624976662810383487397308275256128637703e-1;
+        b[13] = 0.1386196319366293903628308621201843492468;
+        b[14] = 0.3080010168319435405203560375162404389992e-1;
+        b[15] = 0.0;
+
+        bh[0] = 0.1897210532481101330735875918987801428423e-1;
+        bh[1] = 0.0;
+        bh[2] = 0.0;
+        bh[3] = 0.0;
+        bh[4] = 0.0;
+        bh[5] = 0.0;
+        bh[6] = 0.0;
+        bh[7] = 3.408110314549493848404398964228060776034;
+        bh[8] = 0.1260323883820920906560270507988839661845;
+        bh[9] = 0.1188375063451149770930378540709095182997;
+        bh[10] = 0.2491041997838687569190073177760326138188;
+        bh[11] = -3.269966219928978218713853055139116510335;
+        bh[12] = 0.3023798100228882907409723963501699715077;
+        bh[13] = 0.0;
+        bh[14] = 0.0;
+        bh[15] = 0.4652989552070924159305071272518165020637e-1;
+
+        c[16] = 1.0;
+
+        a[16][0] = 0.1500669014979724795766288712377040980022e-1;
+        a[16][1] = 0.0;
+        a[16][2] = 0.0;
+        a[16][3] = 0.0;
+        a[16][4] = 0.0;
+        a[16][5] = 0.0;
+        a[16][6] = 0.0;
+        a[16][7] = -1.055180992746381278594381685080184474710;
+        a[16][8] = 0.2384947263782183112638140557774851953952;
+        a[16][9] = 0.1288151774282991354622515887144081172371;
+        a[16][10] = 0.2276623111046215614614917896665746508574;
+        a[16][11] = 1.229532587437517443160321815173576421896;
+        a[16][12] = 0.4624976662810383487397308275256128637703e-1;
+        a[16][13] = 0.1386196319366293903628308621201843492468;
+        a[16][14] = 0.3080010168319435405203560375162404389992e-1;
+        a[16][15] = 0.0;
+
+        c[17] = 0.7375018139988810429214526568476200466933;
+
+        a[17][0] = 0.1571801061417788021580499748292804743973e-1;
+        a[17][1] = 0.0;
+        a[17][2] = 0.0;
+        a[17][3] = 0.0;
+        a[17][4] = 0.0;
+        a[17][5] = 0.0;
+        a[17][6] = 0.0;
+        a[17][7] = 0.4853403452657363080900487157345670578815;
+        a[17][8] = 0.2107787568904546684574678306538806242458;
+        a[17][9] = 0.1269802413053354152906551816812865612400;
+        a[17][10] = 0.2319687014513919157472314236852033704833;
+        a[17][11] = -0.3620214714069096774945761015915844860748;
+        a[17][12] = 0.5366106712036344076786687672191174826870e-1;
+        a[17][13] = -0.2806066613385549629370667652641771053798e-1;
+        a[17][14] = -0.2378121372710330174923994809623697280793e-1;
+        a[17][15] = 0.0;
+        a[17][16] = 0.2691804261928988988990035710208180655494e-1;
+
+        c[18] = 0.749;
+
+        a[18][0] = 0.1569705832522204237038521016179271074328e-1;
+        a[18][1] = 0.0;
+        a[18][2] = 0.0;
+        a[18][3] = 0.0;
+        a[18][4] = 0.0;
+        a[18][5] = 0.0;
+        a[18][6] = 0.0;
+        a[18][7] = 0.4616075242202111936044636619654233606961;
+        a[18][8] = 0.2113946516698113234089715994366982070261;
+        a[18][9] = 0.1270330917167109232354671760560477406684;
+        a[18][10] = 0.2318540550298708286361100159494550129997;
+        a[18][11] = -0.3385266406688372942274816228586594758312;
+        a[18][12] = 0.5298251972194235542699858758068190007453e-1;
+        a[18][13] = -0.2750461365887187780272257875389040879148e-1;
+        a[18][14] = -0.2361906185395527177468217904834875318962e-1;
+        a[18][15] = 0.0;
+        a[18][16] = 0.2668458089504036186694391521145359328377e-1;
+        a[18][17] = 0.1139683460285541525554621429934611232038e-1;
+
+        c[19] = 0.65;
+
+        a[19][0] = 0.1438964884291216402951708307100553583630e-1;
+        a[19][1] = 0.0;
+        a[19][2] = 0.0;
+        a[19][3] = 0.0;
+        a[19][4] = 0.0;
+        a[19][5] = 0.0;
+        a[19][6] = 0.0;
+        a[19][7] = -1.206901219123788519767229935392578222389;
+        a[19][8] = 0.2505628554639376294459724790398627595052;
+        a[19][9] = 0.1303332915702127271150218433265955255131;
+        a[19][10] = 0.2246717759263522068724895908997148014087;
+        a[19][11] = 1.308419325781946480763221457213122900404;
+        a[19][12] = 0.2589750180376235922034567811002287469857e-2;
+        a[19][13] = 0.8070743254562856892787425005539711000810e-2;
+        a[19][14] = -0.1267568255392829384757055330207124946149e-1;
+        a[19][15] = 0.0;
+        a[19][16] = 0.1129158072373321675992773077865249355325e-1;
+        a[19][17] = 0.3422056680709749787491027385482635813798e-1;
+        a[19][18] = -0.1149726368734142020610819623056729009783;
+
+        c[20] = 0.487;
+
+        a[20][0] = 0.1452348029801041972968851600082942318526e-1;
+        a[20][1] = 0.0;
+        a[20][2] = 0.0;
+        a[20][3] = 0.0;
+        a[20][4] = 0.0;
+        a[20][5] = 0.0;
+        a[20][6] = 0.0;
+        a[20][7] = -0.5214243102465818231380106106018875132310;
+        a[20][8] = 0.1866698844204603947548300448071067931901;
+        a[20][9] = 0.1299316354451273004594203171440036506875;
+        a[20][10] = 0.2262141085765719298250459612902177827497;
+        a[20][11] = 0.6104582639466710747030070634860296980845;
+        a[20][12] = 0.1418715607022412453299982942636029187870e-1;
+        a[20][13] = 0.1480061054412245756808618945399364705863e-1;
+        a[20][14] = -0.3711471609871774475268028540581604884370e-2;
+        a[20][15] = 0.0;
+        a[20][16] = 0.1393256979572558890317889514899316376009e-2;
+        a[20][17] = 1.114731020624073270253135239909350910479;
+        a[20][18] = -1.021208555757145859230933049243060207745;
+        a[20][19] = -0.2795650792912340738723193626472621878286;
+
+        c[21] = 0.97e-2;
+
+        a[21][0] = 0.8711816186418632397551552134391273160887e-2;
+        a[21][1] = 0.0;
+        a[21][2] = 0.0;
+        a[21][3] = 0.0;
+        a[21][4] = 0.0;
+        a[21][5] = 0.0;
+        a[21][6] = 0.0;
+        a[21][7] = 0.1532993247326560354009297246756198757646e-1;
+        a[21][8] = -0.1943250606288015181452523235039064995331e-2;
+        a[21][9] = 0.1072095950570478409673351209542485874475e-2;
+        a[21][10] = 0.2601233036074381124605301648955068308034e-3;
+        a[21][11] = -0.1817718521410219268803888649170562657549e-1;
+        a[21][12] = -0.9877668338996712912794853387601975290298e-3;
+        a[21][13] = -0.3332383192417756458522769371199819990865e-2;
+        a[21][14] = -0.8605833352714280922890371244764817761631e-3;
+        a[21][15] = 0.0;
+        a[21][16] = 0.1392810143886650671629507962655268250424e-2;
+        a[21][17] = -0.1617419987897630058856682050471490927435;
+        a[21][18] = 0.1485156561306044255987175208430691701911;
+        a[21][19] = 0.2289051095253062003982633508518682297336e-1;
+        a[21][20] = -0.1429777169141779172700863258972231247118e-2;
+
+        c[22] = 0.138;
+
+        a[22][0] = 0.2710092628714179537584868171073201523422e-1;
+        a[22][1] = 0.0;
+        a[22][2] = 0.0;
+        a[22][3] = 0.0;
+        a[22][4] = 0.0;
+        a[22][5] = 0.0;
+        a[22][6] = 0.0;
+        a[22][7] = 0.6923104986003885114772983986172611691902;
+        a[22][8] = -0.6725243132164494445073150422370697609664e-1;
+        a[22][9] = 0.8808440659269460280192719417821072306037e-1;
+        a[22][10] = 0.5982566312199630889836790781586188392095e-1;
+        a[22][11] = -0.8251290314815523858109567180972702723517;
+        a[22][12] = -0.4886453387508305661509501250020800230341e-1;
+        a[22][13] = -0.1682608371566511953625070499451460275789;
+        a[22][14] = -0.4443170503743608176862370639456174098455e-1;
+        a[22][15] = 0.0;
+        a[22][16] = 0.7037844639432780238885500484343585742713e-1;
+        a[22][17] = -8.818686397504859095347055843377871115177;
+        a[22][18] = 8.028821919603921347277176943010114087042;
+        a[22][19] = 1.306591406499157862229641342220359676579;
+        a[22][20] = -0.1624883307224014710941456378572112779619;
+        a[22][21] = 0.0;
+
+        c[23] = 0.249;
+
+        a[23][0] = 0.1466442426857961230207869561545353695312e-1;
+        a[23][1] = 0.0;
+        a[23][2] = 0.0;
+        a[23][3] = 0.0;
+        a[23][4] = 0.0;
+        a[23][5] = 0.0;
+        a[23][6] = 0.0;
+        a[23][7] = 0.1898399561399124450873282966092465494738;
+        a[23][8] = 0.3711530651907537053995301753121055816571e-1;
+        a[23][9] = 0.1316247981199510429999180335475037001729;
+        a[23][10] = 0.1466622928822685713088228930531811168096;
+        a[23][11] = -0.2377331162467838377764310139745632904067;
+        a[23][12] = -0.2493071232192949761907468241344412740688e-1;
+        a[23][13] = -0.9427777049221307692559326299260474943801e-1;
+        a[23][14] = -0.2726667627690941070300686079610559177653e-1;
+        a[23][15] = 0.0;
+        a[23][16] = 0.4097659404063362064133671142526181935526e-1;
+        a[23][17] = -5.538725321977494545432278985839620727708;
+        a[23][18] = 4.973442175247713118429917494679987165227;
+        a[23][19] = 0.8626294449189858956229369969895335684221;
+        a[23][20] = -0.2250213948217893084759073334350395278429;
+        a[23][21] = 0.0;
+        a[23][22] = 0.0;
+
+        c[24] = 0.439;
+
+        a[24][0] = 0.1191252689920919922350166270042464848683e-1;
+        a[24][1] = 0.0;
+        a[24][2] = 0.0;
+        a[24][3] = 0.0;
+        a[24][4] = 0.0;
+        a[24][5] = 0.0;
+        a[24][6] = 0.0;
+        a[24][7] = -0.6723137858130876080443132301001699876552;
+        a[24][8] = 0.1801284266825688933352476317750604591900;
+        a[24][9] = 0.1365689447449347163335592887842694005196;
+        a[24][10] = 0.2111038393798906615990851223053694372026;
+        a[24][11] = 0.7775854726244728792738759030556073621755;
+        a[24][12] = 0.2362117948505939633346199470372118351743e-1;
+        a[24][13] = 0.6391325607581230059121991748062989083367e-1;
+        a[24][14] = 0.1197654289246164466950805226357270574198e-1;
+        a[24][15] = 0.0;
+        a[24][16] = -0.2188622145087031801816344581960035242509e-1;
+        a[24][17] = 3.500218060218659348421859254969160434525;
+        a[24][18] = -3.195765244251373581009030415350620096609;
+        a[24][19] = -0.5994401048600148264885439684062323011428;
+        a[24][20] = 0.1137710737227729377873223163880721563872e-1;
+        a[24][21] = 0.0;
+        a[24][22] = 0.0;
+        a[24][23] = 0.0;
+
+        c[25] = 0.794;
+
+        a[25][0] = 0.1184012014074604271397800922251586670558e-1;
+        a[25][1] = 0.0;
+        a[25][2] = 0.0;
+        a[25][3] = 0.0;
+        a[25][4] = 0.0;
+        a[25][5] = 0.0;
+        a[25][6] = 0.0;
+        a[25][7] = -0.6641263289768790103963252230448945748496;
+        a[25][8] = 0.1788914061873194018510863179806909040625;
+        a[25][9] = 0.1367562642287085016908037148859544845819;
+        a[25][10] = 0.2107758119730286112356534765669813731420;
+        a[25][11] = 0.7679184744806335059253231677825531412084;
+        a[25][12] = 0.2313496869580297355994450248598930620113e-1;
+        a[25][13] = 0.6230605705999706953753333719939087802061e-1;
+        a[25][14] = 0.1157099029193866324646273674695267419004e-1;
+        a[25][15] = 0.0;
+        a[25][16] = -0.2129441601042184344575923744216391881547e-1;
+        a[25][17] = 3.163166251075345445930896810245694869610;
+        a[25][18] = -2.819544872276817190650813744596266559171;
+        a[25][19] = -0.3978769046309758754403585402077121882020;
+        a[25][20] = 0.1304821777615737042415746721743137433164;
+        a[25][21] = 0.0;
+        a[25][22] = 0.0;
+        a[25][23] = 0.0;
+        a[25][24] = 0.0;
+
+        bi9[0][0] = 1.0;
+        bi9[0][1] = -60.67156499096275825192551363602396761243;
+        bi9[0][2] = 669.4173339890965298528834018742984105094;
+        bi9[0][3] = -3377.878946225199351520226690513332256243;
+        bi9[0][4] = 9286.468967391046692225695247237821541998;
+        bi9[0][5] = -14780.47713681102054116333940623447596217;
+        bi9[0][6] = 13604.99386328224720102647507806334531178;
+        bi9[0][7] = -6724.705443356114485986782323617396379907;
+        bi9[0][8] = 1381.867933411056511065177869712887072054;
+
+        bi9[1][0] = 0.0;
+        bi9[1][1] = 0.0;
+        bi9[1][2] = 0.0;
+        bi9[1][3] = 0.0;
+        bi9[1][4] = 0.0;
+        bi9[1][5] = 0.0;
+        bi9[1][6] = 0.0;
+        bi9[1][7] = 0.0;
+        bi9[1][8] = 0.0;
+
+        bi9[2][0] = 0.0;
+        bi9[2][1] = 0.0;
+        bi9[2][2] = 0.0;
+        bi9[2][3] = 0.0;
+        bi9[2][4] = 0.0;
+        bi9[2][5] = 0.0;
+        bi9[2][6] = 0.0;
+        bi9[2][7] = 0.0;
+        bi9[2][8] = 0.0;
+
+        bi9[3][0] = 0.0;
+        bi9[3][1] = 0.0;
+        bi9[3][2] = 0.0;
+        bi9[3][3] = 0.0;
+        bi9[3][4] = 0.0;
+        bi9[3][5] = 0.0;
+        bi9[3][6] = 0.0;
+        bi9[3][7] = 0.0;
+        bi9[3][8] = 0.0;
+
+        bi9[4][0] = 0.0;
+        bi9[4][1] = 0.0;
+        bi9[4][2] = 0.0;
+        bi9[4][3] = 0.0;
+        bi9[4][4] = 0.0;
+        bi9[4][5] = 0.0;
+        bi9[4][6] = 0.0;
+        bi9[4][7] = 0.0;
+        bi9[4][8] = 0.0;
+
+        bi9[5][0] = 0.0;
+        bi9[5][1] = 0.0;
+        bi9[5][2] = 0.0;
+        bi9[5][3] = 0.0;
+        bi9[5][4] = 0.0;
+        bi9[5][5] = 0.0;
+        bi9[5][6] = 0.0;
+        bi9[5][7] = 0.0;
+        bi9[5][8] = 0.0;
+
+        bi9[6][0] = 0.0;
+        bi9[6][1] = 0.0;
+        bi9[6][2] = 0.0;
+        bi9[6][3] = 0.0;
+        bi9[6][4] = 0.0;
+        bi9[6][5] = 0.0;
+        bi9[6][6] = 0.0;
+        bi9[6][7] = 0.0;
+        bi9[6][8] = 0.0;
+
+        bi9[7][0] = 0.0;
+        bi9[7][1] = -0.5972167605668447080195299461102185885272;
+        bi9[7][2] = 47.93604836413167338885438264277611366496;
+        bi9[7][3] = -565.7170417002332153647406623941231321861;
+        bi9[7][4] = 2799.622707202858561151696501297251806365;
+        bi9[7][5] = -7032.161794919678731491945450324980521259;
+        bi9[7][6] = 9404.154583498766261417380396057323248088;
+        bi9[7][7] = -6374.661804344304055026346155757703309071;
+        bi9[7][8] = 1720.369337666279969354526136740485828512;
+
+        bi9[8][0] = 0.0;
+        bi9[8][1] = 0.1349844707960069717220566866826298984923;
+        bi9[8][2] = -10.83462914594452681222359026508445675428;
+        bi9[8][3] = 127.8648231869934298266652232809372858609;
+        bi9[8][4] = -632.7779367772187332338007906200698487630;
+        bi9[8][5] = 1589.427325412255464804263544956685619289;
+        bi9[8][6] = -2125.551246305557207529871063524735793374;
+        bi9[8][7] = 1440.817483665754881321490009540733392622;
+        bi9[8][8] = -388.8423097807010970369815759993713435843;
+
+        bi9[9][0] = 0.0;
+        bi9[9][1] = 0.7290747606753265739553888434183906298602e-1;
+        bi9[9][2] = -5.851972901033216313427390997286151384252;
+        bi9[9][3] = 69.06202973876299355875761438134035666890;
+        bi9[9][4] = -341.7744427162097135566927248434513873995;
+        bi9[9][5] = 858.4775271201373000458419246386670445384;
+        bi9[9][6] = -1148.047443579864082962792949259101806969;
+        bi9[9][7] = 778.2107496409214137189987901610599544952;
+        bi9[9][8] = -210.0205396013539280126185513768554408959;
+
+        bi9[10][0] = 0.0;
+        bi9[10][1] = 0.1288534847345785561116203329301814360204;
+        bi9[10][2] = -10.34252098059179829195550043573389974374;
+        bi9[10][3] = 122.0572110662541635697477898873947855333;
+        bi9[10][4] = -604.0372032137003894354892742901494195352;
+        bi9[10][5] = 1517.235637581122757484801301501208740856;
+        bi9[10][6] = -2029.008844152904067945761282194453917620;
+        bi9[10][7] = 1375.375645376320937242441893999796663414;
+        bi9[10][8] = -371.1811168501315596184350570113265596896;
+
+        bi9[11][0] = 0.0;
+        bi9[11][1] = 0.6958971720762387362164791425187492767466;
+        bi9[11][2] = -55.85670513574830649397484217620113261823;
+        bi9[11][3] = 659.1926340795728211557235694809221532847;
+        bi9[11][4] = -3262.215084140846354068888231363556701923;
+        bi9[11][5] = 8194.112807588290125367213373592291507372;
+        bi9[11][6] = -10958.03904467296551539422209426971705612;
+        bi9[11][7] = 7427.971576643475789982515741208313369971;
+        bi9[11][8] = -2004.632548946417281841423673799397312817;
+
+        bi9[12][0] = 0.0;
+        bi9[12][1] = 0.2617668058132623028300450746186815539953e-1;
+        bi9[12][2] = -2.101090775094604504096600471528167656924;
+        bi9[12][3] = 24.79601256660032039131082591571668336326;
+        bi9[12][4] = -122.7106039105787684003939157931166303491;
+        bi9[12][5] = 308.2275402436813943146648421712827254341;
+        bi9[12][6] = -412.1946451000669582597347477727378421190;
+        bi9[12][7] = 279.4085780645544831469608316472195929468;
+        bi9[12][8] = -75.40571800304908908412026712154566848821;
+
+        bi9[13][0] = 0.0;
+        bi9[13][1] = 0.7845665161262074618502665073713581595690e-1;
+        bi9[13][2] = -6.297381611696195354835186970953877598132;
+        bi9[13][3] = 74.31851847203782644803150125553425852543;
+        bi9[13][4] = -367.7877747056500197543410562506219227864;
+        bi9[13][5] = 923.8184599908639672968037927810710633244;
+        bi9[13][6] = -1235.428287659738944177317308989226766106;
+        bi9[13][7] = 837.4423716056154246604113713897017147785;
+        bi9[13][8] = -226.0057431111080504745753090041214216040;
+
+        bi9[14][0] = 0.0;
+        bi9[14][1] = 0.1743239982411997083542666099618776565927e-1;
+        bi9[14][2] = -1.399224563421080747566112426047905797818;
+        bi9[14][3] = 16.51294188206731169216889160389405043672;
+        bi9[14][4] = -81.71931133065192111822155020120302944168;
+        bi9[14][5] = 205.2645942498128099152915051189419910813;
+        bi9[14][6] = -274.5016441798805281613128379066882052764;
+        bi9[14][7] = 186.0725630195001464648099116049710397144;
+        bi9[14][8] = -50.21655137556766366195319885111250443823;
+
+        bi9[15][0] = 0.0;
+        bi9[15][1] = 0.0;
+        bi9[15][2] = 0.0;
+        bi9[15][3] = 0.0;
+        bi9[15][4] = 0.0;
+        bi9[15][5] = 0.0;
+        bi9[15][6] = 0.0;
+        bi9[15][7] = 0.0;
+        bi9[15][8] = 0.0;
+
+        bi9[16][0] = 0.0;
+        bi9[16][1] = -0.3120237034739356102149090910458972991542e-1;
+        bi9[16][2] = 2.505323513858665425821148146925028556689;
+        bi9[16][3] = -29.63209944182268891094187050958569485080;
+        bi9[16][4] = 147.1735748991918299269418428659506735081;
+        bi9[16][5] = -371.6682626297095949736614911358789582154;
+        bi9[16][6] = 500.8032537625967383990239072447982187459;
+        bi9[16][7] = -342.9490465144388046578978125156518258548;
+        bi9[16][8] = 93.79845878067124835173576681254714784020;
+
+        bi9[17][0] = 0.0;
+        bi9[17][1] = 1.005189184600557566129408870772027118073;
+        bi9[17][2] = -80.14524762849856177105235169139266835049;
+        bi9[17][3] = 904.2862871618540482735522275321299612604;
+        bi9[17][4] = -4159.377900797260916077366144135436208057;
+        bi9[17][5] = 9382.948440638739069789096069230867579132;
+        bi9[17][6] = -10865.07436232189535132287509765058944820;
+        bi9[17][7] = 6171.218731586134383883573674542623005216;
+        bi9[17][8] = -1354.861137823673230341057786698974248115;
+
+        bi9[18][0] = 0.0;
+        bi9[18][1] = -1.018790304892149008339208622085912341214;
+        bi9[18][2] = 81.22968543433086703732306275594490877087;
+        bi9[18][3] = -916.5220998408498750879913972119183817869;
+        bi9[18][4] = 4215.658051870923057302442161815205855483;
+        bi9[18][5] = -9509.908233268860695396133182984225322485;
+        bi9[18][6] = 11012.08866136426918043466675789153630399;
+        bi9[18][7] = -6254.720911673139287616309544212696219769;
+        bi9[18][8] = 1373.193636418218902334341350568238768134;
+
+        bi9[19][0] = 0.0;
+        bi9[19][1] = 0.1486756383162591888438322851466844348297e-1;
+        bi9[19][2] = -1.185413256701197518560901363041446866220;
+        bi9[19][3] = 13.37512808773948803650703726266929826638;
+        bi9[19][4] = -61.52057482048160294775557914196592133996;
+        bi9[19][5] = 138.7814224498311940264420678029949672145;
+        bi9[19][6] = -160.7032676951998445023342766109082536156;
+        bi9[19][7] = 91.27733347752088221300420564599480170832;
+        bi9[19][8] = -20.03949580654054522618693682425811381100;
+
+        bi9[20][0] = 0.0;
+        bi9[20][1] = -0.2611963106691748849299256055037070183077e-2;
+        bi9[20][2] = 0.2082557524387797755360510166314825074305;
+        bi9[20][3] = -2.349769034664475310156364256215305171863;
+        bi9[20][4] = 10.80805662268303734097853101791700152378;
+        bi9[20][5] = -24.38139559637047359365680950022456660620;
+        bi9[20][6] = 28.23266885539013812966436802211627788468;
+        bi9[20][7] = -16.03578301196446470276182164155397173990;
+        bi9[20][8] = 3.520578375594150109245344597384118672243;
+
+        bi9[21][0] = 0.0;
+        bi9[21][1] = 61.75136568508868110286254101938221664703;
+        bi9[21][2] = -750.1268910552884473493229930939554325311;
+        bi9[21][3] = 3898.685586244841930988256009725848784326;
+        bi9[21][4] = -10841.82272864745777154320543067614124209;
+        bi9[21][5] = 17321.57566462516617366130887449349951271;
+        bi9[21][6] = -15937.70492491813691151403407581204545506;
+        bi9[21][7] = 7853.047625862506649651583373646970497362;
+        bi9[21][8] = -1605.405697796720304997448299303558881367;
+
+        bi9[22][0] = 0.0;
+        bi9[22][1] = -1.934298042224096808013043975920800289243;
+        bi9[22][2] = 147.5886431665603647558155476856964112175;
+        bi9[22][3] = -1181.453462187681932247747922750006635641;
+        bi9[22][4] = 4124.816268317402451398433861948686547907;
+        bi9[22][5] = -7708.555106592447263782269817638341609752;
+        bi9[22][6] = 8052.349419898650085369881466158633077027;
+        bi9[22][7] = -4443.023055054952034626751291729687203499;
+        bi9[22][8] = 1010.211590494692425940651200300940213030;
+
+        bi9[23][0] = 0.0;
+        bi9[23][1] = 0.9126428307869666507744889095249555815836;
+        bi9[23][2] = -71.31710609019478576924959388866173624547;
+        bi9[23][3] = 695.4561446075446391059957659052646164863;
+        bi9[23][4] = -2620.882193436484363083577561241318740147;
+        bi9[23][5] = 4957.008141154350233563895859921266117993;
+        bi9[23][6] = -5038.298085544127560098020485475941711703;
+        bi9[23][7] = 2633.333935059078624368760947592332675986;
+        bi9[23][8] = -556.2134785809537547385794217224661779516;
+
+        bi9[24][0] = 0.0;
+        bi9[24][1] = -0.4654439970747991651050782556914121007009;
+        bi9[24][2] = 37.09099810444154184865658933003385399191;
+        bi9[24][3] = -417.1382228339854991229014922810663023285;
+        bi9[24][4] = 1920.758799061188217073321625257641207277;
+        bi9[24][5] = -4409.960156149581696033720218281819632770;
+        bi9[24][6] = 5383.383503749768100112081288811655341483;
+        bi9[24][7] = -3353.518501772742141737984950593448341369;
+        bi9[24][8] = 839.8490238379862770256522360126952858160;
+
+        bi9[25][0] = 0.0;
+        bi9[25][1] = -0.1176451708255218561268102928705214692156;
+        bi9[25][2] = 9.481894819354298841374880327580666327841;
+        bi9[25][3] = -114.9156758298319354820100563154045258044;
+        bi9[25][4] = 591.3193291312467068002224871165564177657;
+        bi9[25][5] = -1559.765475086581493834896780108830295683;
+        bi9[25][6] = 2198.545841718649266979102957216738477165;
+        bi9[25][7] = -1564.562048273728342299716850911579457005;
+        bi9[25][8] = 440.0137786917170208520501729678092387026;
+
+        Self {
+            c,
+            a,
+            b,
+            bh: Some(bh),
+            bi: Some(bi9),
+        }
+    }
+}
